% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{a4me_cohen_d}
\alias{a4me_cohen_d}
\title{Adjust Cohen's d for measurement error}
\usage{
a4me_cohen_d(d, n1, n2, rxx = 1, ryy = 1)
}
\arguments{
\item{d}{(num) Cohen's d.}

\item{n1}{(int) Sample size for group 1.}

\item{n2}{(int) Sample size for group 2.}

\item{rxx}{(num) Reliability of x.}

\item{ryy}{(num) Reliability of y.}
}
\description{
Adjust' Cohen's d for measurement error using the Pearson r approach
}
\details{
Convert the d to Pearson r, then applies Spearman's correction, and then converts back.
}
\examples{
a4me_cohen_d(1, 100, 100) #no adjustment
a4me_cohen_d(1, 100, 100, rxx = .5) #adjust based on x
a4me_cohen_d(1, 100, 100, ryy = .5) #based on y, same result
a4me_cohen_d(1, 100, 100, rxx = .5, ryy = .5) #based on both; large effect!
a4me_cohen_d(1, 100, 100, rxx = 0) #infinite d if no reliability
a4me_cohen_d(10, 100, 100, rxx = .1, ryy = .1) #infinite also if corrected correlation exceeds the bounds
}
