% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_column_affix}
\alias{add_column_affix}
\title{Add affix to column names.}
\usage{
add_column_affix(data, prefix, suffix)
}
\arguments{
\item{data}{(an object) An object whose colnames should be changed}

\item{prefix}{(str) A prefix to add.}

\item{suffix}{(str) A suffix to add.}
}
\description{
Add a prefix to column names on data.frame or matrix or similar.
}
\examples{
test_iris = iris[1:10, ] #small test dataset
add_column_affix(test_iris, prefix = "P_") #ad P_ prefix
colnames(test_iris)
add_column_affix(test_iris, suffix = "_S") #ad _S suffix
colnames(test_iris)
#one can also use assign
test_iris2 = add_column_affix(iris, prefix = "A_", suffix = "_B")
colnames(test_iris2)
}

