% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirt.R
\name{add_item_associations_with_criterion_var}
\alias{add_item_associations_with_criterion_var}
\title{Calculate item associations with a criterion variable}
\usage{
add_item_associations_with_criterion_var(
  .item_data,
  .data,
  criterion_var,
  control_vars = NULL,
  reverse_negative_loadings = T,
  winsorize_r2_adj = T
)
}
\arguments{
\item{.item_data}{A data frame with item data from \code{get_mirt_stats()}}

\item{.data}{A data frame with the case-level data. Must have the items, the criterion variable, and any control variables.}

\item{criterion_var}{The name of the criterion variable}

\item{control_vars}{A vector of names of control variables to use in linear regression.}

\item{reverse_negative_loadings}{Whether to reverse the sign of the loadings, correlations, and model incremental R if the loading is negative. Default is TRUE.}

\item{winsorize_r2_adj}{Whether to winsorize the incremental R2 values to avoid negative values (probelmatic with square root transformation). Default is TRUE.}
}
\value{
The input data frame with added columns for the correlation with the criterion variable and the incremental R if control variables are given.
}
\description{
For using Jensen's method, one needs to calculate the correlation between each item and the criterion variable. This function does that, and also calculates the incremental R for each item if control variables are given.
}
