% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_newlines}
\alias{add_newlines}
\title{Insert newlines into text every nth character.}
\usage{
add_newlines(x, line_length = 95)
}
\arguments{
\item{x}{(chr vector) The strings to split with newlines.}

\item{line_length}{(num scalar) The desired max length of each line. Defaults to 95 (suitable for ggplot2).}
}
\description{
Returns a character string with newlines every nth character. Works for character vectors too.
}
\examples{
set.seed(2)
add_newlines(paste0(sample(c(letters, " "), size = 100, replace = T), collapse = ""))
}

