% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{add_spatial_knn}
\alias{add_spatial_knn}
\title{Add spatial k nearest neighbor (sknn) values}
\usage{
add_spatial_knn(x, vars, k = 3, lon = "lon", lat = "lat", suffix = "_lag")
}
\arguments{
\item{x}{A data frame with variables for lon and lat}

\item{vars}{A vector of variables to regress with sknn}

\item{k}{The value for 3, default 3}

\item{lon}{The lon variable}

\item{lat}{The lat variable}

\item{suffix}{The suffix to use, default "_lag"}
}
\value{
A data frame with added variables
}
\description{
Add spatial k nearest neighbor (sknn) values
}
