% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{adj_d_reliability}
\alias{adj_d_reliability}
\title{Adjust Cohen's d for measurement error}
\usage{
adj_d_reliability(x, rel)
}
\arguments{
\item{x}{A vector of d values}

\item{rel}{A vector of reliabilities}
}
\value{
A vector of adjusted d values
}
\description{
Adjust' Cohen's d for measurement error using the Pearson r approach
}
\examples{
adj_d_reliability(1.00, 0.8)
}
