% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{all_different}
\alias{all_different}
\title{Are all elements of a vector or list different?}
\usage{
all_different(x)
}
\arguments{
\item{x}{A vector or list to test.}
}
\value{
A boolean scalar.
}
\description{
Are all elements of a vector or list different?
}
\examples{
all_different(rep(1, 100))
all_different(c(1:5, 5))
all_different(c(NA, NA))
all_different(1:5)
all_different(list(1:2, 1:3, 1:4))
all_different(list(1:2, 1:3, 1:4, NA))
all_different(list(1:2, 1:3, 1:4, 1:4))
}
