% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{all_elements_the_same}
\alias{all_elements_the_same}
\title{Check whether all elements of a list are identical}
\usage{
all_elements_the_same(list)
}
\arguments{
\item{list}{(list) The list to check.}
}
\value{
TRUE if all elements are identical, FALSE if not.
}
\description{
Check whether all elements of a list are identical.
}
\examples{
#test lists
l_testlist_ok = list(1:3, 1:3, 1:3, 1:3, 1:3, 1:3)
l_testlist_bad = list(1:3, 1:3, 1:4, 1:3, 1:3, 1:3)
all_elements_the_same(l_testlist_ok) #TRUE
all_elements_the_same(l_testlist_bad) #FALSE
all_elements_the_same(list()) #TRUE if list is empty
}

