% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{alternate}
\alias{alternate}
\title{Merge vectors by alternating elements.}
\usage{
alternate(x)
}
\arguments{
\item{x}{(a list of vectors) The list of vectors.}
}
\description{
Inputs a list of equal length vectors, outputs a vector which is the merged vector by picking elements from each vector in alternating fashion.
}
\examples{
alternate(list(1:3, letters[1:3]))
}

