% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{as_abbrev}
\alias{as_abbrev}
\title{Abbreviate country and regional names to ISO-3.}
\usage{
as_abbrev(names, georgia = "country")
}
\arguments{
\item{names}{A character vector of the full names of countries and regions.}
}
\description{
To enable easier merging of datasets of international data. You need to download the countrylist.csv file yourself.
}
\examples{
as_abbrev()
}
\keyword{ISO}
\keyword{abbreviate,}
\keyword{names,}
\keyword{shorten;}

