% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{as_abbrev2}
\alias{as_abbrev2}
\title{Abbreviate country and regional names to ISO-3. For megadataset version >=3.}
\usage{
as_abbrev2(x, mega, georgia = "country", miss_msg = T)
}
\arguments{
\item{x}{A character vector of the full names of countries and regions to abbreviate.}

\item{mega}{Full path to the mega file. This is the .xlsx file you downloaded above.}

\item{georgia}{Whether to treat georgia as a country or US state. Defaults to country. Use "state" to use as state.}

\item{miss_msg}{Whether to output messages when abbreviations are missing. Defaults to T.}
}
\description{
To enable easier merging of datasets of international and regional data. Data file can be downloaded from https://osf.io/zdcbq/files/
}
\examples{
as_abbrev2()
}
\keyword{ISO}
\keyword{abbreviate,}
\keyword{names,}
\keyword{shorten;}

