% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as_decimal_year}
\alias{as_decimal_year}
\title{Convert date to decimal year}
\usage{
as_decimal_year(x)
}
\arguments{
\item{x}{A date vector}
}
\value{
A decimal year vector
}
\description{
Convert date to decimal year
}
\examples{
as_decimal_year(as.Date("2020-01-01"))
today() |> as_decimal_year() |>  decimal_year_to_date()
}
