% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{as_long}
\alias{as_long}
\title{Get full country names from ISO-3.}
\usage{
as_long(x)
}
\arguments{
\item{x}{(character vector) The ISO-3 codes.}
}
\description{
To enable easier merging of datasets of international data. You need to download the countrylist.csv file yourself.
}
\examples{
as_long()
}
\keyword{ISO}
\keyword{names,}

