% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{as_num_df}
\alias{as_num_df}
\title{Convert convertible columns in a data.frame to numeric.}
\usage{
as_num_df(df, stringsAsFactors = F, skip_factors = T, remove_commas = T)
}
\arguments{
\item{df}{A data.frame.}

\item{stringsAsFactors}{Whether to convert strings to factors. Default is F.}

\item{skip_factors}{Whether to skip factors. Default is T.}

\item{remove_commas}{Whether to remove commas from the cells first. If present, they will make the conversion fail. Defaults to T.}
}
\description{
Attempts to convert all columns in a data.frame to numeric. If that fails, keeps the original.
}
\examples{
iris_chr = as.data.frame(lapply(iris, as.character)) #convert iris to strings
as_num_df(iris_chr) #convert back
}

