% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{as_num_matrix}
\alias{as_num_matrix}
\title{Convert a data.frame to a numeric matrix, including factors.}
\usage{
as_num_matrix(df, skip_chr = T)
}
\arguments{
\item{df}{(data.frame) A data.frame with variables.}

\item{skip_chr}{(log scalar) Whether to skip character columns (default). If false, they are converted to non-ordered factors.}
}
\description{
Returns a numeric matrix. Ordered factors are converted to numbers, while non-ordered factors are split into dummy variables using the first level as the the reference.
}
\details{
Factors with only two levels are kept as they are.
}
\examples{
as_num_matrix(iris)
}

