% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{assert_installed}
\alias{assert_installed}
\title{Assert that packages are installed without attaching them}
\usage{
assert_installed(...)
}
\arguments{
\item{...}{Unquoted package names separated by comma}
}
\value{
On success invisibly returns TRUE, otherwise an error.
}
\description{
Assert that packages are installed without attaching them
}
\examples{
assert_installed(ggplot2)
testthat::expect_error(assert_installed(ggplot2, ggplot3))
}
