% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{averages}
\alias{averages}
\title{Averages}
\usage{
averages(x, trim = 0.1, types = c("arithmetic", "geometric", "harmonic",
  "mode", "median", "trimmed", "midrange"))
}
\arguments{
\item{x}{(num vectr) A vector of values.}

\item{trim}{(num vectr) Used for trimmed means. Default = .10.}

\item{types}{(chr vectr) Which types of averages to calculate?}
}
\description{
Calculates a number of different types of averages for a vector.
}
\examples{
#simualte some data
set.seed(1)
x = sample(1:10, replace = T, size = 1000, prob = seq(.1, .5, length.out = 10))
#look at it
GG_denhist(x, vline = NULL)
#calcualte averages
averages(x)
#normal data
set.seed(1)
averages(rnorm(1000)) #some functions give warnings when there are negative numbers
}
