% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{browse_if}
\alias{browse_if}
\title{Browse on condition}
\usage{
browse_if(expr, condition)
}
\arguments{
\item{expr}{(expr) An expression.}

\item{condition}{(expr) A condition to evaluate. y. is the output of expr.}
}
\description{
Browse on condition. Calls browse in the parent.frame so that one can see the context. Made for easier debugging. Pipe-friendly: returns output if condition is false.
}
\examples{
browse_if(1+1, .y == 2)
}

