% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_object.R
\name{cache_object}
\alias{cache_object}
\title{Cache object}
\usage{
cache_object(expr, filename, renew = F, ...)
}
\arguments{
\item{expr}{An expression}

\item{filename}{A filename to use}

\item{renew}{Whether to force renewal of the object}

\item{...}{Other arguments passed to \code{\link[readr:read_rds]{readr::write_rds()}}}
}
\value{
An object
}
\description{
Cache object
}
\examples{
i = cache_object(MASS::mvrnorm(1e6, rep(0, 3), matrix(c(1, .4, .4, .4, 1, .4, .4, .4, 1), ncol = 3)), "test_object.rds")
i = cache_object(stop("This is not evaluated"), "test_object.rds")
i = cache_object(warning("But this is"), "test_object.rds", renew = T)
}
