% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirt.R
\name{calc_item_gaps}
\alias{calc_item_gaps}
\title{Calculate item gaps}
\usage{
calc_item_gaps(x, group, return_data_frame = T)
}
\arguments{
\item{x}{Item data frame}

\item{group}{A grouping variable}

\item{return_data_frame}{Whether you want a data frame back, or just a vector}
}
\value{
a data frame or a vector
}
\description{
Calculate item gaps
}
\examples{
set.seed(1)
X = matrix(rbinom(10000, 1, .5), ncol = 10)
group = rbinom(1000, 1, .5)
calc_item_gaps(X, group)
}
