% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calc_row_representativeness}
\alias{calc_row_representativeness}
\title{Calculate row-wise representativeness}
\usage{
calc_row_representativeness(
  x,
  central_tendency_fun = median,
  central_tendencies = NULL,
  standardize = T,
  absolute = T
)
}
\arguments{
\item{x}{Data frame of numerical data}

\item{central_tendency_fun}{Function to use for central tendency}

\item{central_tendencies}{If custom values for central tendency, input here}

\item{standardize}{Whether to standardize values to ensure common metric. If not done, variables with larger SDs will be given more influence.}

\item{absolute}{Whether to use absolute errors (if not, then there is also an attempt to ensure the error directions are balanced across variables)}
}
\value{
A data frame of error values and their mean and medians
}
\description{
Calculate row-wise representativeness
}
\examples{
calc_row_representativeness(iris)
}
