% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{check_if_in}
\alias{check_if_in}
\title{Check whether object is in a list/vector}
\usage{
check_if_in(x, list)
}
\arguments{
\item{x}{(any object) The object to test.}

\item{list}{(list) The list of accepted values}
}
\value{
Logical scalar or error.
}
\description{
Checks whether an object is in a list/vector. If not, returns an informative error.
}
\examples{
check_if_in("a", letters[1:10])
check_if_in("a", letters[2])
}

