% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{check_missing}
\alias{check_missing}
\title{Check if arguments are missing and raise error if they are}
\usage{
check_missing(var_names,
  error_msg = "[VAR] was missing! Please supply the input and try again.")
}
\arguments{
\item{str}{(a character scalar) A string of a half math conditional}

\item{convert_equal}{(boolean) Converts = to ==. Default=T.}

\item{silent_try}{(boolean) Whether to use a silent trial. Default=T. Change to F to get potentially useful debugging information.}

\item{test_function}{(boolean) Whether to test the function. This is done by trying on 0. Default=T.}
}
\description{
Checks if arguments are missing and raises an error if they are. Put this in the beginning of your functions to check the input and give useful errors without writing checking code over and over again.
}
\examples{
math_to_function()
}

