% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC_helper.R
\name{check_spatial_input}
\alias{check_spatial_input}
\title{Check for input for spatial data.}
\usage{
check_spatial_input(df, dists = NULL, lat_var = NULL, lon_var = NULL,
  distance_method = NULL, auto_detect_dist_method = T)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{dists}{A matrix of distances between cases.}

\item{lat_var}{A string with the name of the variable which has the latitude/east-west data.}

\item{lon_var}{A string with the name of the variable which has the longitude/north-south data.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean. If using euclidean it doesn't matter which variable is coded as lat or lon.}

\item{auto_detect_dist_method}{Whether to try to autodetect the distance method. If the dataset contains variables with the names "lat" and "lon", it will be detected as spherical. If it contains "x" and "y", it will be detected as euclidean. Defaults to true.}
}
\description{
Returns a list with the results of the check. This is generally only used as a helper function for other spatial statistics functions.
}
