% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{combine_upperlower}
\alias{combine_upperlower}
\title{Combine upper and lower part of a matrix.}
\usage{
combine_upperlower(.upper.tri, .lower.tri, .diag = NA)
}
\arguments{
\item{.upper.tri}{a matrix or data.frame for the upper trianagle.}

\item{.lower.tri}{a matrix or data.frame for the lower trianagle.}

\item{.diag}{a value to insert in the diagonal. Defaults to NA.}
}
\description{
Returns a matrix composed of an upper and a upper part.
}
\examples{
combine_upperlower()
}
\keyword{combine,}
\keyword{data.frame,}
\keyword{lower}
\keyword{matrix,}
\keyword{upper,}

