% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{compare_predictors}
\alias{compare_predictors}
\title{Fit linear models for a set of predictors to compare their effects alone and together}
\usage{
compare_predictors(data, outcome, predictors, conf.level = 0.95)
}
\arguments{
\item{data}{The data frame to use}

\item{outcome}{The name of the outcome variable. Must be numeric.}

\item{predictors}{A character vector of predictors to use.}

\item{conf.level}{The confidence level to use. Default is .95.}
}
\value{
A data frame with the coefficients for each model
}
\description{
This function fits a set of linear models to the data, where each model includes only one predictor. It then fits a full model with all predictors. The function returns a data frame with the coefficients for each model, as well as the full model.
}
\examples{
compare_predictors(iris, names(iris)[1], names(iris)[-1])
compare_predictors(mpg, names(mpg)[3], names(mpg)[-3])
}
