% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{conditional_change}
\alias{conditional_change}
\title{Change values conditionally}
\usage{
conditional_change(x, func = NULL, func_str = NULL, new_value,
  handle_NA = T)
}
\arguments{
\item{x}{(vector, data.frame, matrix or list) An object whose values to change.}

\item{func}{(function) A function that returns a boolean.}

\item{func_str}{(character scalar) A string to make a function from using math_to_function().}

\item{new_value}{(scalar vector) A value to use in the cases.}

\item{handle_NA}{(boolean) Whether to handle missing values.}
}
\description{
Change values based on a condition given as a function.
}
