% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{conf_interval_width}
\alias{conf_interval_width}
\title{Compute width of confidence interval}
\usage{
conf_interval_width(se, n = Inf, confidence_level = 0.95, single_arm = T)
}
\arguments{
\item{se}{(numeric scalar) The standard error.}

\item{n}{(numeric scalar) The sample size. Defaults to Inf.}

\item{confidence_level}{(numeric scalar) The confidence level. Defaults to .95.}

\item{single_arm}{(logical scalar) Whether to calculate the width for a single arm. Defaults to TRUE.}
}
\value{
The width of the confidence interval.
}
\description{
Calculate the width of a confidence interval given a standard error and sample size. Based on the t distribution.
}
\examples{
conf_interval_width(1) #about 1.96
conf_interval_width(1, confidence_level = .99) #about 2.58
}
