% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{copy_columns}
\alias{copy_columns}
\title{Copy columns between data.frames}
\usage{
copy_columns(from, to, columns, pattern)
}
\arguments{
\item{from}{(data.frame) The source data.frame.}

\item{to}{(data.frame) The destination data.frame.}

\item{columns}{(character vector or numeric vector) The columns to copy. Can be either their names or whole numbers indicating their position in the source data.frame. By default, it will copy all columns.}

\item{pattern}{(character scalar) Alternatively, a regex pattern to use to match the desired columns. Uses the stringr package.}
}
\description{
Copy columns from one data.frame to another by name or pattern.
}
\examples{
merge_rows_by_name(df = t1, names = c("C", "D"), func = mean)
}

