% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{copy_names}
\alias{copy_names}
\title{Copy names from one object to another.}
\usage{
copy_names(x, y, partialmatching = T)
}
\arguments{
\item{x}{(an object) An object whose dimnames should be copied.}

\item{y}{(an object) An object whose dimensions that should be renamed.}
}
\description{
Attempts to copy names that fit the dimensions of vectors, lists, matrices and data.frames.
}
\examples{
m = matrix(1:9, nrow=3)
n = m
rownames(m) = letters[1:3]
colnames(m) = LETTERS[1:3]
copy_names(m, n)
n
}
\keyword{colnames,}
\keyword{copy}
\keyword{names,}
\keyword{rownames,}

