% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{cor_CI_from_SE}
\alias{cor_CI_from_SE}
\title{Compute correlation confidence intervals from correlations and standard errors}
\usage{
cor_CI_from_SE(cor, se, ci = 0.95, winsorise = T)
}
\arguments{
\item{cor}{A correlation value or matrix}

\item{se}{A standard error value or matrix}

\item{ci}{The desired confidence level}

\item{winsorise}{Keep values within -1 to 1}
}
\value{
If you entered a single value, you get 2 values back: lower and upper confidence numbers. If you entered matrices, you get 2 matrices back in a list.
}
\description{
Compute correlation confidence intervals from correlations and standard errors
}
\examples{
#simple results
cor_CI_from_SE(.50, .10)
cor_CI_from_SE(.50, .10, ci = .99)
#matrices
iris_res = suppressWarnings(weights::wtd.cor(iris[-5]))
cor_CI_from_SE(iris_res$correlation, iris_res$std.err)
}
