% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{cor_matrix_weights}
\alias{cor_matrix_weights}
\title{Calculate a correlation matrix with and without weights.}
\usage{
cor_matrix_weights(df, weight_var)
}
\arguments{
\item{df}{A data.frame.}

\item{weight_var}{A character string of the name of the weights variable.}
}
\description{
Inputs a data.frame and a set of weights. Outputs a correlation matrix where the lower triangle are weighted correlations and the upper triangle are unweighted. Diagonals are set as NA. The weights variable is excluded from the matrix.
}
\examples{
cor_matrix_weights()
}
\keyword{correlation,}
\keyword{matrix,}
\keyword{weights}

