% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{count_decimals}
\alias{count_decimals}
\title{Count decimals}
\usage{
count_decimals(x)
}
\arguments{
\item{x}{Vector of values}
}
\value{
Vector of values
}
\description{
Count decimals
}
\examples{
c(1, 1.2, 1.23) \%>\% count_decimals()
c() \%>\% count_decimals()
}
