% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{count_NA}
\alias{count_NA}
\title{Count missing data}
\usage{
count_NA(x, reverse = F)
}
\arguments{
\item{x}{(any suitable object) An object for which to count NAs.}

\item{reverse}{(logical scalar) Whether to count non-NAs instead (default false).}
}
\description{
A simple wrapper for is.na() and sum(). Returns an integer.
}
\examples{
m = matrix(c(NA, 1:3, NA, 4:6, NA), nrow=3)
count_NA(m)
}

