% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{decimal_year_to_date}
\alias{decimal_year_to_date}
\title{Convert decimal year to date}
\usage{
decimal_year_to_date(decimal_year)
}
\arguments{
\item{decimal_year}{A decimal year vector}
}
\value{
A date vector
}
\description{
Convert decimal year to date
}
\examples{
decimal_year_to_date(2020.5)
today() |> as_decimal_year() |>  decimal_year_to_date()
}
