% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{describe2}
\alias{describe2}
\title{Describe data}
\usage{
describe2(...)
}
\arguments{
\item{...}{Inut for \link[psych:describe]{psych::describe}}
}
\value{
Tibble subset
}
\description{
Describe data
}
\examples{
describe2(iris)
describe2(mpg)
}
