% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_add_affix}
\alias{df_add_affix}
\title{Add affix to column names.}
\usage{
df_add_affix(data, prefix = "", suffix = "", exclude = c())
}
\arguments{
\item{data}{(an object) An object whose colnames should be changed}

\item{prefix}{(chr) A prefix to add.}

\item{suffix}{(chr) A suffix to add.}

\item{exclude}{(chr) Variables to exclude.}
}
\description{
Add a prefix to column names on data.frame or matrix or similar.
}
\examples{
df_add_affix(iris, prefix = "P_") |> head() #add P_ prefix
df_add_affix(iris, suffix = "_S") |> head() #add _S suffix
df_add_affix(iris, suffix = "_X", exclude = "Species") |> head() #add _X suffix but not for Species
}
