% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_add_id}
\alias{df_add_id}
\title{Add ID column to data.frame}
\usage{
df_add_id(data, id, id_var = "ID")
}
\arguments{
\item{data}{(data.frame) The data.frame.}

\item{id}{(vector) A vector of values to use. Will be repeated silently if not long enough. Will throw an error if it's not possible.}

\item{id_var}{(character scalar) What to call the id column (default "ID").}
}
\description{
Add an ID column to a data.frame. Convenience function.
}
\examples{
head(df_add_id(iris, "A")) #Add an ID column named "ID" filled with "A"
}
