% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_as_num}
\alias{df_as_num}
\title{Convert convertible columns in a data.frame to numeric.}
\usage{
df_as_num(
  df,
  stringsAsFactors = F,
  smart_factor_conversion = T,
  always_skip_factors = F,
  remove_commas = T,
  skip_factors = NULL
)
}
\arguments{
\item{df}{A data.frame.}

\item{stringsAsFactors}{Whether to convert strings to factors. Default is F.}

\item{smart_factor_conversion}{Whether to automatically detect numeric factors. Default=T. This is done by checking whether the levels can be converted to numeric values. If they can, the factor is converted.}

\item{always_skip_factors}{Whether to always skip non-ordered factors. Default=F.}

\item{remove_commas}{Whether to remove commas from the cells first. If present, they will make the conversion to numeric fail. Defaults to T.}

\item{skip_factors}{Depreciated.}
}
\description{
Attempts to convert all columns in a data.frame to numeric. If that fails, keeps the original.
}
\examples{
iris_chr = as.data.frame(lapply(iris, as.character)) #convert iris to strings
str(iris_chr)
str(df_as_num(iris_chr)) #convert back
all(iris == df_as_num(iris_chr)) #confirm identity
}
