% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_copy_columns}
\alias{df_copy_columns}
\title{Copy columns between data.frames}
\usage{
df_copy_columns(from, to, columns, pattern)
}
\arguments{
\item{from}{(data.frame) The source data.frame.}

\item{to}{(data.frame) The destination data.frame.}

\item{columns}{(chr vec or num vec) The columns to copy. Can be either their names or whole numbers indicating their position in the source data.frame. By default, it will copy all columns.}

\item{pattern}{(chr scl) Alternatively, a regex pattern to use to match the desired columns. Uses the stringr package.}
}
\description{
Copy columns from one data.frame to another by name or pattern.
}
\examples{
t = data.frame(x = 1:5, y = 6:10); t
t2 = data.frame(a = letters[1:5], b = LETTERS[1:5]); t2
df_copy_columns(t, t2) #same as cbind(t, t2)
df_copy_columns(t, t2, columns = 1) #same as cbind(t[1], t2)
df_copy_columns(t, t2, pattern = "[^x]") #copy columns that doesn't have x in their name. complicated in base-r!
}

