% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_flexsubset}
\alias{df_flexsubset}
\title{Subset a data.frame flexibilly}
\usage{
df_flexsubset(data, vars, messages = T)
}
\arguments{
\item{data}{(data.frame) Data to subset.}

\item{vars}{(chr vectr) Variables to subset.}
}
\value{
Returns a data.frame that is a subset of the original.
}
\description{
Subsets a data.frame by a vector of variables given as names. Only subsets the ones found in the data.
}
\details{
Subsets the intersection of vars and the columns in the data. Throws a warning if no overlap. Subsets using drop=F.
}
\examples{
df_flexsubset(iris, c("Species")) \%>\% str
df_flexsubset(iris, c("Species", "Sepal.Length")) \%>\% str
df_flexsubset(iris, c("Species", "test")) \%>\% str
df_flexsubset(iris, c("test")) \%>\% str
}

