% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{df_func}
\alias{df_func}
\title{Calculate summary statistics by row from multiple columns in a data.frame.}
\usage{
df_func(..., standardize = F, func = mean, pattern, ignore_NA = T)
}
\arguments{
\item{...}{(data.frames, matrices, vectors) The variables to use. They will be coerced to a single data.frame.}

\item{standardize}{(boolean) Whether to standardize the data before calculating. Defaults to F.}

\item{func}{(function) Which base function to call. Can be mean, median, sd, var and any other suitable function. Default to mean.}

\item{pattern}{(string) A pattern to use for finding the columns names.}

\item{ignore_NA}{(boolean) Whether to ignore missing data. Defaults to T.}
}
\description{
Calculate mean/median/sd/etc values by row. Can be given multiple data.frames, matrices or vectors which are coerced into one data.frame. Can standardize data before calculating. Ignores missing data by default. Can also subset columns from a data.frame using regex of the colnames.
}
\examples{
df_func()
}
\keyword{date.frame,}
\keyword{function}
\keyword{mean,}
\keyword{median,}
\keyword{standardize,}

