% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_gather_by_pattern}
\alias{df_gather_by_pattern}
\title{Gather by pattern}
\usage{
df_gather_by_pattern(data, pattern, key_col = ".varying", id_col = ".id",
  method = "pure_tidyr")
}
\arguments{
\item{data}{(data.frame) A data.frame to tidy.}

\item{pattern}{(chr sclr) The regex pattern to group variables by.}

\item{varying_name}{(chr sclr) A name to use for the column of varying part of the name, e.g. the year. Defaults to ".varying".}
}
\value{
A tidy data.frame.
}
\description{
Makes a data.frame tidyr by gathering columns that are variants of each other, such as for different years.
}

