% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_merge_cols}
\alias{df_merge_cols}
\title{Merge columns}
\usage{
df_merge_cols(df, cols)
}
\arguments{
\item{df}{(df) Data frame.}

\item{cols}{(chr) Variable names to merge.}
}
\value{
A vector.
}
\description{
Combines columns by filling in missing values with values from other columns in a designated order.
}
\examples{
data_frame(
a = c(1, NA, NA),
b = c(-1, 2, NA),
c = c(-5, -9, 3)
) \%>\% df_merge_cols(letters[1:3])
}
