% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_no_list_cols}
\alias{df_no_list_cols}
\title{Remove list columns}
\usage{
df_no_list_cols(x)
}
\arguments{
\item{x}{Data frame}
}
\value{
Data frame
}
\description{
Remove list columns
}
\examples{
iris$list = list(1)
iris$df = list(iris)
map(iris, typeof)
map(df_no_list_cols(iris), typeof)
}
