% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_remove}
\alias{df_remove}
\alias{df_remove_vars}
\title{Remove variables from a data.frame}
\usage{
df_remove(data, vars)
}
\arguments{
\item{data}{(data.frame) A data.frame.}

\item{vars}{(vector) The variables to remove. Can be logical, character or integer.}
}
\description{
Remove variables from a data.frame by name, position or logical.
}
\examples{
#remove two variables by name
#does not work in base r by expected syntax:
#iris[-c("Sepal.Length", "Species")]
names(df_remove(iris, c("Sepal.Length", "Species")))
}

