% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_remove_NA_vars}
\alias{df_remove_NA_vars}
\title{Remove NA columns}
\usage{
df_remove_NA_vars(data, keep = "")
}
\arguments{
\item{data}{(data.frame) Data to subset.}

\item{keep}{(num or chr vector) A vector of names of columns to keep no matter what. Can also be numeric indices which are then replaced with the colnames.}
}
\value{
Returns the subset of the object.
}
\description{
Detect variables with no data in a data.frame and remove them. Can be given a vector of names or indices of columns to keep no matter what.
}
\examples{
df_remove_NA_vars(data.frame(a = NA, b = 1))
df_remove_NA_vars(data.frame(a = NA, b = 1), keep = "a")
df_remove_NA_vars(data.frame(a = NA, b = 1), keep = 1)
}

