% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_remove_vars}
\alias{df_remove_vars}
\title{Rename variables from a data.frame by name}
\usage{
df_remove_vars(data, names)
}
\arguments{
\item{data}{(data.frame) The data.frame.}

\item{names}{(chr vector) The variables to remove.}
}
\description{
Remove variables from a data.frame by name.
}
\examples{
str(df_remove_vars(iris, c("Sepal.Length", "Species")))
}

