% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_rename}
\alias{df_rename}
\title{Rename variables in a data.frame}
\usage{
df_rename(data, current_names, new_names)
}
\arguments{
\item{data}{(data.frame) The data.frame.}

\item{current_names}{(chr vector) The current names of the variables.}

\item{new_names}{(chr scalar) The new names of the variables.}
}
\description{
Rename variables in a data.frame and keep their positions.
}
\examples{
#rename one variable
names(df_rename(iris, "Sepal.Length", "Sepal_Length"))
#rename multiple variables
names(df_rename(iris, c("Sepal.Length", "Species"), c("SEPAL_LENGTH", "SPECIES")))
#randomly rename iris
df_rename(iris, current_names = names(iris), new_names = rev(names(iris)[sample(1:5)])) \%>\% names
#warning on non-existent names
head(df_rename(iris, "bleh", "blah"))
}
