% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_rename_vars}
\alias{df_rename_vars}
\title{Rename variables in a data.frame}
\usage{
df_rename_vars(data, current_names, new_names)
}
\arguments{
\item{data}{(data.frame) The data.frame.}

\item{current_names}{(chr vector) The current names of the variables.}

\item{new_names}{(chr scalar) The new names of the variables.}
}
\description{
Rename variables in a data.frame and keep their positions.
}
\examples{
str(df_rename_vars(iris, "Sepal.Length", "Sepal_Length"))
}

