% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_reorder_columns}
\alias{df_reorder_columns}
\title{Reorder columns in a data.frame by name}
\usage{
df_reorder_columns(data, vars)
}
\arguments{
\item{data}{(data.frame) The data frame.}

\item{vars}{(named vector) The variables to reorder.}
}
\value{
The modified data frame.
}
\description{
Reorder columns in a data.frame by name using a named vector.
}
\details{
Copied and modified from http://stackoverflow.com/a/37009127/3980197.
}
\examples{
#remove Species to front
head(df_reorder_columns(iris, c("Species" = 1)))
#check for identity
all(df_reorder_columns(iris, c("Species" = 1)) == iris[c(5, 1:4)])
#move multiple
head(df_reorder_columns(iris, c("Species" = 1, "Petal.Length" = 2)))
#throws error if not given a named vector
throws_error("df_reorder_columns(iris, 1)")
#or if names are not there
throws_error("df_reorder_columns(iris, c('abc' = 1))")
#throws warning if one tries to move the same multiple times as this is probably not intended
throws_error('df_reorder_columns(iris, c("Species" = 1, "Species" = 2))')
}
