% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_residualize}
\alias{df_residualize}
\alias{residualize_DF}
\title{Residualized data.frame.}
\usage{
df_residualize(data, resid.vars, suffix = "", exclude.resid.vars = T,
  return.resid.vars = T, print.models = T, exclude_vars = NULL)
}
\arguments{
\item{data}{(data.frame or matrix) The data object.}

\item{resid.vars}{(character vector) The names of the variables to use for the residualization.}

\item{exclude.resid.vars}{(logical) Whether to exclude the residualization variables from residualization. Defaults to TRUE.}

\item{return.resid.vars}{(logical) Whether to include the residualization variables in the returned data.frame. Defaults to TRUE.}

\item{print.models}{(logical) Wether to print the lm models used in the process. Defaults to TRUE.}

\item{exclude_vars}{(character vector) Names of variables are that excluded from the residualization.}
}
\description{
Returns a residualized data.frame given a set of variables to partial out.
}
\examples{
df = data.frame(a = rnorm(5), b = rnorm(5), c = rnorm(5))
df_residualize(df, resid.vars = "c")
}
\keyword{modeling,}
\keyword{partialing}
\keyword{residualize,}

