% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_round}
\alias{df_round}
\title{Round numeric variables of a data frame.}
\usage{
df_round(df, digits = 3)
}
\arguments{
\item{df}{A data.frame or matrix.}

\item{digits}{The number of digits to round to.}
}
\description{
Returns a data.frame where numeric variables have been rounded to the desired number of digits.
}
\examples{
head(df_round(iris)) #round isis. Automatically skips non-numeric columns.
head(df_round(iris, 0)) #round isis to 0 digits.
}

