% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_sort}
\alias{df_sort}
\title{Sort a data.frame by one or more variables. Can sort by decreasing or increasing order}
\usage{
df_sort(df, vars = names(df), decreasing = F)
}
\arguments{
\item{df}{(df) A data.frame.}

\item{vars}{(str/int) variables to use for sorting.}

\item{decreasing}{Whether to use decreasing order. Default=F.}
}
\description{
Sort a data.frame by one or more variables. Can sort by decreasing or increasing order. Improvement of the reshape::order because it preserves rownames and can sort by decreasing order as well.
}
\examples{
head(df_sort(iris, 1)) #sort by the first variable, increasing order
head(df_sort(iris, 1, decreasing = T)) #sort by the first variable, decreasing order
head(df_sort(iris, "Species")) #sort by Species variable
}
