% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_subset_by_pattern}
\alias{df_subset_by_pattern}
\title{Subset by pattern}
\usage{
df_subset_by_pattern(data, pattern, inverse = FALSE)
}
\arguments{
\item{data}{(data.frame or matrix) Object to subset.}

\item{pattern}{(character scalar) A regex pattern.}

\item{inverse}{(logical scalar) Whther to keep the non-matches instead.}
}
\value{
Returns the subset of the object.
}
\description{
Subset a data.frame or matrix by a pattern in the column names. A simple wrapper using str_detect().
}
\details{
Subsets using [] and with drop=FALSE.
}
\examples{
df_subset_by_pattern(iris, "Length") # length columns
df_subset_by_pattern(iris, "Length", T) # non-length columns
}
