% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_to_ldf}
\alias{df_to_ldf}
\title{Transform data.frame to a list of data.frames}
\usage{
df_to_ldf(data, by, remove_by = T)
}
\arguments{
\item{data}{(data.frame) The data.frame.}

\item{by}{(character scalar or vector) The values to split by. If input is a character scalar, will use that variable from the data.frame. If it is a longer vector, will use that to split with.}

\item{remove_by}{(logical scalar) Whether to remove the vector used to split with from the data.frame. Only relevant when using a character scalar as input to by.}
}
\description{
Splits a data.frame by a variable and returns a named list.
}
\examples{
df_to_ldf(iris, "Species")
}

