% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_to_v}
\alias{df_to_v}
\title{Transform data.frame to vector}
\usage{
df_to_v(data, fact_to_chr = T)
}
\arguments{
\item{data}{(data.frame) The data.frame.}

\item{fact_to_chr}{(logical scalar) Convert factors to characters? (Default yes)}
}
\description{
Appends each column after each other. A wrapper for unlist() with some extra features.
}
\examples{
df_to_v(iris)
}

