% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_var_table}
\alias{df_var_table}
\title{Make variable table}
\usage{
df_var_table(x)
}
\arguments{
\item{x}{Data frame}
}
\value{
Data frame
}
\description{
Make variable table
}
\examples{
df_var_table(iris)
}
