% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_winsorise}
\alias{df_winsorise}
\title{Winsorsise data}
\usage{
df_winsorise(x, variables = NULL, z = NULL, rz = NULL, centile = NULL)
}
\arguments{
\item{x}{Data frame}

\item{variables}{Names of variables to winsorsise. If not given, defaults to all numeric variables.}

\item{z}{Standard z score to winsorise to (mean+sd)}

\item{rz}{Robust z score to winsorise to (median+mad)}

\item{centile}{Centile to winsorsise to}
}
\value{
Data frame with the same variables as input.
}
\description{
Winsorsise data
}
\examples{
iris[1, 1] = 100
df_winsorise(iris, z = 2)
df_winsorise(iris, rz = 2)
df_winsorise(iris, centile = .99)
}
