% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{DIF_test}
\alias{DIF_test}
\title{Test for differential item functioning (DIF)}
\usage{
DIF_test(
  items,
  model,
  group,
  fscores_pars = list(full.scores = T, full.scores.SE = T),
  messages = T,
  method = "EM",
  technical = list(),
  itemtype = NULL,
  verbose = T,
  ...
)
}
\arguments{
\item{items}{Item data to use}

\item{model}{Item model}

\item{group}{Group (2 groups at most)}

\item{fscores_pars}{Any extra scoring parameters used}

\item{messages}{Show messages}

\item{method}{Method}

\item{technical}{Further technical args to pass to mirt}

\item{...}{Other arguments passed to mirt functions}
}
\value{
A list of results
}
\description{
Tests are done following the mirt package approach outlined by Chalmers.
}
