% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_operations.R
\name{discretize}
\alias{discretize}
\title{Discretize a variable}
\usage{
discretize(x, breaks, equal_range = T, labels = "numbers",
  include_end = T, right = T, ordered_factor = F, ...)
}
\arguments{
\item{x}{(numeric vector) A vector of datapoints to discretize.}

\item{breaks}{(integer scalar or numeric vector) The number of levels to create, or a numeric vector of values to use as cut-offs.}

\item{equal_range}{(logical scalar) Whether to use equal ranges. If false, will use equal sizes. Defaults to TRUE.}

\item{labels}{(character scalar) Which labels to use. By default, it will use the intervals. Can also be "midpoint", which uses the interval midpoints, "numbers" will use integers.}

\item{include_end}{(logical scalar) Whether to include datapoints equal to the last break value. Default=TRUE.}

\item{right}{(logical scalar) Whether to use intervals that are closed on the right. Default=TRUE.}

\item{ordered_factor}{(logical scalar) Should the result be an ordered factor? Default=FALSE.}

\item{...}{(additional arguments) Any other arguments passed to cut().}
}
\description{
A wrapper for base-r's cut() with some improved features.
}
\examples{
x = discretize(rnorm(100), 5)
hist(x)
y = discretize(rnorm(100), 5, equal_range = F)
hist(y)
}

