% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{encode_combinations}
\alias{encode_combinations}
\title{Encode combinations}
\usage{
encode_combinations(x, collapse = ", ")
}
\arguments{
\item{x}{A data frame of binary values, which will be converted to logicals.}
}
\value{
A chracter vector of encoded combinations
}
\description{
Given a data frame of binary values, encode the combinations of TRUE values. Useful for combining multiple yes/no options in different variables into a single column, e.g. self-identified race/ethnicity.
}
\examples{
tibble(A = c(T, F, F, T, F), B = c(F, T, F, F, F), C = c(F, F, T, F, F), D = c(F, F, F, T, F)) \%>\% encode_combinations()
}
