% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{extract_last}
\alias{extract_last}
\title{Subset object using Python's negative indexing}
\usage{
extract_last(x, margin_1, margin_2, drop = FALSE)
}
\arguments{
\item{x}{(vector, data.frame or matrix) Object to subset.}

\item{margin_1}{(integer vector) Indices for the first (vertical) margin.}

\item{margin_2}{(integer vector) Indices for the second (horizontal) margin.}

\item{drop}{(logical scalar) Whether to convert to a vector if there is only 1 column. Default=FALSE.}
}
\value{
Returns the subset of the object.
}
\description{
Subset object using Python's negative indexing, that is, by counting indices backwards.
}
\examples{
extract_last(iris, 1) # last row
extract_last(iris, , 1) # last column
extract_last(iris, 5:1, 2:1) #last 5 elements rows and last 2 columns
}

