% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{extract_nonnum_vars}
\alias{extract_nonnum_vars}
\title{Extract non-numerical variables}
\usage{
extract_nonnum_vars(data)
}
\arguments{
\item{data}{(data.frame or matrix) The data.}
}
\value{
Returns the subset of the data while keeping the type (using drop = FALSE).
}
\description{
Extract the non-numerical variables from a data.frame or matrix.
}
\examples{
extract_nonnum_vars(iris)
}

