% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{extract_num_vars}
\alias{extract_num_vars}
\title{Extract numerical variables}
\usage{
extract_num_vars(data)
}
\arguments{
\item{data}{(data.frame or matrix) The data.}
}
\value{
Returns the subset of the data while keeping the type (using drop = FALSE).
}
\description{
Extract the numerical variables from a data.frame or matrix.
}
\examples{
extract_num_vars(iris)
}
