% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_all_methods}
\alias{FA_all_methods}
\title{Factor analyze with all methods}
\usage{
FA_all_methods(DF, ..., skip_methods = "")
}
\arguments{
\item{DF}{A data.frame to extract factors from.}

\item{...}{Parameters to fa().}

\item{skip_methods}{A character vector of methods to skip. Defaults to none.}
}
\description{
Runs factor analysis on a dataset with all 30 possible combinations of extraction and scoring methods. Returns a list with all scores and loadings for further use.
}
\examples{
FA_all_methods()
}
\keyword{analysis}
\keyword{factor}
\keyword{psychometrics,}

