% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_CFS}
\alias{FA_CFS}
\title{Calculates the effect of removing a case on size of the first factor, for each case.}
\usage{
FA_CFS(data, sort = T)
}
\arguments{
\item{data}{a data.frame.}

\item{sort}{whether to sort the results or not. Defaults to true.}
}
\description{
Returns a numerical vector of the change in the size of the first factor by removing that case.
}
\examples{
FA_CFS()
}
\keyword{analysis}
\keyword{factor}
\keyword{latent}
\keyword{psychology,}
\keyword{psychometrics,}
\keyword{variable,}

