% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_congruence_matrix}
\alias{FA_congruence_matrix}
\title{Calculate congruence coefficient matrix.}
\usage{
FA_congruence_matrix(x)
}
\arguments{
\item{x}{A list of factor analysis objects from fa() or a data.frame/matrix of factor loadings.}
}
\description{
Takes an input either a list of factor analysis objects from fa() (psych package) or a data.frame/matrix of loadings. Returns a matrix of congruence coefficients. These are calculated using factor.congruence() from psych package.
}
\examples{
FA_congruence_matrix()
}
\keyword{congruence,}
\keyword{factor,}
\keyword{matrix}

