% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fa_loadings}
\alias{fa_loadings}
\title{Get loadings from a factor analysis object.}
\usage{
fa_loadings(fa)
}
\arguments{
\item{fa}{(object of class "fa") A factor analysis object.}
}
\description{
Extract loadings from a factor analysis object of class "fa") into a data.frame.
}
\examples{
iris_fa = psych::fa(iris[-5])
fa_loadings(iris_fa)
}

