% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fa_MAR}
\alias{fa_MAR}
\title{Calculates the mean absolute residual at the case-level.}
\usage{
fa_MAR(data, sort = T, ...)
}
\arguments{
\item{data}{a data.frame.}

\item{sort}{whether to sort the results or not. Defaults to true.}
}
\description{
Extracts the first factor, then uses the factor scores to predict the indicator values for each indicator and for each case. Returns a data.frame with residuals. Then calculates the mean absolute value of these by case.
}
\examples{
fa_MAR(iris[-5])
}

