% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fa_rank_fa}
\alias{fa_rank_fa}
\title{Rank order factor analysis}
\usage{
fa_rank_fa(x, ...)
}
\arguments{
\item{x}{A data.frame to extract factors from.}

\item{...}{Parameters to fa().}
}
\description{
Runs factor analysis on a rank-ordered dataset which prevents outliers from having strong effects.
}
\examples{
fa_rank_fa()
}

