% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fa_residuals}
\alias{fa_residuals}
\title{Find residuals on case-level basis for all indicators in a factor analysis.}
\usage{
fa_residuals(data, standardize = T, ...)
}
\arguments{
\item{data}{(data.frame) The data.}

\item{standardize}{(log scalar) Whether to standardize the residuals (default true). If not done, they may not have the same standard deviation.}

\item{...}{(arguments to fa) Further arguments to \code{\link{fa}}.}
}
\description{
Extracts the first factor, then uses the factor scores to predict the indicator values for each indicator and for each case. Returns a data.frame with residuals.
}
\examples{
fa_residuals(iris[-5])
}

