% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_robust_cormatrix}
\alias{FA_robust_cormatrix}
\title{Robust correlation matrix}
\usage{
FA_robust_cormatrix(x, method = "lmrob")
}
\arguments{
\item{x}{A data.frame to correlate.}

\item{x}{method Which robust method to use. Options are "lmrob" [robustbase] or "rlm" [MASS]. Defaults to the first.}
}
\description{
Returns a correlation matrix with robust correlations. These are derived from either rlm() [MASS] or lmrob() [robustbase].
}
\examples{
FA_robust_cormatrix()
}
\keyword{correlation,}
\keyword{matrix}
\keyword{psychometrics,}
\keyword{robust,}

