% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fa_robust_fa}
\alias{fa_robust_fa}
\title{Robust factor analysis.}
\usage{
fa_robust_fa(x, ..., .method = "lmrob")
}
\arguments{
\item{x}{A data.frame to factor analyze.}
}
\description{
Performs a robust factor analysis. It is done using the output from fa_robust_cormatrix(). Note that this uses a correlation matrix, so factor scores are not available.
}
\examples{
fa_robust_cormatrix()
}
