% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fa_salient_loadings}
\alias{fa_salient_loadings}
\title{Get salient loadings from a factor analysis object.}
\usage{
fa_salient_loadings(fa, threshold = 0.4, to_df = T)
}
\arguments{
\item{fa}{(object of class "fa") A factor analysis object.}

\item{threshold}{(numeric scalar) The cutoff to use for salient loadings. Default to .40.}
}
\description{
Find indicators that load above a given threshold for each factor. Can also force them into a data.frame for easy output. Uses input object from psych package's fa().
}

