% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_splitsample_repeat}
\alias{FA_splitsample_repeat}
\title{Repeated splithalf reliability with factor analysis}
\usage{
FA_splitsample_repeat(df, runs = 100, save_scores = F, ...)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{var}{A string of the name of the variable to use.}
}
\description{
Divides a dataset into 2 at random, extracts a factor from each and correlates them. Then saves the correlation. Repeats this any desired number of times. Can also return the factor scores instead of correlations.
}
\examples{
FA_splitsample_repeat()
}
\keyword{density,}
\keyword{ggplot2,}
\keyword{histogram}
\keyword{plot,}

