% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{fail_if}
\alias{fail_if}
\alias{stop_if}
\title{Fail conditionally}
\usage{
fail_if(x, msg = "There was an error", extended = F)
}
\arguments{
\item{x}{(any object) The object to test.}

\item{msg}{(chr) An error message.}

\item{extended}{(lgl) Whether to convert to logical.}
}
\value{
Invisibly returns the input.
}
\description{
Fail on not TRUE and give a custom error message.
}
\examples{
fail_if(T)
fail_if(1) #does not convert types
fail_if(1, extended = T)
}
