% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{fail_if_NA}
\alias{fail_if_NA}
\title{Fail if input contains NA}
\usage{
fail_if_NA(x, msg = "Input contained NA.")
}
\arguments{
\item{x}{(any object) The object to test.}

\item{msg}{(chr) An error message.}
}
\description{
Inputs any object, checks if it contains NA. If yes, throws an error. If not, returns the input.
}
\examples{
fail_if_NA(1:10) #no NAs
fail_if_NA(c(1:3, NA, 1:3)) #NAs
}
