% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fill_in}
\alias{fill_in}
\title{Fill in values in a vector}
\usage{
fill_in(x, length, value = NA)
}
\arguments{
\item{x}{(an object) An object whose dimnames should be copied.}

\item{length}{(numeric scalar) The desired length.}

\item{value}{(numeric/character/logical scalar) The value to fill in.}
}
\description{
Fill in values in a vector until it reaches a specific length.
}
\examples{
fill_in(1:5, 10)
}
\keyword{fill}
\keyword{vector,}

