% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{filter_by_missing_values}
\alias{filter_by_missing_values}
\title{Filter data by missing values per row.}
\usage{
filter_by_missing_values(data, missing = 0)
}
\arguments{
\item{data}{(data.frame or something coersible to a data.frame) The data.}

\item{missing}{(whole number scalar) The maximum number of missing values in cases. Defaults to 0 (keep only cases with no missing values).}
}
\description{
Counts the number of missing values per row and then keeps rows that have at most a chosen number of missing values.
}
\examples{
df = data.frame(1:10, letters[1:10])
df = miss_add_random(df)
filter_by_missing_values(df)
}
\keyword{missing}
\keyword{subset}
\keyword{values,}

