% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{find_cutoff}
\alias{find_cutoff}
\title{Find a cutoff of a normal distribution that results in a given mean trait value above the cutoff}
\usage{
find_cutoff(mean_above, mean_pop = 100, sd_pop = 15, n = 10000,
  precision = 0.1, below = F)
}
\arguments{
\item{mean_above}{(num scalar) Mean trait level of population above the cutoff.}

\item{mean_pop}{(num scalar) Mean trait level of population. Default = 100 (IQ scale).}

\item{sd_pop}{(num scalar) Standard deviation of the trait in the population. Default = 15 (IQ scale).}

\item{n}{(num scalar) Sample size to generate in the process. More results in higher precision and memory use. Default = 1e4.}

\item{precision}{(num scalar) The precision to use. Default = .1.}

\item{below}{(log scalar) Reverse the model to find cutoffs for}
}
\description{
Assuming a normal distribution for a trait and a cutoff value. Estimate what this cutoff value is to obtain a population above the cutoff with a known mean trait value.
}
\examples{
#what cutoff is needed to get a population above the cutoff with a mean of 115 when the population mean is 100?
find_cutoff(115)
#try impossible
find_cutoff(95)
}
