% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC.R
\name{get_distances_mat}
\alias{get_distances_mat}
\title{Calculate distance matrices for all variables in a data.frame.}
\usage{
get_distances_mat(df, lat_var, lon_var, distance_method,
  auto_detect_dist_method = T)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{lat_var}{A string with the name of the variable which has the latitude data. Defaults to "lat".}

\item{lon_var}{A string with the name of the variable which has the longitude data. Defaults to "lon".}

\item{auto_detect_dist_method}{Whether to try to autodetect the distance method. If the dataset contains variables with the names "lat" and "lon", it will be detected as spherical. If it contains "x" and "y", it will be detected as euclidean. Defaults to true.}

\item{dists}{A matrix of distances between cases.}
}
\description{
Inputs a data.frame and outputs a list with distance matrices. One can specify that a pair of variables are spherical coordinates which are then treated in a special fashion using get_spherical_dists().
The function requires either that it is given a distance matrix or that it can find variables to calculate distances with.
This is a sister function to get_distances(), which outputs vectors.
}

