% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{get_each_subset_minus_1}
\alias{get_each_subset_minus_1}
\title{Get every subset of the data where one case is missing.}
\usage{
get_each_subset_minus_1(df)
}
\arguments{
\item{df}{(df) A data.frame.}
}
\description{
Returns a list of data.frames. This function is useful for doing leave one out cross validation (LOOCV).
}
\examples{
df = data.frame(n = 1:5, a = letters[1:5])
get_each_subset_minus_1(df)
}
