% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC.R
\name{get_euclidean_dists}
\alias{get_euclidean_dists}
\title{Calculate pairwise euclidean distances for all pairs.}
\usage{
get_euclidean_dists(df, output = "vector", remove_diag = T)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{output}{A character vector stating which output form is desired. Can be either vector or matrix. Vector results in a vector of all the unique combinations of distances, which is n*(n-1)/2, where n is the number of cases. Matrix results in the full distance matrix with every permutation including diagonals, which are always 0.}
}
\description{
Returns a numeric vector with the pairwise euclidean distances for all pairs.
}

