% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{get_loadings}
\alias{get_loadings}
\title{Get loadings from a factor analysis object.}
\usage{
get_loadings(fa)
}
\arguments{
\item{fa}{(object of class "fa") A factor analysis object.}
}
\description{
Extract loadings from a factor analysis object of class "fa") into a data.frame.
}
\examples{
library("psych")
iris_fa = fa(iris[-5])
get_loadings(iris_fa)
}

