% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_model_coefs}
\alias{get_model_coefs}
\title{Extract model coefficients from a list of models}
\usage{
get_model_coefs(models, conf.level = 0.95, nicer_factor_levels = T)
}
\arguments{
\item{models}{A list of models}

\item{conf.level}{Which confidence level to use. Default is .95.}

\item{nicer_factor_levels}{Whether to make factor levels nicer. Default is TRUE.}
}
\value{
Data frame
}
\description{
Uses \code{broom::tidy} to extract coefficients from models. If the model doesn't support this, it will fail. There is a special case for \code{rlm} models, where the p value is calculated manually. If the model is of class \code{rms}, then the \code{summary.lm} function is used to extract coefficients. There is some special handling for factor levels, where the levels are made nicer, but this is not done for \code{rms} models.
}
\examples{
list(
a = lm(Sepal.Width ~ Sepal.Length, data = iris),
b = lm(Sepal.Width ~ Sepal.Length + Petal.Width, data = iris)
) \%>\% get_model_coefs()
}
