% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAC.R
\name{get_Morans_I}
\alias{get_Morans_I}
\title{Calculate Moran's I using spherical or euclidean geomtry.}
\usage{
get_Morans_I(df, var, lat_var = "lat", lon_var = "lon",
  distance_method = "spherical")
}
\arguments{
\item{df}{A data.frame with variables.}

\item{var}{A string with the name of the variable for which Moran's I should be calculated.}

\item{lat_var}{A string with the name of the variable which has the latitude data.}

\item{lon_var}{A string with the name of the variable which has the longitude data.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean. If using euclidean it doesn't matter which variable is coded as lat or lon.}
}
\description{
Returns Moran's I coefficient and related NHST statistics. Excludes cases with missing data.
}
\examples{
get_Morans_I()
}
\keyword{I}
\keyword{Moran's}
\keyword{autocorrelation,}
\keyword{distance,}
\keyword{latitude,}
\keyword{longitude,}
\keyword{spatial}

