% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAC.R
\name{get_Morans_I_multi}
\alias{get_Morans_I_multi}
\title{Calculate multiple Moran's I's using spherical or euclidean geomtry. It's a wrapper function for get_Morans_I().}
\usage{
get_Morans_I_multi(df, vars, lat_var = "lat", lon_var = "lon",
  distance_method = "spherical")
}
\arguments{
\item{df}{A data.frame with variables.}

\item{vars}{A character vector with the names of the variables for which Moran's I should be calculated.}

\item{lat_var}{A string with the name of the variable which has the latitude data.}

\item{lon_var}{A string with the name of the variable which has the longitude data.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean. If using euclidean it doesn't matter which variable is coded as lat or lon.}
}
\description{
Returns a numeric vector with Moran's I coefficients.
}
\examples{
get_Morans_I_multi()
}
\keyword{I,}
\keyword{Moran's}
\keyword{autocorrelation,}
\keyword{distance,}
\keyword{latitude,}
\keyword{longitude,}
\keyword{spatial}
\keyword{wrapper}

