% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC.R
\name{get_pairwise_means}
\alias{get_pairwise_means}
\title{Calculate pairwise means.}
\usage{
get_pairwise_means(x, weight_method = "harmonic")
}
\arguments{
\item{x}{A numeric object that can be forced into a data.frame.}

\item{weight_method}{A character string indicating which averaging method to use when combining weights for cases. Defaults to harmonic mean. Other options are arithmic and geometric.}
}
\description{
Returns a data.frame with the pairwise means of the desired type.
}

