% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{get_prop_table}
\alias{get_prop_table}
\title{Cut into bins and get proportions}
\usage{
get_prop_table(x, breaks_ = 20)
}
\arguments{
\item{x}{A numeric vector.}

\item{breaks_}{The number of bins to use.}
}
\description{
Cuts a vector into a specified number of equal sized bins and calculations the proportion of datapoints in each bin. Returns a data.frame.
}
\examples{
get_prop_table()
}
\keyword{bins,}
\keyword{cut,}
\keyword{proportion,}
\keyword{table}

