% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirt.R
\name{get_reliabilities}
\alias{get_reliabilities}
\title{Get reliability distribution of a \code{mirt} model}
\usage{
get_reliabilities(x)
}
\arguments{
\item{x}{A \code{mirt} fit object}
}
\value{
A data frame of reliability values
}
\description{
Get reliability distribution of a \code{mirt} model
}
\examples{
library(mirt)
data = simdata(seq(0.2, 2, length.out = 5), seq(-2, 2, length.out = 5), 1000, itemtype = "2PL")
fit = mirt(data, 1)
get_reliabilities(fit)
}
