% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAC.R
\name{get_SAC_measures}
\alias{get_SAC_measures}
\title{Calculate multiple spatial autocorrelation measures.}
\usage{
get_SAC_measures(df, vars, lat_var = "lat", lon_var = "lon",
  distance_method = "spherical", k = 3, weights_var = "",
  weight_method = "harmonic")
}
\arguments{
\item{df}{A data.frame with variables.}

\item{vars}{A character vector with the names of the variables for which SAC measures should be calculated.}

\item{lat_var}{A string with the name of the variable which has the latitude/east-west data.}

\item{lon_var}{A string with the name of the variable which has the longitude/north-south data.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean. If using euclidean it doesn't matter which variable is coded as lat or lon.}

\item{k}{A vector of k values to use for knsnr. Defaults to 3.}

\item{weights_var}{A string with the name of the variable which has the case weights. Optional.}

\item{weight_method}{A string with the weighing method to use. Defaults to harmonic.}
}
\description{
Returns a data.frame with measures of SAC using Moran's I, CD, CD_sqrt and KNSNR.
}
\examples{
get_SAC_measures()
}
\keyword{I,}
\keyword{Moran's}
\keyword{autocorrelation,}
\keyword{distance,}
\keyword{latitude,}
\keyword{longitude,}
\keyword{spatial}
\keyword{wrapper}

