% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAC.R
\name{get_spherical_dists}
\alias{get_spherical_dists}
\title{Calculate pairwise spherical distances for all pairs.}
\usage{
get_spherical_dists(df, lat_var = "lat", lon_var = "lon")
}
\arguments{
\item{df}{A data.frame with variables.}

\item{lat_var}{A numeric vector with the latitudes.}

\item{lon_var}{A numeric vector with the longitudes.}
}
\description{
Returns a vector with the pairwise spherical distances for all pairs.
}
\examples{
get_spherical_dists()
}
\keyword{autocorrelation,}
\keyword{distance}
\keyword{latitude,}
\keyword{longitude,}
\keyword{spatial}

