% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_t_value}
\alias{get_t_value}
\title{Get t value by confidence-level and degree of freedom}
\usage{
get_t_value(conf, df, ...)
}
\arguments{
\item{conf}{(numeric scalar) The confidence level desired as a fraction.}

\item{df}{(numeric scalar) The degrees of freedom.}

\item{...}{(other named arguments) Other arguments to pass to \code{qt}. See that function for details.}
}
\description{
Wrapper function for \code{qt} to get the t value needed for finding confidence intervals or p values.
}
\examples{
#get t value for 95 pct. confidence interval with df = 20
get_t_value(.95, 20)
}

