% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_BMA}
\alias{GG_BMA}
\title{Plot BMA results}
\usage{
GG_BMA(x, confidence_level = 0.95)
}
\arguments{
\item{x}{BMA model fit from BMA, BAS (call \code{coef()} first), or BMS}

\item{confidence_level}{Confidence level for the error bars. Default is 0.95.}
}
\value{
A ggplot2 object
}
\description{
A ggplot2 function for plotting the output of BMA models, from the BMA, BAS, or BMS packages.
}
\examples{
#fit a model
fit = BAS::bas.lm(Sepal.Length ~ ., data = iris)
#get coefs
fit_coefs = coef(fit)
#plot
GG_BMA(fit_coefs)
GG_BMA(fit_coefs, confidence_level = .99)
}
