% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_denhist}
\alias{GG_denhist}
\title{Histogram with an empirical density curve and a vertical line at the mean}
\usage{
GG_denhist(df, var, binwidth = NULL)
}
\arguments{
\item{df}{(data.frame or something coercible into) A data.frame with variables.}

\item{var}{(character scalar) The name of the variable to use. Not needed if df is a vector.}
}
\description{
Plots a histogram with an empirical density curve and a vertical line at the mean using ggplot2.
}
\examples{
GG_denhist(iris, "Sepal.Length")
}

