% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{GG_funnel}
\alias{GG_funnel}
\title{Funnel plot with ggplot2}
\usage{
GG_funnel(.analysis, .CI = 0.95, .study_CI = F)
}
\arguments{
\item{.analysis}{(rma object) The rma analysis from \link[metafor:metafor-package]{metafor::metafor}.}

\item{.CI}{(chr vector) Confidence interval to use.}

\item{.study_CI}{(lgl vector) Whether to plot confidence intervals for individual studies.}
}
\description{
Makes a pretty funnel plot using \link[ggplot2:ggplot2-package]{ggplot2::ggplot2}.
}
\examples{
library(metafor); data(european_ancestry)
meta = rma(european_ancestry$r, sei = european_ancestry$SE_r)
GG_funnel(meta)
GG_funnel(meta, .study_CI = T)
}
