% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_group_means}
\alias{GG_group_means}
\title{ggplot2 with group means and error bars.}
\usage{
GG_group_means(df, var, groupvar, CI = 0.95, type = "bar")
}
\arguments{
\item{df}{(data.frame) A data.frame with variables.}

\item{var}{(character scalar) The name of the variable to plot.}

\item{groupvar}{(character scaler) The name of the grouping variable.}

\item{CI}{(numeric scalar) The confidence interval to use. Default = .95.}

\item{type}{(character scalar) The type of plot. Options: bar, point, points. Default = bar.}
}
\description{
Draws a nice ggplot2 with group means and error bars.
}
\examples{
GG_group_means()
}
\keyword{bars}
\keyword{confidence}
\keyword{error}
\keyword{ggplot2,}
\keyword{group,}
\keyword{interval,}
\keyword{means,}
\keyword{plot,}

