% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_lines}
\alias{GG_lines}
\title{Plot lines with labels}
\usage{
GG_lines(data, x, y, color, right_margin = 100, points = T)
}
\arguments{
\item{data}{A data frame with x, y, and color variables}

\item{x}{A string with the x variable name}

\item{y}{A string with the y variable name}

\item{color}{A string with the color variable name}

\item{right_margin}{Right margin for the plot. Default is 100.}

\item{points}{Whether to add points to the plot. Default is TRUE.}
}
\value{
A ggplot2 object
}
\description{
Plot lines with labels
}
\examples{
tidyr::population \%>\%
filter(country \%in\% (.env$population$country \%>\% unique() \%>\% str_subset(pattern = "^A") \%>\% head(10))) \%>\%
GG_lines("year", "population", "country") +
scale_y_log10()
}
