% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_matrix}
\alias{GG_matrix}
\title{Matrix of data values}
\usage{
GG_matrix(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A ggplot2 object
}
\description{
Matrix of data values
}
\examples{
iris[1:4] \%>\% GG_matrix()
}
