% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_proportions}
\alias{GG_proportions}
\title{ggplot2 proportions plot}
\usage{
GG_proportions(
  x,
  group,
  drop_empty = F,
  stacked = F,
  add_values = F,
  repel = F,
  seed = 1
)
}
\arguments{
\item{x}{A factor variable}

\item{group}{Grouping variable}

\item{drop_empty}{Drop empty combinations}

\item{stacked}{Stacked bar}

\item{add_values}{Add values to the plot}

\item{repel}{Repel the text labels in case of problematic overplotting}

\item{seed}{Seed for the repel algorithm}
}
\value{
A ggplot2 of proportions
}
\description{
ggplot2 proportions plot
}
\examples{
#plot the proportions of cylinders by year
GG_proportions(mpg$year, mpg$cyl)

#remove the 0\%'s
GG_proportions(mpg$year, mpg$cyl, drop_empty = T)
#stacked
GG_proportions(mpg$year, mpg$cyl, drop_empty = T, stacked = T)
#with values
GG_proportions(mpg$year, mpg$cyl, drop_empty = T, stacked = T, add_values = T)
#with text repel, but it doesn't work so well
GG_proportions(mpg$year, mpg$cyl, drop_empty = T, stacked = T, add_values = T, repel = T)
}
