% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_quantiles}
\alias{GG_quantiles}
\title{Plot quantiles of y on x}
\usage{
GG_quantiles(
  data,
  x_var,
  y_var,
  quantiles = c(0.1, 0.25, 0.5, 0.75, 0.9),
  method = "qgam"
)
}
\arguments{
\item{data}{Data frame}

\item{x_var}{X variable name}

\item{y_var}{Y variable name}

\item{quantiles}{Which quantiles to plot}

\item{method}{Method to use with \code{\link[=quantile_smooth]{quantile_smooth()}}}
}
\value{
A ggplot2 object
}
\description{
Plot quantiles of y on x
}
\examples{
GG_quantiles(diamonds \%>\% slice_sample(n=2000), "carat", "price")
}
