% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_save}
\alias{GG_save}
\title{Save a ggplot2 figure}
\usage{
GG_save(filename, plot = last_plot(), device = NULL, path = NULL,
  scale = 1, width = 10, height = 6.5, units = c("in", "cm", "mm"),
  dpi = 300, limitsize = TRUE, ...)
}
\description{
Save a ggplot2 figure. Now with a sensible default size (10 x 6.5 inches). Same arguments as \code{ggplot2::ggsave}
}
