% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_save_pdf}
\alias{GG_save_pdf}
\title{Save list of ggplot2 objects to single pdf}
\usage{
GG_save_pdf(list, filename)
}
\arguments{
\item{list}{(list) List of ggplot2 objects.}

\item{filename}{(chr) What to call the pdf.}
}
\value{
Invisible NULL.
}
\description{
Save list of ggplot2 objects to single pdf
}
\examples{
#plot histogram of each numeric variable in iris
list_iris = map(names(iris[-5]), ~ggplot(iris, aes_string(.)) + geom_histogram())
#save to a single pdf
GG_save_pdf(list_iris, "test.pdf")
file.remove("test.pdf")
}
