% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_scale_abbreviation}
\alias{GG_scale_abbreviation}
\title{Plot results from scale abbreviation}
\usage{
GG_scale_abbreviation(
  x,
  vars = c("reliability_frac", "mean_criterion_cors_frac", "r_full_score", "reliability"),
  add_lines_for_full_model = F
)
}
\arguments{
\item{x}{Output from \code{scale_abbreviation()}}

\item{vars}{Which variables to plot. Default is c("reliability_frac", "mean_criterion_cors_frac", "r_full_score", "reliability")}

\item{add_lines_for_full_model}{Add lines for the full model for comparison. Default is FALSE.}
}
\value{
A ggplot2 object
}
\description{
Plots results from the output of \code{scale_abbreviation()}.
}
