% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_scatter}
\alias{GG_scatter}
\title{Scatter plot with regression line and correlation information using ggplot2}
\usage{
GG_scatter(df, x_var, y_var, text_pos, case_names = T, CI = 0.95,
  clean_names = T)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{x_var}{X variable string.}

\item{y_var}{Y variable string.}

\item{text_pos}{Where to put the text. Defaults to top right ("tl") if correlation is positive, or tr if negative. Can be tl, tr, bl, or br.}

\item{case_names}{Whether to add case names or not. Defaults to true. Row names are used for case names.}

\item{CI}{Confidence interval. Defaults to .95. Set to NULL to disable.}

\item{clean_names}{(boolean) Whether to clean the axes names using str_clean(). Default=T.}
}
\description{
Plots a scatterplot with a regression line and correlation information. Returns a ggplot2 object.
}
\examples{
GG_scatter()
}
\keyword{ggplot2,}
\keyword{plot,}
\keyword{scatter}

