% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%equals\%}
\alias{\%equals\%}
\title{Equality infix operator}
\usage{
x \%equals\% y
}
\arguments{
\item{x}{An object}

\item{y}{An object}
}
\value{
A boolean scalar
}
\description{
Equality infix operator
}
\examples{
#simple
1 \%equals\% 1
1 \%equals\% 2
#works for vectors too
1:3 \%equals\% 1:3
#compare with
1:3 == 1:3
}
