% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{helper_breaks}
\alias{helper_breaks}
\title{Augment logarithmic values with intuitive additional breaks}
\usage{
helper_breaks(x, factors = c(2.5, 5))
}
\arguments{
\item{x}{(num vectr) A vector of values.}

\item{factors}{(num vectr) Factors to use to create the new values. Default is 2.5 and 5.}
}
\description{
Adds intuitive middle values to a vector of values. Useful for when plotting data on logarithmnic plots.
}
\examples{
#standard log10 breaks
10^(0:5)
#add extra
helper_breaks(10^(0:5))
}

