% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{homogeneity}
\alias{homogeneity}
\title{Calculate homogeneity/heterogeneity}
\usage{
homogeneity(x, reverse = F)
}
\arguments{
\item{x}{(a vector) A vector of values.}

\item{reverse}{(log scalar) Whether to reverse the index to index heterogeneity.}
}
\description{
Calculate a simple index of homogeneity for nominal data, that is variously called Simpson's, Herfindahl's or Hirschman's index.
}
\examples{
homogeneity(iris$Species)
homogeneity(iris$Species, reverse = T)
}

