% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inf_to_NA}
\alias{inf_to_NA}
\title{Conert infinite values to \code{NA}}
\usage{
inf_to_NA(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of the same length as x
}
\description{
Conert infinite values to \code{NA}
}
\examples{
c(1, 2, Inf, 3) \%>\% inf_to_NA()
}
