% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_table}
\alias{inv_table}
\title{Reverse \code{table()}}
\usage{
inv_table(x)
}
\arguments{
\item{x}{A table}
}
\value{
A data frame
}
\description{
Function from https://stackoverflow.com/questions/30481496/is-there-a-general-inverse-of-the-table-function.
}
\examples{
mpg \%>\% nrow() #324
inv_table(table(mpg$cyl, mpg$class, mpg$manufacturer)) #also 324
inv_table(Titanic)

}
