% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{is_}
\alias{is_}
\title{Check whether object is the right class, size and type}
\usage{
is_(x, class, size, type, error_on_false = F)
}
\arguments{
\item{x}{(any object) The object to test.}

\item{class}{(chr vector) The accepted classes. This is checked with \code{\link{is}}.}

\item{size}{(num vector) The accepted lengths. This is checked with \code{\link{get_dims}}.}

\item{type}{(chr vector) The accepted types. This is checked with \code{\link{typeof}}.}

\item{error_on_false}{(log scalar) Whether to throw an error instead of returning false (default false).}
}
\value{
Logical scalar or error.
}
\description{
Checks both the class, size and type of an object. Returns a logical or an error if false.
}
\details{
Vectors are treated as 1-dimensional.
}
\examples{
is_(iris, class = "data.frame") #check for one class
is_(iris, class = c("data.frame", "logical", "matrix")) #can check for multiple classes
is_(iris, class = "data.frame", size = c(150, 5)) #check for one class and size
is_(iris, size = c(150, 5)) #check for size
is_(iris, size = 1) #check for wrong size
is_(iris, type = "list") #check for type
is_(iris, type = "factor") #check for wrong type
is_(iris, class = "list", error_on_false = T) #check for one class, error
}

