% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{is_between}
\alias{is_between}
\title{Check whether numbers are between two other numbers.}
\usage{
is_between(x, a, b, include_lower = TRUE, include_upper = TRUE)
}
\arguments{
\item{x}{(numeric vector) A vector of values to test.}

\item{a}{(numeric scalar) The lower limit.}

\item{b}{(numeric scalar) The upper limit.}

\item{include_lower}{(boolean) Whether to include the lower limit itself. Defaults to TRUE.}

\item{include_upper}{(boolean) Whether to include the upper limit itself. Defaults to TRUE.}
}
\description{
Returns a logical vector of the same length as x. Can use inclusive or exclusive limits.
}
\examples{
is_between(1:10, 3, 6)
}
\keyword{between}
\keyword{limit,}

