% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{is_error}
\alias{is_error}
\title{Is an object a try-error?}
\usage{
is_error(x)
}
\arguments{
\item{x}{(an object) An object to check.}
}
\description{
Checks whether an object is a try-error. Returns a boolean.
}

