% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{is_logical}
\alias{is_logical}
\title{Is logical?}
\usage{
is_logical(x, scalar = F)
}
\arguments{
\item{x}{(obj) An object to check.}

\item{scalar}{(lgl) Only accept scalar.}
}
\value{
Logical scalar.
}
\description{
Check whether an object is logical
}
\examples{
is_scalar_NA(NA)
is_scalar(1:3)
is_scalar_NA(c(1, NA))
is_scalar_NA(c(NA, 1))
}

