% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{is_negative}
\alias{is_negative}
\title{Test if value is negative}
\usage{
is_negative(x)
}
\arguments{
\item{x}{(any object compatible with <) Some values to test.}
}
\description{
A wrapper for "<0", returns a boolean.
}
\examples{
is_negative()
}
\keyword{boolean}
\keyword{negative,}
\keyword{value,}

