% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_nullish}
\alias{is_nullish}
\title{Is x nullish?}
\usage{
is_nullish(x, false = T, na = T, zero = T, null = T)
}
\arguments{
\item{x}{Length 0 or 1 input}

\item{false}{Whether False means "no"}

\item{na}{Whether NA means "no"}

\item{zero}{Whether 0 means "no"}

\item{null}{Whether NULL means "no"}
}
\value{
Logical scalar
}
\description{
Sometimes useful to have flexible input to disable a feature that accepts NULL, NA, or F as meaning "no".
}
\examples{
is_nullish(NULL)
is_nullish(NA)
is_nullish(F)
is_nullish(0)
is_nullish(NA, na = F)
}
