% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_numeric}
\alias{is_numeric}
\title{Is object thoroughly numeric?}
\usage{
is_numeric(x, recursive = TRUE)
}
\arguments{
\item{x}{(any object) An object to test.}

\item{recursive}{(logical scalar) Whether to use recursive checking. Default=TRUE.}
}
\value{
Returns a logical scalar indicating whether the object is thoroughly numeric.
}
\description{
A more advanced version of \code{\link{is.numeric}}. It wraps the base-r function, but allows for recursive checking inside lists and hence data.frames as well.
}
\examples{
is_numeric(iris)
is_numeric(iris[-5])
}

