% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_numeric_by_col}
\alias{is_numeric_by_col}
\title{Check numericalness by column.}
\usage{
is_numeric_by_col(df)
}
\arguments{
\item{x}{(something coercible to a data.frame) An object to test.}
}
\value{
Returns a logical vector the same length as the number of columns in x.
}
\description{
A simple wrapper for \code{vapply}.
}
\examples{
is_numeric_by_col(iris)
}

