% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{is_positive}
\alias{is_positive}
\title{Test if value is positive}
\usage{
is_positive(x)
}
\arguments{
\item{x}{(any object compatible with <) Some values to test.}
}
\description{
A wrapper for ">0", returns a boolean.
}
\examples{
is_positive()
}
\keyword{boolean}
\keyword{positive,}
\keyword{value,}

