% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{is_scalar}
\alias{is_scalar}
\title{Is scalar?}
\usage{
is_scalar(x, error_on_null = T)
}
\arguments{
\item{x}{(obj) An object to check.}

\item{erro_on_null}{(lgl) Whether to throw an error on NULL, or treat as TRUE.}
}
\value{
Logical scalar.
}
\description{
Check whether an object is a scalar
}
\examples{
is_scalar(1)
is_scalar(1:3)
is_scallar(NULL, error_on_null = F)
}
