% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{is_unordered_factor}
\alias{is_unordered_factor}
\title{Is object an unordered factor?}
\usage{
is_unordered_factor(x)
}
\arguments{
\item{x}{(any object) An object to test.}
}
\description{
A simple wrapper for is.factor and is.ordered.
}
\examples{
is_unordered_factor(factor(1:3))
is_unordered_factor(ordered(1:3))
}

