% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{is_whole_number}
\alias{is_whole_number}
\title{Is vector a whole number?}
\usage{
is_whole_number(x)
}
\arguments{
\item{x}{(numeric vector) A numeric vector.}
}
\description{
Tests via x = is.integer(x).
}
\examples{
is_whole_number(1L)
is.integer(1)
is_whole_number(1)
}

