% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{is_zero}
\alias{is_zero}
\title{Test if value is zero}
\usage{
is_zero(x)
}
\arguments{
\item{x}{(any object compatible with <) Some values to test.}
}
\description{
A wrapper for "==0", returns a boolean.
}
\examples{
is_zero()
}
\keyword{boolean}
\keyword{value,}
\keyword{zero,}

