% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Jensens_method}
\alias{Jensens_method}
\title{Scatter plot of Jensens method.}
\usage{
Jensens_method(fa, df, criteria, reverse_factor = F, loading_reversing = T,
  text_pos)
}
\arguments{
\item{fa}{A factor analysis object from fa().}

\item{df}{A data.frame that contains all the variables.}

\item{criteria}{A character string of the name of the criteria variable.}

\item{reverse_factor}{Whether to reverse the factor first.}

\item{loading_reversing}{Whether to use loading reversing to avoid inflated results. Defaults to TRUE.}

\item{text_pos}{Which corner to put the text in. Defaults to "tl". Other options: tr, bl, br.}
}
\description{
Takes a factor analysis, data.frame and name of the criteria variable as inputs and returns a ggplot2 scatter plot with Jensen's method applied.
}
\examples{
Jensens_method()
}
\keyword{Jensen,}
\keyword{analysis,}
\keyword{correlated}
\keyword{factor}
\keyword{method}
\keyword{of}
\keyword{vectors}

