% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAC.R
\name{knsn_reg}
\alias{knsn_reg}
\title{Perform k nearest spatial neighbor regression.}
\usage{
knsn_reg(df, dependent, k = 3, lat_var = "lat", lon_var = "lon",
  weights_var = "", distance_method = "spherical", output = "scores")
}
\arguments{
\item{df}{A data.frame with variables.}

\item{dependent}{A string with the name of the dependent variable.}

\item{k}{The number of neighbors taken into account. Defaults to 3.}

\item{lat_var}{A string with the name of the variable which has the latitude data.}

\item{lon_var}{A string with the name of the variable which has the longitude data.}

\item{weights_var}{A string with the name of the variable that contains case weights.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean. If using euclidean it doesn't matter which variable is coded as lat or lon.}

\item{weight}{The weight to assign the value from the neighbors in each iteration.}
}
\description{
Predict the value of cases on the basis of values of nearby cases. A useful measure of spatial autocorrelation. Returns a data.frame. Keeps rownames.
}
\examples{
knsn_reg()
}
\keyword{autocorrelation,}
\keyword{distance,}
\keyword{knn,}
\keyword{knsn,}
\keyword{latitude,}
\keyword{longitude,}
\keyword{simulation}
\keyword{spatial}

