% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{last_value}
\alias{last_value}
\title{Last value}
\usage{
last_value(x, na.rm = T)
}
\arguments{
\item{x}{(a vector) A vector.}

\item{na.rm}{(log scalar) Whether to ignore missing values (default yes).}
}
\value{
The last value (scalar).
}
\description{
Find the last value of a vector. By default, will take the last non-NA value.
}
\examples{
last_value(1:3)
last_value(c(1:3, NA))
last_value(rep(NA, 3))
}

