% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{ldf_to_df}
\alias{ldf_to_df}
\title{Transform list of data.frames to a data.frame.}
\usage{
ldf_to_df(list, add_by = T, by_name = "group", rownames_to_var = F,
  rownames_name = "rownames")
}
\arguments{
\item{list}{(list of data.frames) The data.frames.}

\item{add_by}{(logical scalar) Whether to add a variable indicating the origin of each row.}

\item{by_name}{(character scalar) What should the added variable by called? Only used when add_by = TRUE.}

\item{rownames_to_var}{(logical scalar) Should rownames be saved in a variable? (Default no)}

\item{rownames_name}{(logical scalar) The variable name to save rownames under. Only used if rownames_to_var = TRUE.}
}
\description{
Combines a split data.frame back to the original.
}
\examples{
ldf_to_df(df_to_ldf(iris), "Species")
}

