% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{lengths_match}
\alias{lengths_match}
\title{Do the objects have the same lengths?}
\usage{
lengths_match(..., dimension = 1)
}
\arguments{
\item{...}{(objects) The objects to test.}

\item{dimension}{(num scalar) The dimension to test.}
}
\value{
Logical scalar.
}
\description{
Checks whether the nth dimension lengths match for the given objects.
}
\details{
Vectors are treated as 1-dimensional.
}
\examples{
lengths_match(1:4, 5:8) #same lengths
lengths_match(iris, iris[1:2]) #same nrow
lengths_match(iris, iris[1:2], dimension = 2) #different ncol
lengths_match(iris, 1:3, dimension = 2) #incompatible objects
}

