% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{lm_best}
\alias{lm_best}
\title{Find the best model by R2 adj. value.}
\usage{
lm_best(model_list)
}
\arguments{
\item{model_list}{A list of model fits e.g. from lm().}
}
\description{
Returns the index of the model with the highest R2 adj. value.
}
\examples{
lm_best()
}

