% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{lm_beta_matrix}
\alias{lm_beta_matrix}
\title{Matrix of beta coefficients of all simple linear models.}
\usage{
lm_beta_matrix(dependent, predictors, data, standardized = T, .weights = NA,
  return_models = "b")
}
\arguments{
\item{dependent}{A string with the name of the dependent variable.}

\item{predictors}{A character vector with the predictor variables.}

\item{data}{A data.frame with the variables.}

\item{standardized}{Whether to standardize the results. Defaults to true.}

\item{.weights}{Weights to use. Leave blank not to use weights.}
}
\description{
Returns a data.frame with beta coefficients for all possible simple (without interactions) linear models given a set of predictor variables and a dependent variable.
}
\examples{
lm_beta_matrix()
}
\keyword{automatic}
\keyword{linear,}
\keyword{lm,}
\keyword{modeling,}

