% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modeling.R
\name{lm_CI}
\alias{lm_CI}
\title{Convenient summary of an lm() model with confidence intervals.}
\usage{
lm_CI(fitted_model, level = 0.95, round = 2)
}
\arguments{
\item{fitted_model}{the fitted model to summarize.}

\item{level}{the level of confidence to use. Defaults to .95 (95\%).}

\item{round}{at which digit to round the numbers. Defaults to 2.}
}
\description{
Returns beta coefficients and confidence intervals from a fitted lm() model.
}
\examples{
lm_CI()
}
\keyword{model}
\keyword{modeling,}
\keyword{summary,}

