% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{lm_get_fits}
\alias{lm_get_fits}
\title{Get R2 and R2 adj. for each model.}
\usage{
lm_get_fits(model_list)
}
\arguments{
\item{model_list}{(list) A list of model fits e.g. from lm().}
}
\description{
Returns a data.frame with each models R2 and R2 adj.
}
\examples{
lm_get_fits()
}

