% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{load_packages}
\alias{load_packages}
\title{Load packages}
\usage{
load_packages(...)
}
\arguments{
\item{...}{Unquoted package names separated by comma}
}
\value{
On success invisibly returns TRUE, otherwise an error.
}
\description{
Load packages
}
\examples{
load_packages(ggplot2)
testthat::expect_error(load_packages(ggplot3))
}
