% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirt.R
\name{loading_to_slope}
\alias{loading_to_slope}
\title{Convert from slopes to factor loadings}
\usage{
loading_to_slope(x, logit_scaling = T)
}
\arguments{
\item{x}{A vector of slopes}

\item{logit_scaling}{Whether to use logit scaling}
}
\value{
A vector of loadings
}
\description{
Convert from slopes to factor loadings
}
