% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{locf}
\alias{locf}
\title{Last observation carried forward}
\usage{
locf(x, reverse = F)
}
\arguments{
\item{x}{A vector}

\item{reverse}{Whether to do it in reverse}
}
\value{
A vector
}
\description{
Last observation carried forward
}
\examples{
c(NA, 1, NA, 2, NA) \%>\% locf()
c(NA, 1, NA, 2, NA) \%>\% locf(reverse = T)
c(NA, 1, NA, 2, NA, NA, NA) \%>\% locf()
}
