% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_list_array}
\alias{make_list_array}
\title{Create list-array}
\usage{
make_list_array(...)
}
\arguments{
\item{...}{(any number of whole numbers or vectors) If given a whole number scalar, will use that to define the length of that dimension. If given a vector, will use the length of that to determine the length of the dimension and use the values as names.}
}
\value{
An n-dimensional list-array.
}
\description{
A convenience function to create a list-array. A list with dimensions like an array.
}
\examples{
make_list_array(1:3, letters[1:3], LETTERS[1:3])
}
