% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{make_xlsx_write_function}
\alias{make_xlsx_write_function}
\title{Make a function to write to an XLSX file with}
\usage{
make_xlsx_write_function(filename)
}
\arguments{
\item{filename}{(character scalar) The filename to write to. Requires full path.}
}
\description{
This creates a copy of writeWorksheet where the object is filled out. Uses the XLConnect package.
}
\examples{
#set up the function
write_to_test = make_xlsx_write_function("test.xlsx")
#test it by writing iris dataset to a sheet in the file
write_to_test(data = iris, sheet = "iris")
}

