% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mapvalues}
\alias{mapvalues}
\title{Map values}
\usage{
mapvalues(x, from, to, warn_missing = T)
}
\arguments{
\item{x}{A vector}

\item{from}{A vector to change from}

\item{to}{A vector to change to, can be same length or length 1 (recycled)}

\item{warn_missing}{Warn if from values not present}
}
\value{
A vector
}
\description{
Copy of plyr::mapvalues
}
\examples{
mapvalues(0:9, from = 3:5, to = c(-1, -1, -1))
mapvalues(0:9, from = 3:5, to = -1)
}
