% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{MAT_divide_rowwise}
\alias{MAT_divide_rowwise}
\title{Divide a matrix or df row-wise by a vector}
\usage{
MAT_divide_rowwise(data, divisor)
}
\arguments{
\item{data}{(mat or df) Data to divide.}

\item{divisor}{(num/int vectr) Divisor to use.}
}
\value{
Returns a data.frame/matrix with the orignal dimnames.
}
\description{
The default use of / divides column-wise, but sometimes row-wise is needed. This function restores the dimnames.
}
\examples{
m = matrix(1:10, ncol=2)
m
m/c(1, 2) #divide column-wise
MAT_divide_rowwise(m, c(1, 2)) #row-wise
#works on data.frames too
MAT_divide_rowwise(iris[-5], c(1, 999, 1, 999)) \%>\% head
}
