% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{MAT_find_size}
\alias{MAT_find_size}
\title{Determine the dimensions of a symmetric matrix from the length of a vector of the upper or lower half.}
\usage{
MAT_find_size(x, diag = F)
}
\arguments{
\item{x}{The length of the vector of one half of the matrix.}

\item{diag}{Whether the diagonal values are included or not. Defaults to F.}
}
\description{
Returns a single number which is the length of the sides of the original matrix.
}
\examples{
MAT_find_size()
}
\keyword{lower,}
\keyword{matrix,}
\keyword{size,}
\keyword{symmetric,}
\keyword{upper}

