% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{MAT_get_half}
\alias{MAT_get_half}
\title{Get half of a matrix.}
\usage{
MAT_get_half(x, lower = T, diag = F)
}
\arguments{
\item{x}{A matrix or matrix-coercable object.}

\item{lower}{Whether to extract the lower half. Defaults to T. If F, then the upper half is extracted.}

\item{diag}{Whether the diagonal values should be included or not. Defaults to F.}
}
\description{
Returns a vector with the values in one half of the matrix. Which half and whether diagonal values are included can be specified.
}
\examples{
MAT_get_half()
}
\keyword{half}
\keyword{lower,}
\keyword{matrix,}
\keyword{upper,}

