% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{MAT_half}
\alias{MAT_half}
\title{Get half of a matrix.}
\usage{
MAT_half(x, lower = T, diag = F)
}
\arguments{
\item{x}{A matrix or matrix-coercable object.}

\item{lower}{Whether to extract the lower half. Defaults to T. If F, then the upper half is extracted.}

\item{diag}{Whether the diagonal values should be included or not. Defaults to F.}
}
\description{
Returns a vector with the values in one half of the matrix. Which half and whether diagonal values are included can be specified.
}
\examples{
cor(iris[-5]) #can't summarize this data due to diagonal
MAT_half(cor(iris[-5])) #this data we can
}
