% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{MAT_half2full}
\alias{MAT_half2full}
\title{Half to full matrix}
\usage{
MAT_half2full(mat, lower = T, diag = F)
}
\arguments{
\item{mat}{(matrix) A symmetric matrix}

\item{lower}{(lgl) Whether to use the lower or higher triangle.}

\item{diag}{(lgl) Whether the diagonal values should be included or not.}
}
\value{
A matrix
}
\description{
From a half matrix, recreate a full symmetric matrix. Uses either lower or upper triangle.
}
\examples{
m = matrix(1:9, ncol=3)
rownames(m) = letters[1:3]
colnames(m) = LETTERS[1:3]
m \%>\% MAT_half2full
m \%>\% MAT_half2full(lower=F)
m \%>\% MAT_half2full(diag=T)
m \%>\% MAT_half2full(lower=F, diag=T)
}
