% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{math_to_function}
\alias{math_to_function}
\title{Convert a half-complete math condition to a function.}
\usage{
math_to_function(str, convert_equal = T, silent_try = T,
  test_function = T)
}
\arguments{
\item{str}{(a character scalar) A string of a half math conditional}

\item{convert_equal}{(boolean) Converts = to ==. Default=T.}

\item{silent_try}{(boolean) Whether to use a silent trial. Default=T. Change to F to get potentially useful debugging information.}

\item{test_function}{(boolean) Whether to test the function. This is done by trying on 0. Default=T.}
}
\description{
Takes a string like "<0" and outputs a function to test for that condition. Returns a boolean.
}
\examples{
math_to_function()
}
\keyword{condition}
\keyword{math,}

