% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{matrixplot2}
\alias{matrixplot2}
\title{Wrapper for matrixplot()}
\usage{
matrixplot2(df)
}
\arguments{
\item{df}{a data.frame.}
}
\description{
Wrapper for matrixplot() from VIM, but automatically substrings the variable names to 8 characters.
}
\examples{
matrixplot2()
}
\keyword{VIM}
\keyword{data,}
\keyword{matrix,}
\keyword{missing}
\keyword{plot,}

