% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_operations.R
\name{mean_abs_diff}
\alias{mean_abs_diff}
\title{Calculate mean absolute difference for values in a vector}
\usage{
mean_abs_diff(x, na.rm = T)
}
\arguments{
\item{x}{(numeric vector) A vector of values.}

\item{na.rm}{(logical scalar) Whether to ignore missing values. Default=TRUE.}
}
\description{
Calculates all the pairwise absolute differences, then averages them.
}
\examples{
mean_abs_diff(iris[[1]])
}
