% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Dataset merger function}
\usage{
merge_datasets(DF1, DF2, main = 1, time = F, join = "both")
}
\arguments{
\item{DF1}{the first data.frame}

\item{DF2}{the second data.frame}

\item{main}{which data.frame should be used as the main? Choose the larger one if working with large datasets. Default to using neither.}

\item{join}{(character scalar) Which data.frames to use cases from. Defaults to "both". Can be: both, left, right.}
}
\description{
This function allows you to merge two data.frames by their overlapping rownames.
}
\examples{
merge_datasets()
}
\keyword{combining}
\keyword{data.frame}
\keyword{datasets}
\keyword{merging}

