% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{merge_rows}
\alias{merge_rows}
\title{Merge rows in data.frame.}
\usage{
merge_rows(df, key, func = sum, numeric = TRUE)
}
\arguments{
\item{data}{(data.frame or matrix) The data object.}

\item{resid.vars}{(character vector) The names of the variables to use for the residualization.}

\item{exclude.resid.vars}{(logical) Whether to exclude the residualization variables from residualization. Defaults to TRUE.}

\item{return.resid.vars}{(logical) Whether to include the residualization variables in the returned data.frame. Defaults to TRUE.}

\item{print.models}{(logical) Wether to print the lm models used in the process. Defaults to TRUE.}

\item{exclude_vars}{(character vector) Names of variables are that excluded from the residualization.}
}
\description{
Find blocks of rows with a matching
}
\examples{
t = data.frame(key = c("a", "a", "b", "b", "c"), value = 1:5)
merge_rows(t, "key") #rows merged by sum
merge_rows(t, "key", func = mean) #rows merged by mean
}
\keyword{modeling,}
\keyword{partialing}
\keyword{residualize,}

