% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{merge_rows_by_name}
\alias{merge_rows_by_name}
\title{Merge rows in data.frame by name.}
\usage{
merge_rows_by_name(df, names, new_name, func = mean, numeric = TRUE)
}
\arguments{
\item{df}{(data.frame or matrix) The data object.}

\item{names}{(character vector) The rownames to merge.}

\item{new_name}{(character scalar) The new rownames to use. Defaults to the first member of the names parameter.}

\item{func}{(function) The function to use. Note that if you set numeric = FALSE, then the function must be able to handle non-numeric data. Defaults to mean.}

\item{numeric}{(logical scalar) Whether to apply the function only to the numeric columns. Default=TRUE.}
}
\description{
Find blocks of rows with a rowname that matches a given vector of names and merge them into one.
}
\examples{
#suppse you had a data.frame with data for multiple variables
#but accidentally, one observation was given two names, "C" and "D".
#and data has been dispersed among the rows
#we can move all the data into one row without data loss.
t1 = data.frame(X = c(1, 2, 3, NA), Y = c(1, 2, NA, 3));rownames(t1) = LETTERS[1:4]
t1
#here the real values for the C observation are both 3, but it has accidentally been called "D".
merge_rows_by_name(df = t1, names = c("C", "D"), func = mean)
}

