% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_extract_data}
\alias{meta_extract_data}
\title{Extract data from \code{rma} object}
\usage{
meta_extract_data(.rma)
}
\arguments{
\item{.rma}{(rma object) The rma analysis from metafor.}
}
\value{
A data frame.
}
\description{
Extract data from fitted meta-analysis of type \code{rma}.
}
\details{
Extract effect size, sampling variance/standard error from \code{rma} object. Then adds a couple of more useful columns with z scores, p values (two-tailed).
}
\examples{
library(metafor); data(european_ancestry)
meta = rma(european_ancestry$r, sei = european_ancestry$SE_r)
meta_extract_data(meta)
}
