% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_pcurve}
\alias{meta_pcurve}
\title{P-curve plot with ggplot2}
\usage{
meta_pcurve(.analysis, p_cutoff = 0.05, print_plot = T,
  binom_method = "exact", text_pos = "tl")
}
\arguments{
\item{.analysis}{(rma object) The rma analysis from \link{metafor}.}

\item{p_cutoff}{(num) P-value cutoff to use.}

\item{print_plot}{(lgl) Print the plot?}

\item{binom_method}{(chr) Which method to use to calculate binomial p-value.}
}
\description{
Makes a pretty p-curve plot using \link{ggplot2}.
}
\examples{
library(metafor)
data(european_ancestry)
meta = rma(european_ancestry$r, sei = european_ancestry$SE_r)
meta_pcurve(meta)
}
