% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_add_random}
\alias{miss_add_random}
\title{Insert random NAs into a data.frame.}
\usage{
miss_add_random(df, prop = 0.1)
}
\arguments{
\item{df}{(data.frame) A data.frame.}

\item{prop}{(num vector) The proportion of NAs to add.}
}
\value{
A a data.frame.
}
\description{
Inserts missing data into a data.frame at random, thus creating data that are Missing Completely At Random (MCAR). THis isn't how data usually are missing in the real world, but may be sufficient for some situations.
}
\examples{
df = data.frame(c(1:10), letters[1:10]) #example data
miss_add_random(df) #add 10\% random NAs
miss_amount(df) #verify that 10\% really is missing
}

