% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_amount}
\alias{miss_amount}
\title{How much data is missing?}
\usage{
miss_amount(data, round = 2)
}
\arguments{
\item{data}{(data.frame/matrix) The data.}
}
\value{
A length 3 vector with the proportions.
}
\description{
Calculates the proportions of missing data by case, variable and cell.
}
\examples{
miss_amount(df_addNA(iris))
}

