% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_by_case}
\alias{miss_by_case}
\title{Missing datapoint counter, case-level}
\usage{
miss_by_case(x)
}
\arguments{
\item{x}{(data.frame/matrix) The data.}
}
\description{
Counts the number of missing datapoints per case
}
\examples{
miss_by_case(miss_add_random(iris))
}

