% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_by_group}
\alias{miss_by_group}
\title{Calculate proportion missing data by group variables}
\usage{
miss_by_group(data, grouping_vars, vars = NULL)
}
\arguments{
\item{data}{Data frame}

\item{grouping_vars}{Names of grouping variables}

\item{vars}{Names of data variables}
}
\value{
A data frame
}
\description{
Calculate proportion missing data by group variables
}
