% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_by_var}
\alias{miss_by_var}
\alias{miss_table}
\title{Missing datapoint counter, variable-level}
\usage{
miss_by_var(x)
}
\arguments{
\item{x}{a matrix or data.frame}
}
\description{
Counts the number of missing datapoints per variable
}
\examples{
miss_by_var(df_addNA(iris))
}

