% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_case}
\alias{miss_case}
\title{Missing datapoint counter, case-level}
\usage{
miss_case(x)
}
\arguments{
\item{x}{a matrix or data.frame}
}
\description{
Counts the number of missing datapoints per case
}
\examples{
miss_case()
}
\keyword{data}
\keyword{missing}

