% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_count}
\alias{miss_count}
\title{Count missing data}
\usage{
miss_count(x, reverse = F, prop = F)
}
\arguments{
\item{x}{(any suitable object) An object for which to count NAs.}

\item{reverse}{(logical scalar) Whether to count non-NAs instead (default false).}
}
\value{
An integer.
}
\description{
A simple wrapper for \code{\link{is.na}} and \code{\link{sum}}.
}
\examples{
m = matrix(c(NA, 1:3, NA, 4:6, NA), nrow = 3)
miss_count(m)
miss_count(m, reverse = T)
miss_count(m, prop = T)
miss_count(m, reverse = T, prop = T)
}
