% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_drop_vars}
\alias{miss_drop_vars}
\title{Drop variables with high missingness}
\usage{
miss_drop_vars(data, na_threshold = 1, inclusive = F)
}
\arguments{
\item{data}{A data frame.}

\item{na_threshold}{Threshold for allowed missingness. Default = 1 (drop variables with completely missing data)}

\item{inclusive}{Whether to use inclusive comparison (<=) or exclusive (<).}
}
\value{
A data frame
}
\description{
Drop variables with high missingness
}
\examples{
iris \%>\% miss_add_random(0.5) \%>\% miss_drop_vars(0.5) #will drop variables with more than 50\% missing, which is approximately half at random
}
