% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_fill}
\alias{miss_fill}
\title{Fill in missing values based on other variables/vectors}
\usage{
miss_fill(...)
}
\arguments{
\item{...}{Can be a data frame, list of vectors, or just vectors. Vectors must have same length.}
}
\value{
A vector with imputed values
}
\description{
Handy function to imputing missing values based on values from other sources instead of statistical methods.
}
\examples{
list(c(1, NA, NA), c(9, 2, NA), c(9, 9, 3)) \%>\% miss_fill()
}
