% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_matrix}
\alias{miss_matrix}
\title{Missing data matrix}
\usage{
miss_matrix(data)
}
\arguments{
\item{data}{(data.frame/matrix) The data.}
}
\value{
A matrix where missing/present values are coded as 1/0.
}
\description{
Create the missing data matrix. Not so useful in itself, but useful for other functions.
}
\examples{
miss_matrix(miss_add_random(iris))
}
