% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_pattern}
\alias{miss_pattern}
\title{Calculates the patterns of missing data}
\usage{
miss_pattern(data)
}
\arguments{
\item{data}{(data.frame/matrix) The data.}
}
\value{
A vector of patterns the same length as the number of rows in data.
}
\description{
Calculates the patterns of missing data. This is the missing data matrix multiplied row-wise by a vector of the 2 to the 1 to n power, where n is the number of variables. This results in a unique id for each way the data can be missing.
}
\examples{
miss_pattern(df_addNA(iris))
}

