% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_patterns}
\alias{miss_patterns}
\title{Calculates the patterns of missing data}
\usage{
miss_patterns(df, digits = 3)
}
\arguments{
\item{data}{(data.frame/matrix) The data.}
}
\value{
A data frame showing the pattern of missing data in variables
}
\description{
Calculates the patterns of missing data
}
\examples{
miss_patterns(miss_add_random(iris))
}
