% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_table}
\alias{miss_table}
\title{Missing datapoint counter, variable-level}
\usage{
miss_table(x)
}
\arguments{
\item{x}{a matrix or data.frame}
}
\description{
Counts the number of missing datapoints per variable
}
\examples{
miss_table()
}
\keyword{data}
\keyword{missing}

