% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_add_str_to_formula}
\alias{MOD_add_str_to_formula}
\title{Add terms to a formula from character vector}
\usage{
MOD_add_str_to_formula(formula., terms., env. = NULL)
}
\arguments{
\item{formula.}{A formula to modify.}

\item{terms.}{Terms to add.}

\item{env.}{Optionally, input to \code{\link[base:parent.frame]{base::parent.frame()}} so as to use another environment than the function's.}
}
\value{
a formula
}
\description{
Add terms to a formula from character vector.
}
\examples{
MOD_add_str_to_formula(y~x, "z")
MOD_add_str_to_formula(y~x, c("z", "t", "w"))
#change environment
MOD_add_str_to_formula(y~x, "z", 1)
}
