% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_APSLM}
\alias{MOD_APSLM}
\title{Fit all possible simple linear models}
\usage{
MOD_APSLM(dependent, predictors, data, standardized = T, .weights = NA,
  messages = F, progress = T, cv_runs = 20)
}
\arguments{
\item{dependent}{(chr scalar) The name of the dependent variable.}

\item{predictors}{(chr vector) The names of the redictor variables.}

\item{data}{(data.frame) A data.frame with the variables.}

\item{standardized}{(log scalar) Whether to standardize the results. Defaults to true.}

\item{.weights}{(num vector) A numeric vector of weights to use. Defaults to NA, which causes it to use unit weights for all cases.}

\item{messages}{(log scalar) Whether to show messages with the models. Default no.}

\item{progress}{(lgl scalar) Whether to show a progress bar. Default yes.}

\item{cv_runs}{(num scalar) How many runs to use in cross-validation. Default 20. Increase this value for smaller datasets to get more reliable results.}
}
\value{
Returns a list with two elements: 1) A data.frame with betas and other summary statistics from the models. 2) A list with all the models.
}
\description{
Fits all possible simple (no interactions) linear models using \code{\link{lm}}.
}
\details{
Use \code{\link{MOD_summarize_models}} on the data.frame with model summary information to get an overview of the results.
}
\examples{
#try all models in iris dataset to predict sepal length
MOD_APSLM(dependent = "Sepal.Length", predictors = c("Sepal.Width", "Petal.Length", "Petal.Width", "Species"), data = iris)
}
