% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_etas}
\alias{MOD_etas}
\title{Get etas from analysis of variance}
\usage{
MOD_etas(fitted_model)
}
\description{
Converts a \code{lm} or \code{glm} to an analysis of variance, and calculates the etas (square rooted values of traditional eta^2).
}

