% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_k_fold_r2}
\alias{MOD_k_fold_r2}
\title{Calculate k fold cross validated r2}
\usage{
MOD_k_fold_r2(lmfit, folds = 10, runs = 20, seed = 1, progress = interactive())
}
\arguments{
\item{lmfit}{(lm fit) A lm fit object.}

\item{folds}{(whole number scalar) The number of folds to use (default 10).}

\item{runs}{(int scalar) The number of runs (default 20).}

\item{seed}{(int scalar) The seed for the random number generator (default 1). This ensures reproducible results.}

\item{progress}{(lgl scalar) Whether to show a progress bar, default yes.}
}
\description{
Using k fold cross-validation, estimate the true r2 in a new sample. This is better than using adjusted r2 values.
}
\examples{
fit = lm("Petal.Length ~ Sepal.Length", data = iris) #a fit
fit_wtd = lm("Petal.Length ~ Sepal.Length", data = iris, weight = Sepal.Width) #with weights
MOD_k_fold_r2(fit) #raw r2 and cv r2
MOD_k_fold_r2(fit_wtd) #different r2 due to weights
}
