% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_LASSO}
\alias{MOD_LASSO}
\title{Repeatedly run glmnet.cv()}
\usage{
MOD_LASSO(data, dependent, predictors, weights_ = NA, standardize = T,
  runs = 100, alpha_ = 1, NA_ignore = T, seed = 1, messages = T,
  progress = T)
}
\arguments{
\item{data}{(data.frame) A data.frame with the data. Must contain dependent and predictor variables.}

\item{dependent}{(chr scalar) The name of the dependent variable.}

\item{predictors}{(chr vector) The names of the predictor variables.}

\item{weights_}{(num vector) If weights should be used, a numeric vector of values to use. Defaults to equal weights.}

\item{standardize}{(log scalar) Whether to standardize the data beforehand. Defaults to true.}

\item{runs}{(int scalar) Number of times to run. Defaults to 100.}

\item{alpha_}{(num scalar) The penalty to use. 1 = lasso regression, 0 = ridge regression. Defaults to 1.}

\item{NA_ignore}{(log scalar) Whether to remove cases with missing data. Defaults to T.}

\item{seed}{(int scalar) The seed to use (default 1). For reproducible results.}

\item{messages}{(log scalar) Whether to show messages. Default yes.}

\item{progress}{(log scalar) Whether to show a progress bar. Default yes.}
}
\description{
Returns a data frame of beta coefficients from glmnet.cv() fits.
}
\details{
The messages parameter is depreceated in favor of the progress parameter.
}
\examples{
MOD_LASSO(iris, "Sepal.Length", predictors = colnames(iris)[-1])
}

