% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{MOD_partial}
\alias{MOD_partial}
\title{Calculate a partial correlation.}
\usage{
MOD_partial(df, x, y, z, weights_var)
}
\arguments{
\item{df}{A data.frame.}

\item{x}{String with the name of the first variable.}

\item{y}{String with the name of the second variable.}

\item{z}{String with the name of the control variable.}

\item{weights}{String with the name of the weights variable. Can be left out.}
}
\description{
Calculates the partial correlation.
}
\examples{
MOD_partial()
}

