% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_summarize_models}
\alias{MOD_summarize_models}
\title{Summarize model coefficients}
\usage{
MOD_summarize_models(df, digits = 3, desc = c("mean", "median", "sd",
  "fraction_zeroNA"), include_intercept = F)
}
\arguments{
\item{df}{A data frame with betas for predictors across models.}

\item{digits}{The number of digits to round the results to. Defaults to 3.}

\item{desc}{A character vector of the desired descriptive statistics. These are extracted using describe() from the psych package. Defaults to c("mean", "median", "sd").}
}
\description{
Calculates summary statistics for model betas.
}
\examples{
MOD_summarize_models()
}
\keyword{describe}
\keyword{fit,}
\keyword{model,}
\keyword{summary,}

