% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_summary}
\alias{MOD_summary}
\title{Convenient summary of a \code{\link{lm}} or \code{\link{glm}} with analytic confidence intervals.}
\usage{
MOD_summary(fitted_model, level = 0.95, standardize = T, kfold = T,
  folds = 10, runs = 20, ...)
}
\arguments{
\item{fitted_model}{(lm or glm) The fitted model.}

\item{level}{(num scalar) The level of confidence to use.}

\item{standardize}{(log scalar) Whether to report standardized betas.}

\item{kfold}{(log scalar) Whether to also calculate a k fold cross-validated r2 value.}

\item{folds}{(num scalar) The number of folds to use if using cross-validation.}

\item{runs}{(int scalar) The number of runs to use for cross-validation.}

\item{...}{(other args) Other arguments passed to \code{\link{MOD_k_fold_r2}}.}

\item{round}{(num scalar) At which digit to round the numbers.}
}
\description{
Tidys information from linear models or generalized linear models.
}
\examples{
fit1 = lm("Sepal.Length ~ Sepal.Width + Petal.Length", iris)
#then summarize the model
MOD_summary(fit1) #unstd. data, std. betas
MOD_summary(fit1, standardize = F) #unstd. data, don't std. betas, similar to base \\code{summary}
}
