% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{NA_to_F}
\alias{NA_to_F}
\title{\code{NA} to \code{FALSE}}
\usage{
NA_to_F(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of the same length as x
}
\description{
Helpful small function when dealing with non-tidyverse subsetting.
}
\examples{
x = c(T, F, NA, T)
NA_to_F(x)
}
