% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{named_vector_to_df}
\alias{named_vector_to_df}
\title{Convert a named data to a data frame}
\usage{
named_vector_to_df(x, name_col = "name", value_col = "value")
}
\arguments{
\item{x}{A vector}

\item{name_col}{The name of the variable to use for the names}

\item{value_col}{The name of the variable to use for the values}
}
\value{
A data frame
}
\description{
Convert a named data to a data frame
}
\examples{
named_vector_to_df(c(a = 1, b = 2))
named_vector_to_df(c(a = 1, b = 2), "x", "y")
}
