% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{named_vectors_to_df}
\alias{named_vectors_to_df}
\title{Reshape named vectors to a data.frame.}
\usage{
named_vectors_to_df(list, name_suffix = "_name", value_suffix = "_value")
}
\arguments{
\item{list}{(a list of vectors) The list of vectors.}

\item{name_suffix}{(character scalar) The suffix to use on the names.}

\item{valie_suffix}{(character scalar) The the suffix to use on the values.}
}
\description{
Construct a data.frame from a list of named vectors by filling in the shorter vectors with NAs.
}
\examples{
l = list(A = c(a = 1, b = 2, c = 3), B = c(a = 3, b = 2, c = 1))
named_vectors_to_df(l)
}
\keyword{data.frame,}
\keyword{reshape}
\keyword{vectors,}

