% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ndims}
\alias{ndims}
\title{How many dimensions does an object have?}
\usage{
ndims(x)
}
\arguments{
\item{x}{(an object) An object.}
}
\description{
Useful wrapper for \code{\link[=get_dims]{get_dims()}}.
}
\examples{
#flexible
ndims(1:3) #vector
ndims(iris) #data frame
ndims(array(1:27, dim = rep(3, 3))) #array
ndims(mean) #function
}
