% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{new_lines_adder}
\alias{new_lines_adder}
\title{Insert newlines into text every nth character.}
\usage{
new_lines_adder(x, interval)
}
\arguments{
\item{x}{A character string.}

\item{interval}{How often the newlines are added.}
}
\description{
Returns a character string with newlines every nth character. See also add_newlines().
}
\examples{
set.seed(2)
new_lines_adder(paste0(sample(c(letters, " "), size = 100, replace = T), collapse = ""), interval = 30)
}
