% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{object_to_string}
\alias{object_to_string}
\title{Object to string}
\usage{
object_to_string(x)
}
\arguments{
\item{x}{(an object) An object to convert to a string.}
}
\description{
Converts an object to a string using deparse and substitute.
}
\examples{
object_to_string()
}
\keyword{call,}
\keyword{convert}
\keyword{object,}
\keyword{string,}

