% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{overlap_metrics}
\alias{overlap_metrics}
\title{Calculate overlap metrics}
\usage{
overlap_metrics(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
A named vector of metrics of overlap
}
\description{
Calculate overlap metrics
}
\examples{
overlap_metrics(
c(1:3),
c(3:6)
)
}
