% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{p_to_asterisk}
\alias{p_to_asterisk}
\title{Add asterisks to p values}
\usage{
p_to_asterisk(
  x,
  asterisks = c(0.01, 0.005, 0.001),
  digits = max(count_decimals(asterisks)),
  asterisks_only = F
)
}
\arguments{
\item{x}{p values vector}

\item{asterisks}{Thresholds}

\item{digits}{Number of digits}

\item{asterisks_only}{Only print asterisks}
}
\value{
character vctor
}
\description{
Add asterisks to p values
}
\examples{
c(.123, .009, .004, .0009, .0001) \%>\% p_to_asterisk()
c(.123, .009, .004, .0009, .0001) \%>\% p_to_asterisk(asterisks = c(.05, .01))
c(.123, .009, .004, .0009, .0001) \%>\% p_to_asterisk(asterisks = c(.05))
c(.123, .009, .004, .0009, .0001) \%>\% p_to_asterisk(asterisks = c())
c(.123, .009, .004, .0009, .0001) \%>\% p_to_asterisk(asterisks = NULL)
c(.123, .009, .004, .0009, .0001) \%>\% p_to_asterisk(asterisks_only = T)
p_to_asterisk(NA)
}
