% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{percent_cutoff}
\alias{percent_cutoff}
\title{Find percentage of numbers that are above given thresholds.}
\usage{
percent_cutoff(x, cutoffs = c(0.3, 0.5), digits = 2)
}
\arguments{
\item{x}{(numeric vector) A vector of numbers.}

\item{cutoffs}{(numeric vector) A vector of thresholds. Default=(.30, .50)}

\item{digits}{(numeric scalar) The number of digits to round output to. Default=2.}
}
\description{
Takes a numeric vector and a numeric vector of thresholds. Returns the percent of numbers in the first above each of the numbers in the second.
}
\examples{
percent_cutoff()
}
\keyword{percent,}
\keyword{proportion}
\keyword{threshold,}

