% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_kmeans}
\alias{plot_kmeans}
\title{Scatter plot with kmeans clustering}
\usage{
plot_kmeans(df, clusters, runs, standardize = T)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{clusters}{The number of clusters to find.}

\item{runs}{Number of runs to use. The best run is used in the plot.}

\item{standardize}{Whether to standardize the data first. Defaults to TRUE.}
}
\description{
Performs kmeans and factor analysis on a dataset. Then plots a scatter plot using the first two factors (orthogonal) and color codes cases by their kmeans cluster.
}
\examples{
plot_kmeans()
}
\keyword{density,}
\keyword{ggplot2,}
\keyword{histogram}
\keyword{plot,}

