% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot factor loadings.}
\usage{
plot_loadings(fa.object, reverse = F, text_pos = "tl")
}
\arguments{
\item{fa.object}{a factor analysis object from the fa() function from the psych package.}

\item{reverse}{whether to reverse all loadings. Default to false.}

\item{text_pos}{which corner to write the numerical results in. Options are "tl", "tr", "bl", "br". Defaults to "tl".}
}
\description{
Returns a ggplot2 plot with sorted loadings and numerical results in a corner. Supports reversing of the factor is reversed.
}
\examples{
library(psych)
plot_loadings(fa(iris[-5])) #plot loadings from analysis of iris data
}

