% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{plot_miss}
\alias{plot_miss}
\title{Missing data histogram with ggplot2.}
\usage{
plot_miss(df, percent = T)
}
\arguments{
\item{df}{a data.frame.}

\item{percent}{whether to use percent or not. Defaults to true.}
}
\description{
Returns a ggplot2 histogram plot.
}
\examples{
plot_miss()
}
\keyword{data,}
\keyword{histogram,}
\keyword{missing}
\keyword{plot}

