% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{+}
\alias{+}
\title{Easy character concatenation}
\usage{
`+`(x, y)
}
\arguments{
\item{x}{(character vector) A character vector.}

\item{y}{(character vector) A character vector.}
}
\value{
A character vector. Note that it will have a length longer than one if one of the inputs has that.
}
\description{
A wrapper for stringr's \code{\link{str_c}} and the primitive "+" function.
}
\examples{
#digits still work
1+1
#characters work too
"str" + "ing"
#mixes are converted to characters
"123" + 456
#longer than length 1
1:2 + "a"
}
