% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{pool_sd}
\alias{pool_sd}
\title{Pooled sd}
\usage{
pool_sd(x, group)
}
\arguments{
\item{x}{(numeric vector) The data.}

\item{group}{(vector) Group membership.}
}
\description{
Calculate pooled sd.
}
\examples{
#Wikipedia's example https://en.wikipedia.org/wiki/Pooled_variance
v_test_vals = c(31, 30, 29, 42, 41, 40, 39, 31, 28, 23, 22, 21, 19, 18, 21, 20, 19, 18, 17)
v_test_group = c(rep(1, 3), rep(2, 4), rep(3, 2), rep(4, 5), rep(5, 5))
pool_sd(v_test_vals, v_test_group)
}

