% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{print.model_summary}
\alias{print.model_summary}
\title{Custom printing function for model_summary class}
\usage{
\method{print}{model_summary}(x)
}
\description{
Custom printing function for model_summary class
}
