% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{product}
\alias{product}
\title{Calculate the product}
\usage{
product(...)
}
\arguments{
\item{...}{(any number of numbers) The numbers. Can be multiple arguments, or one argument that is a vector.}
}
\value{
A whole number.
}
\description{
Calculate the product of a given set of numbers.
}
\examples{
product(1:3)
product(1, 2, 3)
}

