% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{prop_tests}
\alias{prop_tests}
\title{Compute many proportion tests}
\usage{
prop_tests(
  x,
  group,
  correct = T,
  conf_level = 0.95,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x}{A factor variable}

\item{group}{A grouping variable}

\item{correct}{a logical indicating whether Yates' continuity correction should be applied where possible.}

\item{conf_level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1. Only used when testing the null that a single proportion equals a given value, or that two proportions are equal; ignored otherwise.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter. Only used for testing the null that a single proportion equals a given value, or that two proportions are equal; ignored otherwise.}
}
\value{
A data frame, which may have missing levels if a combination did not exist in the data.
}
\description{
This is a convenient tidy wrapper for \code{stat::prop.test()}
}
\examples{
prop_tests(mpg$cyl, mpg$manufacturer)
}
