% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{proportion_true}
\alias{proportion_true}
\title{Proportion true}
\usage{
proportion_true(x)
}
\arguments{
\item{x}{(numeric vector) A vector of values.}
}
\description{
Given a suitable input (logical vector/matrix/array, numeric/integer vector/matrix/array), find the proportion of true values. Automatically removes NAs.
}
\examples{
#sample some logical data
x = sample(c(T, F), size = 100, replace = T)
proportion_true(x)
#sample some logical data with NAs
x = sample(c(T, F, NA), size = 100, replace = T)
proportion_true(x) #no errors or NA output
#converts input to logical if possible
x = sample(c(1, 0), size = 100, replace = T)
proportion_true(x)
#throws errors if it gets a nonsensical input
"str" \%>\% proportion_true() #character
data.frame() \%>\% proportion_true() #data.frame
list() \%>\% proportion_true()
}

