% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{quantile_smooth}
\alias{quantile_smooth}
\title{Quantile smoothening}
\usage{
quantile_smooth(
  x,
  y,
  quantile,
  method = c("qgam", "Rq", "running"),
  k = 5,
  window = 50
)
}
\arguments{
\item{x}{Predictor variable}

\item{y}{Outcome variable}

\item{quantile}{Which quantile (e.g., .95)}

\item{method}{Method to use}

\item{k}{Number of knots for method qgam, see \link[qgam:qgam]{qgam::qgam}}

\item{window}{Window size for method running, see \link[caTools:runquantile]{caTools::runquantile}}
}
\value{
Vector of fitted values
}
\description{
Quantile smoothening
}
\examples{
quantile_smooth(iris$Petal.Length, iris$Sepal.Length, quantile = .90)
}
