% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/statistics.R
\name{rcorr2}
\alias{rcorr2}
\title{Wrapper for rcorr()}
\usage{
rcorr2(x, ...)
}
\arguments{
\item{x}{a matrix or data.frame}
}
\description{
A wrapper for rcorr() that takes data.frames or matrices as opposed to only matrices.
}
\examples{
rcorr2()
}
\keyword{correlation}
\keyword{matrix}
\keyword{wrapper}

