% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{read_clipboard}
\alias{read_clipboard}
\title{Read data frame from clipboard}
\usage{
read_clipboard(
  header = T,
  sep = "\\t",
  na.strings = c("", "NA"),
  check.names = T,
  stringsAsFactors = F,
  dec = ".",
  ...
)
}
\arguments{
\item{header}{(lgl) Whether to use the first row as headers (column names).}

\item{sep}{(chr) Which separator to use.}

\item{na.strings}{(chr) Which cell values to interpret as missing data.}

\item{check.names}{(lgl) Whether to convert illegal names to legal ones using \code{\link[base:make.names]{base::make.names()}}.}

\item{stringsAsFactors}{(lgl) Whether to automatically convert strings to factors (GOD PLEASE NO).}

\item{dec}{(chr) Which symbol to interpret as decimal separator.}

\item{...}{Any other arguments passed to \code{\link[utils:read.table]{utils::read.table()}}.}
}
\description{
A wrapper function to \code{\link{write.table}} for reading from the clipboard. Useful to quickly read data copied from a table or spreadsheet.
}
\examples{
iris[-5] \%>\% cor \%>\% write_clipboard
iris \%>\% head \%>\% miss_add_random \%>\% write_clipboard
}
