% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{read_mega}
\alias{read_mega}
\title{A wrapper for read.csv()}
\usage{
read_mega(filename)
}
\arguments{
\item{filename}{the file to be read}
}
\description{
This is a convenient wrapper for read.csv()
}

