% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{read_rvest}
\alias{read_rvest}
\title{Read rvest/xml object}
\usage{
read_rvest(path)
}
\arguments{
\item{path}{(str) Where to read from.}

\item{...}{Other parameters to read_rds.}
}
\value{
An XML object.
}
\description{
Reads an rvest/xml object from disk. This is a wrapper around read_rds.
}

