% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{read_vcf}
\alias{read_vcf}
\title{Read VCF file into data frame}
\usage{
read_vcf(x, var_id = "ID", n_max = Inf)
}
\arguments{
\item{x}{path to VCF}

\item{var_id}{Which column(s) to use for IDs. Can be 'ID' or 'chrpos'}

\item{n_max}{Max lines to read (including comments)}
}
\value{
a data frame
}
\description{
Read VCF file into data frame
}
