% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{remove_NA_vars}
\alias{remove_NA_vars}
\title{Remove NA columns}
\usage{
remove_NA_vars(data, keep)
}
\arguments{
\item{data}{(data.frame) Object to subset.}

\item{keep}{(num or chr vector) A vector of names of columns to keep no matter what. Can also be numeric indices which are then replaced with the colnames.}
}
\value{
Returns the subset of the object.
}
\description{
Detect variables with no data in a data.frame and exclude them. Can be given a vector of names or indices of columns to keep no matter what.
}

