% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/statistics.R
\name{remove_redundant_vars}
\alias{remove_redundant_vars}
\title{Remove the n most redundant variables from a data.frame.}
\usage{
remove_redundant_vars(df, num.to.remove = 1, remove.method = "s")
}
\arguments{
\item{df}{a data.frame.}

\item{num.to.remove}{the number of variables to remove.}

\item{remove.method}{the method to use to remove variables. Methods are "c", "l", "r", "f" and "s": conversative, liberal, random, first or second.}
}
\description{
Removes the n top variables that highly correlated with another variable so as to avoid problems in analysis.
}
\examples{
remove_redundant_vars()
}
\keyword{analysis,}
\keyword{factor}
\keyword{latent}
\keyword{psychology,}
\keyword{psychometrics,}
\keyword{redundant}
\keyword{variable,}

