% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{remove_redundant_vars2}
\alias{remove_redundant_vars2}
\title{Remove redundant variables from a data.frame based on a threshold value. This is done by calculating all the intercorrelations, then finding those that correlate at or above the threshold (absolute value), then removing the second pair of each variable and not removing more variables than strictly necessary.}
\usage{
remove_redundant_vars2(df, threshold = 0.9, cor_method = "pearson")
}
\arguments{
\item{df}{A data.frame with numeric variables.}

\item{threshold}{A threshold above which intercorrelations are removed. Defaults to .9.}

\item{cor_method}{The correlation method to use. Parameter is fed to cor(). Defaults to pearson.}
}
\description{
Remove redundant variables from a data.frame based on a threshold value. This is done by calculating all the intercorrelations, then finding those that correlate at or above the threshold (absolute value), then removing the second pair of each variable and not removing more variables than strictly necessary.
}
\examples{
remove_redundant_vars()
}
\keyword{analysis,}
\keyword{factor}
\keyword{latent}
\keyword{psychometrics,}
\keyword{redundant,}
\keyword{variable}
\keyword{variable,}

