% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder columns in a data.frame by name}
\usage{
reorder_columns(data, vars)
}
\arguments{
\item{data}{(data.frame) The data.frame.}

\item{vars}{(named vector) The variables to reorder.}
}
\value{
The modified data.frame.
}
\description{
Reorder columns in a data.frame by name using a named vector.
}
\details{
Copied and modified from http://stackoverflow.com/a/37009127/3980197.
}
\examples{
#remove Species to front
reorder_columns(iris, c("Species" = 1))
#check for identity
all(reorder_columns(iris, c("Species" = 1)) == iris[c(5, 1:4)])
}

