% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{rescale}
\alias{rescale}
\title{Rescale numbers}
\usage{
rescale(x, new_min, new_max, old_min = min(x), old_max = max(x))
}
\arguments{
\item{x}{(num vectr) A vector of values.}

\item{new_min}{(num sclr) The new scale minimum.}

\item{new_max}{(num sclr) The new scale maximum.}

\item{old_min}{(num sclr) The old scale minimum. Default: min(x).}

\item{old_max}{(num sclr) The old scale maximum. Default: max(x).}
}
\description{
Rescales numbers from one scale to another. By default, assumes that the old scale has min and max values represented in the vector, but this can be overwritten.
}
\examples{
rescale(1:10, new_min = 0, new_max = 1) #converts to 0, ..., 1
rescale(1:10, new_min = 0, new_max = 5) #converts to 0, ..., 5
rescale(rnorm(10), new_min = 1, new_max = 100) \%>\% sort #converts to 1, ..., 1
rescale(c(.1, .5, 1), new_min = 10, new_max = 20, old_min = 0, old_max = 1) #assume old numbers belong to scale 0-1, rescale to 10-20 scale, i.e. 11, 15, 20
}

