% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{restore_NAs}
\alias{restore_NAs}
\title{Restore NAs in vector}
\usage{
restore_NAs(x, na_pos, len = NULL)
}
\arguments{
\item{x}{Non-NA values}

\item{na_pos}{Positions of NA, either logical of total length, or integer indexes}

\item{len}{Length of vector if known (optional)}
}
\value{
A vector with NAs restored in place
}
\description{
Restore NAs in a vector from known positions. Useful for dealing with NAs in functions and predictions.
}
\examples{
restore_NAs(c(1, 3, 5), c())
restore_NAs(c(1, 3, 5), 4:10)
restore_NAs(c(1, 3, 5), 1:5)
restore_NAs(c(1, 3, 5), c(2, 4, 6), 6)
restore_NAs(c(1, 3, 5), c(2, 4, 6))
}
