% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{retry_while_error}
\alias{retry_while_error}
\title{Retry while error}
\usage{
retry_while_error(
  expr,
  retry_interval = 60,
  silent = F,
  max_tries = 10,
  max_time = Inf
)
}
\arguments{
\item{expr}{An expression.}

\item{retry_interval}{(num) Time interval between retries in seconds.}

\item{silent}{(lgl) Whether to output dated errors while trying is ongoing.}

\item{max_tries}{(num) Max number of tries.}

\item{max_time}{(num) Max time to try, including run and sleep time.}
}
\value{
The output of expr.
}
\description{
Run an expression repeatedly until it doesn't produce an error or a max number of attempts has been reached.
}
\examples{
#this one eventually succeeds
retry_while_error(log(unlist(sample(list("", 0), size = 1, prob = c(1, .01)))), retry_interval = 0, max_time = Inf, max_tries = Inf, silent = T)

#note: these produce errors
retry_while_error(log(""), retry_interval = 0)
retry_while_error(log(""), retry_interval = 0, max_time = 1, max_tries = Inf)
}
