% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{reverse_scale}
\alias{reverse_scale}
\title{Reverse a scale}
\usage{
reverse_scale(x, .min = min(x), .max = max(x))
}
\arguments{
\item{x}{(numeric vector) A vector of values.}

\item{.min}{(numeric scalar) The scale maximum.}

\item{.max}{(numeric scalar) The scale minimum.}
}
\description{
Reverses a scale so that the largest number becomes the smallest, next largest becomes next smallest, etc. Can be given scale limits but will determine them empirically if not given.
}
\examples{
x = c(2, 1, 3, 0, -1) #some scale ranging from -1 to 3
reverse_scale(x) #reverse scores: 3 becomes -3
x == reverse_scale(reverse_scale(x)) #reverse twice and get back
reverse_scale(x, .min = -3, .max = 10) #supply other min/max
}

