% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{round_df}
\alias{round_df}
\title{Round numeric variables of a data frame.}
\usage{
round_df(df, digits = 3)
}
\arguments{
\item{df}{A data.frame or matrix.}

\item{digits}{The number of digits to round to.}
}
\description{
Returns a data.frame where numeric variables have been rounded to the desired number of digits.
}
\examples{
df = data.frame(rnorm(5), runif(5))
df
round_df(df)
}
\keyword{data.frame}
\keyword{round,}

