% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC.R
\name{SAC_control}
\alias{SAC_control}
\alias{SAC_control_all_methods}
\title{Control for spatial autocorrelation with multiple methods}
\usage{
SAC_control(df, dependent, predictors, knsn_k = 3, slr_k = 3, dists,
  lat_var, lon_var, distance_method, auto_detect_dist_method = T,
  SLR_weights_method = "inverse", SLR_include_self = F, SLR_central_measure,
  CD_weight_method = "harmonic", weights_var = "", methods = c("KNSNR",
  "SLR"), standardize = T, control_approach = c("partial"))
}
\arguments{
\item{df}{A data.frame with variables.}

\item{dependent}{A character vector with the name of the dependent variable.}

\item{predictors}{A character vector with the name of the predictor variables.}

\item{knsn_k}{The number of neighbors to use with k nearest spatial regression. Defaults to 3.}

\item{slr_k}{The number of neighbors to use with spatial local regression. Defaults to 3.}

\item{dists}{A matrix of distances between cases.}

\item{lat_var}{A string with the name of the variable which has the latitude/east-west data.}

\item{lon_var}{A string with the name of the variable which has the longitude/north-south data.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean. If using euclidean it doesn't matter which variable is coded as lat or lon.}

\item{auto_detect_dist_method}{Whether to try to autodetect the distance method. If the dataset contains variables with the names "lat" and "lon", it will be detected as spherical. If it contains "x" and "y", it will be detected as euclidean. Defaults to true.}

\item{SLR_weights_method}{A string with the name of the weights method to use for spatial local regression. Defaults to "inverse", i.e. weighing cases by their inverse distance to the case being considered. Can also be "none".}

\item{SLR_include_self}{Whether to include the self node or not in SLR. Defaults to F.}

\item{SLR_central_measure}{Which central tendency measure to meta-analyze SLR results with. Defaults to trim10, which is a 10 percent trimmed mean. For other options, see help for SAC_slr.}

\item{CD_weight_method}{A string indicating which averaging method to use when combining weights for cases. Defaults to harmonic mean. Other options are arithmic and geometric.}

\item{weights_var}{A character vector of the name of the weight variable to use. If none given, defaults to equal weights.}

\item{methods}{Which control methods to use. Options are CD, KNSNR, SLR. Defaults to all of them.}

\item{standardize}{Whether to standardize the data to get standardized betas. Defaults to T.}

\item{control_approach}{Which control approaches to use. Options are partial and mr. Defaults to partial.}
}
\description{
Returns a data.frame with predictors for linear regression with and without controls for SAC.
}

