% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAC.R
\name{SAC_control_all_methods}
\alias{SAC_control_all_methods}
\title{Perform spatial local regression.}
\usage{
SAC_control_all_methods(df, dependent, predictor, knsn_k = 3, slr_k = 3,
  dists, lat_var, lon_var, distance_method, auto_detect_dist_method = T,
  SLR_weights_method = "inverse", CD_weight_method = "harmonic",
  weights_var = "", methods_ = c("CD", "KNSNR", "SLR"))
}
\arguments{
\item{df}{A data.frame with variables.}

\item{dependent}{A character vector with the name of the dependent variable.}

\item{predictor}{A character vector with the name of the predictor variable.}

\item{knsn_k}{The number of neighbors to use with k nearest spatial regression. Defaults to 3.}

\item{slr_k}{The number of neighbors to use with spatial local regression. Defaults to 3.}

\item{dists}{A matrix of distances between cases.}

\item{lat_var}{A string with the name of the variable which has the latitude/east-west data.}

\item{lon_var}{A string with the name of the variable which has the longitude/north-south data.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean. If using euclidean it doesn't matter which variable is coded as lat or lon.}

\item{auto_detect_dist_method}{Whether to try to autodetect the distance method. If the dataset contains variables with the names "lat" and "lon", it will be detected as spherical. If it contains "x" and "y", it will be detected as euclidean. Defaults to true.}

\item{SLR_weights_method}{A string with the name of the weights method to use for spatial local regression. Defaults to "inverse", i.e. weighing cases by their inverse distance to the case being considered. Can also be "none".}

\item{CD_weight_method}{A string indicating which averaging method to use when combining weights for cases. Defaults to harmonic mean. Other options are arithmic and geometric.}
}
\description{
Returns a data.frame with measures of SAC using Moran's I, CD, CD_sqrt and KNSNR.
}
\examples{
SAC_control_all_methods()
}
\keyword{autocorrelation,}
\keyword{local}
\keyword{regression}
\keyword{regression,}
\keyword{spatial}

