% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{score_accuracy}
\alias{score_accuracy}
\title{Score accuracy of estimates.}
\usage{
score_accuracy(df, criteria, methods = c("pearson_r", "mean_abs_delta",
  "sd_error_abs", "mean_elevation_error_abs"))
}
\arguments{
\item{df}{(numeric data.frame) A data.frame with estimates. Rows must be cases.}

\item{criteria}{(numeric vector) A vector of criteria values to score estimates against.}

\item{methods}{(character vector) Which measures to return. Defaults to c("pearson_r", "mean_abs_delta", "sd_error_abs", "mean_elevation_error_abs"). Use "all" to get all.}
}
\description{
Calculates accuracy measures from a data.frame of estimates using a vector of criteria values.
}
\examples{
score_accuracy()
}
\keyword{accuracy}
\keyword{estimates,}
\keyword{score,}

