% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereotypes.R
\name{score_bias_metrics}
\alias{score_bias_metrics}
\title{Score stereotype bias metrics}
\usage{
score_bias_metrics(estimates, criterion, bias_var)
}
\arguments{
\item{estimates}{Data frame of estimates}

\item{criterion}{Vector of criterion values}

\item{bias_var}{Vector of values to bias metrics for}
}
\value{
A data frame of scored bias metrics
}
\description{
Score stereotype bias metrics
}
