% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{semi_par}
\alias{semi_par}
\title{Semi-partial correlation with weights}
\usage{
semi_par(x, y, z, weights = NA)
}
\arguments{
\item{x}{A numeric vector to correlate with y.}

\item{y}{A numeric vector to correlate with x after partialing out z.}

\item{z}{A numeric vector to partial out of y.}

\item{weights}{A numeric vector of weights to use. If none given, will return unweighted results.}
}
\description{
Returns the semi-partial correlation. Weights may be used.
}
\examples{
semi_par()
}
\keyword{correlation,}
\keyword{partial,}
\keyword{psychometrics,}
\keyword{semi-partial,}
\keyword{weights}

