% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{semi_par_serial}
\alias{semi_par_serial}
\title{Semi-partial correlation with weights}
\usage{
semi_par_serial(df, dependent, primary, secondaries, weights = NA,
  standardize = T)
}
\arguments{
\item{df}{A data frame with the variables.}

\item{dependent}{A string with the name of the dependent variable.}

\item{primary}{A string with the name of the primary predictor variable.}

\item{secondaries}{A character vector with the names of the secondary predictor variables.}

\item{weights}{A string with the name of the variable to use for weights.}

\item{standardize}{Whether to standardize the data frame before running results. The weights variable will not be standardized.}
}
\description{
Returns a table of semi-partial correlations where a dependent variable has first been regressed on the primary predictor variable, and then correlated with each of the secondary predictors in turn.
}
\examples{
semi_par_serial()
}
