% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{seq_along_rows}
\alias{seq_along_rows}
\title{Seq along rows}
\usage{
seq_along_rows(df)
}
\arguments{
\item{df}{(vector) A vector.}
}
\value{
An integer vector.
}
\description{
Equivalent to seq_along, but for rows.
}
\examples{
#example on iris
seq_along_rows(iris)
#does not fail on 0-row data frames
seq_along_rows(iris[-c(1:150), ])
}
