% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{silence}
\alias{silence}
\title{Silence warnings or messages from expression via parameter.}
\usage{
silence(
  expr,
  warnings = FALSE,
  messages = FALSE,
  startupmessages = FALSE,
  output = FALSE
)
}
\arguments{
\item{expr}{(expression) Some expression to run.}

\item{warnings}{(logical) Show warnings? Default=F.}

\item{messages}{(logical) Show messages? Default=F.}
}
\description{
If warnings and messages need to be toggleable, but there are many expressions that can give them and it would be cumbersome to add an if sentence for every expression.
}
\examples{
silence(warning())
silence(log(-1))
silence(warning("test"))
silence(warning("test"), warnings = T)
silence(message("test"))
silence(message("test"), messages = T)
}
