% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirt.R
\name{slope_to_loading}
\alias{slope_to_loading}
\title{Convert from factor loading to discrimination}
\usage{
slope_to_loading(x, logit_scaling = T)
}
\arguments{
\item{x}{Vector of factor loadings}

\item{logit_scaling}{Whether to use logit scaling}
}
\value{
A vector of slopes (discrimination)
}
\description{
Convert from factor loading to discrimination
}
