% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{SMD_matrix}
\alias{SMD_matrix}
\title{Standardized mean differences}
\usage{
SMD_matrix(x, group, central_tendency = mean, dispersion = "sd",
  dispersion_method = "all", ...)
}
\arguments{
\item{x}{(numeric vector) A vector of values.}

\item{group}{(vector) A vector of group memberships.}

\item{central_tendency}{(function) A function to use for calculating the central tendency. Must support a parameter called na.rm. Ideal choices: mean, median.}

\item{dispersion}{(character or numeric scalar) Either the name of the metric to use (sd or mad) or a value to use.}

\item{dispersion_method}{(character scalar) If using one of the built in methods for dispersion, then a character indicating whether to use the pooled value from the total dataset (all), the pairwise comparison (pair), or the sd from the total dataset (total).}

\item{...}{(other arguments) Additional arguments to pass to the central tendency function.}
}
\description{
Calculate standardized mean differneces between all groups.
}
\examples{
SMD_matrix(iris$Sepal.Length, iris$Species)
}

