% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_lag_cors}
\alias{spatial_lag_cors}
\title{Compute correlatoins for lag variables}
\usage{
spatial_lag_cors(x, suffix = "_lag", long_output = F)
}
\arguments{
\item{x}{A data frame}

\item{suffix}{The suffix on the lag variables}
}
\value{
A named vector
}
\description{
Compute correlatoins for lag variables
}
