% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{split_every_k}
\alias{split_every_k}
\title{Split vector every k elements}
\usage{
split_every_k(x, k, uneven = T)
}
\arguments{
\item{x}{(vector) A vector to split.}

\item{k}{(whole number scalar) Split every k elements.}

\item{uneven}{(logical scalar) Whether to accept a split that would be uneven. If yes, the last group will be smaller than the others. Defaults to TRUE.}
}
\description{
Split a vector every k elements. Returns a list.
}
\examples{
split_every_k(1:12, 4)
split_every_k(1:11, 4) #last group isnt as large as the others
}
\keyword{split}
\keyword{vector,}

