% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{split_text}
\alias{split_text}
\title{Insert newlines into text every nth character.}
\usage{
split_text(text, interval)
}
\arguments{
\item{text}{a character vector to interspace with newlines.}

\item{interval}{how often to insert the newline?}
}
\description{
Returns a character vector with newlines every nth character.
}
\examples{
split_text()
}
\keyword{combine,}
\keyword{data.frame,}
\keyword{lower}
\keyword{matrix,}
\keyword{upper,}

