% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector}
\usage{
standardize(x, w, robust = F, sample = T)
}
\arguments{
\item{x}{(num vector) A vector of values.}

\item{w}{(num vector) A vector of weights.}

\item{sample}{(log scalar) Whether this is a sample as opposed to a population (default true).}
}
\description{
Standardize a vector. Can use weights and robust measures of central tendency and dispersion. Returns a clean vector as opposed to base-r's \code{\link{scale}}.
}
\examples{
set.seed(1)
X = rnorm(100, mean = 10, sd = 5)
set.seed(1)
W = runif(100)
standardize(X, W)
standardize(X, robust = T) #almost the same for these data
}

