% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{std_df}
\alias{std_df}
\title{Standardize data.frame}
\usage{
std_df(df, exclude = "", messages = T)
}
\arguments{
\item{df}{(data.frame) The data.frame.}

\item{exclude}{(character vector) Names of variables to not standardize.}

\item{messages}{(boolean) Whether to output messages.}
}
\description{
Returns a standardized data.frame, i.e. one where every variable has mean 0 and sd 1.
}
\examples{
df = data.frame(rnorm(5, 100, 15))
df
std_df(df)
}
\keyword{data.frame,}
\keyword{standardize,}
\keyword{z-score}

