% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{str_clean}
\alias{str_clean}
\title{Clean string}
\usage{
str_clean(string, underscores = T, spacing_dots = T, end_dots = T,
  all_dots = F)
}
\arguments{
\item{string}{(a character scalar) A string to clean.}

\item{underscores}{(boolean) Whether to clean underscores. Default=T.}

\item{spacing_dots}{(boolean) Whether to clean spacing underscores. Default=T.}

\item{end_dots}{(boolean) Whether to clean dots at the end of the string. Default=T.}

\item{all_dots}{(boolean) Whether to clean all dots. Default=F.}
}
\description{
A simple wrapper str_replace_all() with sensible defaults.
}
\examples{
str_clean()
}
\keyword{character,}
\keyword{clean}
\keyword{string,}

