% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_detect2}
\alias{str_detect2}
\title{Improved str_detect}
\usage{
str_detect2(string, pattern, value = F)
}
\arguments{
\item{string}{(character vector) A character vector.}

\item{pattern}{(character vector) Pattern to look for (REGEX).}

\item{value}{(log scalar) Whether to return the matched values instead of logical values. Default = F.}
}
\value{
A character vector. Note that it will have a length longer than one if one of the inputs has that.
}
\description{
A wrapper for stringr's \code{\link{str_detect}} so that one can get the matches values, not just logical values. This restores the functionality of base-r's equivalent function, \code{\link{gsub}}.
}
\examples{
str_detect2(letters[1:10], pattern = "[acbde]")
str_detect2(letters[1:10], pattern = "[acbde]", value = T)
}

