% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_detect_replace}
\alias{str_detect_replace}
\title{Detect and replace values with regex}
\usage{
str_detect_replace(string, pattern, replacement)
}
\arguments{
\item{string}{(chr vector) A character vector.}

\item{pattern}{(chr vector) Pattern to look for (REGEX).}

\item{replacement}{(chr vector) A vector of values to replace with.}
}
\value{
A character vector.
}
\description{
A wrapper for stringr's \code{\link{str_detect}} so that one can detect strings with a particular pattern and replacement them outright. Vectorized over patterns and replacements.
}
\examples{
str = 1:99 + c(".", ",", ":") #example str vector
str_detect_replace(str, pattern = c("\\\\.", ",", ":"), replacement = c("dot", "comma", "colon")) #replace by pattern
}
