% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_filename}
\alias{str_filename}
\title{Extract filename from a file path.}
\usage{
str_filename(x)
}
\arguments{
\item{x}{(chr) A path to a filename.}
}
\value{
A character.
}
\description{
Extract filename from a file path.
}
\details{
Just a thin wrapper for \code{\link[base:basename]{base::basename()}}.
}
\examples{
dir(full.names = T) \%>\% str_filename()
}
