% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_legalize}
\alias{str_legalize}
\title{Make names legal}
\usage{
str_legalize(x, ascii = T)
}
\arguments{
\item{x}{(chr) A vector of names.}

\item{ascii}{Allow ascii only or all letters?}
}
\value{
a vector
}
\description{
Make variable names legal to use in R formulas etc.
}
\examples{
c("2017x", "male (\%)", "Male/female ratio") \%>\% str_legalize()
}
