% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_replace_multi}
\alias{str_replace_multi}
\title{Multiple replacement}
\usage{
str_replace_multi(string, patterns, replacement, all = T)
}
\arguments{
\item{string}{(a character scalar) A string.}

\item{patterns}{(a character vector) A character vector of things to clean. Regex.}

\item{replacement}{(a character scalar) What to replace matches with.}

\item{all}{(boolean) Whether to clean all instances or just the first. Default=T.}
}
\description{
A simple wrapper for stringr's str_replace() and str_replace_all().
}
\examples{
str_replace_multi()
}

