% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_round}
\alias{str_round}
\title{Round a number to desired number of shown digits}
\usage{
str_round(x, digits = 2, less_than = NULL, more_than = NULL, pad = F)
}
\arguments{
\item{x}{A vector of values}

\item{digits}{How many digits to print}

\item{less_than}{A value to add less than sign for}

\item{more_than}{A value to add greater than sign for}

\item{pad}{Whether to pad whitespace to force equal length strings}
}
\value{
A character vector
}
\description{
Round a number to desired number of shown digits
}
\examples{
str_round(1.123, 2)
str_round(1.100, 2)
str_round(1.1000000, 2)
str_round(1.1000000, 3)
str_round(0.0000001, 3)
str_round(seq(0, 7, by = 1), 2, less_than = 2, more_than = 5)
}
