% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_to_upper_initial}
\alias{str_to_upper_initial}
\title{Capitalize initial letter}
\usage{
str_to_upper_initial(x)
}
\arguments{
\item{x}{(chr) A character vector.}
}
\value{
A character vector same length as the input.
}
\description{
Capitalize the initial letter in a string. Vectorized.
}
\details{
Just a simple wrapper around stringr functions. Mainly useful for when writing tables to clipboard/supplementary files for writing studies.
}
\examples{
#make iris colnames lowercase, then capitalize initial letter.
colnames(iris) \%>\% str_to_lower() \%>\% str_to_upper_initial()
}
