% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_uniquify}
\alias{str_uniquify}
\title{Make strings unique}
\usage{
str_uniquify(string, suffix = " (\%d)", pad = F)
}
\arguments{
\item{string}{(chr vector) A character vector.}

\item{suffix}{(chr vector) A character to use to make unique suffixes. Must contain \verb{\%d}.}

\item{pad}{Whether to pad numbers by adding 0's in front to keep the same length in all suffixes.}
}
\description{
Make strings unique by adding a suffix to indicate which number copy it is.
}
\examples{
str_uniquify(letters[rep(1, 3)])
str_uniquify(letters[rep(1, 3)], suffix = ', \%d')
}
