% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_zero_to_lt}
\alias{str_zero_to_lt}
\title{Round and use less than symbol if zero}
\usage{
str_zero_to_lt(x, digits = 2)
}
\arguments{
\item{x}{numerical vector}

\item{digits}{Digits count}
}
\value{
character vector
}
\description{
Round and use less than symbol if zero
}
\examples{
c(.01) \%>\% str_zero_to_lt()
c(.01, .009, .001) \%>\% str_zero_to_lt()
c(.01, .009, .001) \%>\% str_zero_to_lt(digits = 3)
}
