% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{summarize_models}
\alias{summarize_models}
\title{Concise model comparison summary table}
\usage{
summarize_models(
  x,
  asterisks = c(0.01, 0.005, 0.001),
  asterisks_only = T,
  beta_digits = 2,
  beta_se_digits = beta_digits + 1,
  p_digits = 3
)
}
\arguments{
\item{x}{List of models}

\item{asterisks}{Thresholds for asterisks}

\item{asterisks_only}{Asterisks only? If false, then print rounded p values as well}

\item{beta_digits}{Digits for betas}

\item{beta_se_digits}{Digits for beta SEs}

\item{p_digits}{Digits for p values}
}
\value{
Data frame
}
\description{
Concise model comparison summary table
}
\examples{
models = list(
lm(Sepal.Width ~ Sepal.Length, data = iris),
lm(Sepal.Width ~ Sepal.Length + Petal.Width, data = iris),
lm(Sepal.Width ~ Sepal.Length + Petal.Width + Petal.Length, data = iris)
)
summarize_models(models)
summarize_models(models, asterisks = c(.05))
summarize_models(models, asterisks_only = F)
}
