% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{t_df}
\alias{t_df}
\title{Transpose data.frame}
\usage{
t_df(df)
}
\arguments{
\item{df}{(data.frame) A data.frame.}
}
\value{
A transposed data.frame.
}
\description{
Transpose a data.frame, returning a data.frame that also keeps the dimnames.
}
\examples{
t_df(iris)
}

