% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{test_HS}
\alias{test_HS}
\title{Test for heteroscedasticity}
\usage{
test_HS(resid, x)
}
\arguments{
\item{resid}{Model residuals}

\item{x}{Model predictor of interest}
}
\value{
Data frame of results
}
\description{
Test for heteroscedasticity
}
\examples{
#look for mostly nonexistent HS in iris
test_HS(resid = resid(lm(Sepal.Length ~ Petal.Length, data = iris2)), x = iris2$Petal.Length)
#a lot of HS here
test_HS(resid = resid(lm(Petal.Width ~ Petal.Length, data = iris2)), x = iris2$Petal.Length)
}
