% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{throws_error}
\alias{throws_error}
\title{Does call return an error?}
\usage{
throws_error(expr, silent_try = T)
}
\arguments{
\item{expr}{(expr) An expression.}

\item{silent_try}{(lgl scalar) Whether to use a silent try.}
}
\description{
Parses a string as a call and determines whether it returns an error or not. Returns a boolean.
}
\examples{
throws_error(log("")) #cannot take a logarithm of a string
throws_error(log(0)) #can take logarithm of 0
throws_error("log(0)") #works on strings too
throws_error("log(NULL)") #works on strings too
}

