% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{total_cells}
\alias{total_cells}
\title{Calculate total number of cells in an object}
\usage{
total_cells(x)
}
\arguments{
\item{x}{(any suitable object) The object.}
}
\value{
A whole number.
}
\description{
Calculate the total number of cells in an object. This is done by finding the product of the lengths of each dimension.
}
\examples{
total_cells(iris)
}
