% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{transform_01}
\alias{transform_01}
\title{Transform to 0-1 scale}
\usage{
transform_01(x)
}
\arguments{
\item{x}{Numeric factor}
}
\value{
Numeric factor on scale 0-1
}
\description{
Transform to 0-1 scale
}
\examples{
transform_01(1:5)
transform_01(c(1:3, NA, 4:5))
}
