% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{try_browser}
\alias{try_browser}
\title{Try and browse on error}
\usage{
try_browser(expr)
}
\description{
Opens the browser in the calling envirionment so you can see what went wrong. Useful to putting inside loops and only opening browser on the iterations that cause errors.
}

