% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{try_else}
\alias{try_else}
\title{Try and fail with a default value}
\usage{
try_else(expr, else. = NULL, silent = T)
}
\arguments{
\item{expr}{(expression) An expression to execute.}

\item{else.}{(any) Value returned on error.}

\item{silent}{(lgl) Whether to silence the error messages.}
}
\value{
The output from the expression or the else value.
}
\description{
Try and fail with a default value
}
\examples{
try_else(log(1))
try_else(log("abc"))
try_else(log("abc"), silent = F)
}
