% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unduplicate}
\alias{unduplicate}
\title{Remove duplicate elements}
\usage{
unduplicate(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Object of the same type as input.
}
\description{
Remove duplicated elements from an object. Simple wrapper for \code{\link[base:duplicated]{base::duplicated()}} to use in functional contexts.
}
\details{
Calls either \code{x[!duplicated(x)]} or \code{x[!duplicated(x), ]} depending on input dimensions. Will work for objects with a defined method for \code{\link[base:duplicated]{base::duplicated()}} e.g. \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}}.
}
\examples{
#simple vector
simple_vector = sample(1:5, size = 20, replace = T) \%T>\% print
simple_vector \%>\% unduplicate
#example on iris (data frame)
iris_with_dups = iris[c(1, 1, 2, 2, 3, 4, 1, 5, 2, 4), ] \%T>\% print
iris_with_dups \%>\% unduplicate
}
