% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{uniq_encoding}
\alias{uniq_encoding}
\title{Unique encoding}
\usage{
uniq_encoding(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A list
}
\description{
Unique encoding
}
\examples{
c(1, 2, 2, 3, 3, 3) \%>\% uniq_encoding()
c(3, 1, 3, 2, 3, 2) \%>\% uniq_encoding()
c(NA, 1, 2, 2, 3, 3, 3, NA) \%>\% uniq_encoding()
c(1, 2, 2, 3, 3, 3) \%>\% as.character() \%>\% uniq_encoding()
c(1, 2, 2, 3, 3, 3) \%>\% factor() \%>\% uniq_encoding()
}
