% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{v_to_df}
\alias{v_to_df}
\title{Vector to data frame}
\usage{
v_to_df(x, value_col = "value", name_col = F)
}
\arguments{
\item{x}{(vector) A vector.}

\item{value_col}{(chr) Name of the column with values. If NA, will use the name of x.}

\item{name_col}{(chr/lgl) Name of the column with names. If given, will not use rownames and use an explicit column.}
}
\value{
a data.frame
}
\description{
Transform a vector to a data.frame. If it has names, these will be added as rownames unless name_col is specified to a chr.
}
\examples{
v_to_df(1:4)
v_to_df(c(a = 1, b = 2, c = 3, d = 4))
v_to_df(c(a = 1, b = 2, c = 3, d = 4), name_col = "name")
}

