% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{which_max2}
\alias{which_max2}
\title{Generalized which.max(). Returns the index of the cell or number with the highest value.}
\usage{
which_max2(x)
}
\arguments{
\item{x}{a numeric vector, matrix or data.frame.}
}
\description{
A generalization of which.max() that works on matrices and data.frames too.
}
\examples{
m = matrix(runif(9), nrow=3)
m
which_max2(m)
}

