% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{winsorise}
\alias{winsorise}
\title{Winsorise numbers}
\usage{
winsorise(x, upper, lower)
}
\arguments{
\item{x}{(numeric vector) A vector of values.}

\item{upper}{(numeric scalar) The upper limit.}

\item{lower}{(numeric scalar) The lower limit.}
}
\description{
Winsorizes the numbers, that is, reduces any number above a given limit to that number and likewise for a lower limit.
}
\examples{
winsorise(seq(-3, 13), 10, 0)
}

