% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{write_clipboard.data.frame}
\alias{write_clipboard.data.frame}
\title{Write data frame to clipboard}
\usage{
\method{write_clipboard}{data.frame}(x, digits = 2, clean_names = T,
  clean_what = c("_", "\\\\."), pad_digits = T, print = T,
  .rownames = T, write_to_clipboard = interactive(), return_modified = F,
  capitalize_dimnames = T)
}
\arguments{
\item{x}{(any object that works with write.table) Something to write to the clipboard.}

\item{digits}{(int scalar) A number of digits to round the data to.}

\item{clean_names}{(log scalar) Whether to clean the names. Default=T.}

\item{clean_what}{(chr vector) Which things to clean. Defaults to underscores and dots.}

\item{pad_digits}{(log scalar) Whether to pad zeros to the digits (for prettier tables; default = T).}

\item{print}{(log scalar) Whether to also print the output in R (default T).}

\item{.rownames}{(lgl scalar) Whether to write rownames. Default yes. These are written to a column in front called .rownames.}

\item{write_to_clipboard}{(lgl) Whether to write to the clipboard. Can be useful to disable in rare cases.}

\item{return_modified}{(lgl) Whether to return the modified input instead of the original. Useful if one wants to modify it further.}

\item{capitalize_dimnames}{(lgl) Whether to capitalize the first letter in the dimnames.}
}
\description{
A wrapper function to \code{\link{write.table}} for writing to the clipboard for pasting in a spreadsheet.
}
\examples{
iris[-5] \%>\% cor() \%>\% write_clipboard()
iris \%>\% head() \%>\% miss_add_random() \%>\% write_clipboard()
#does not pad integers
test_df = data.frame(int = 1:5, num = rnorm(5))
test_df \%>\% write_clipboard()
}
