% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{write_clipboard.model_summary}
\alias{write_clipboard.model_summary}
\title{Write model summary to clipboard}
\usage{
\method{write_clipboard}{model_summary}(x, digits = 2)
}
\description{
Restructures a model summary object to a matrix and writes it to the clipboard.
}
\examples{
lm(Sepal.Length ~ Petal.Length, data = iris) \%>\% MOD_summary(kfold = F) \%>\% write_clipboard
}

