% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{write_clipboard}
\alias{write_clipboard}
\title{Write object to clipboard}
\usage{
write_clipboard(...)
}
\description{
S3 generic function to write objects to the clipboard for easy use.
}
\details{
See \code{\link{write_clipboard.data.frame}}, \code{\link{write_clipboard.model_summary}}.
}

