% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{write_mega}
\alias{write_mega}
\title{A wrapper for write.csv()}
\usage{
write_mega(object, filename)
}
\arguments{
\item{object}{the object to be written to a file}

\item{filename}{the name of the file you want to write to}
}
\description{
This is a convenient wrapper for write.csv()
}
