% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{write_rvest}
\alias{write_rvest}
\title{Write rvest/xml object}
\usage{
write_rvest(x, path, ...)
}
\arguments{
\item{x}{(obj) Object to write to disk.}

\item{path}{(str) Where to save it.}

\item{...}{Other parameters to write_rds.}
}
\value{
Invisibly returns x for use in pipelines.
}
\description{
Writes an rvest/xml object to disk for reuse. This is a wrapper around write_rds.
}

