% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_sessioninfo}
\alias{write_sessioninfo}
\title{Write session information to a file.}
\usage{
write_sessioninfo(filename = "sessions_info.txt", print = T)
}
\arguments{
\item{filename}{(chr scalar) The filename of the file to write to. Default is "sessions_info.txt".}

\item{print}{(logical scalar) Whether to also print the output.}
}
\description{
A wrapper for writeLines() and capture.output().
}
\examples{
write_sessioninfo("session_info.txt", print = TRUE)
}
