% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{wtd_mean}
\alias{wtd_mean}
\title{Calculate a weighted mean}
\usage{
wtd_mean(x, w)
}
\arguments{
\item{x}{(num vector) A vector of values.}

\item{w}{(num vector) A vector of weights.}
}
\description{
This is an improvement on \code{\link{weighted.mean}} in \code{base-r}.
}
\details{
The original function returns \code{NA} when there are missing values in the weights vector despite na.rm=T. This function avoids that problem. It also returns a useful error message if there are no complete cases. The function wraps base-r's function.
}
\examples{
set.seed(1)
X = rnorm(100)
set.seed(1)
W = runif(100)
wtd_mean(X) # not using weights
mean(X) #same as above
wtd_mean(X, W) #slightly different
}

