% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{wtd_median}
\alias{wtd_median}
\title{Weighted median}
\usage{
wtd_median(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{Vector of values}

\item{w}{Vector of weights}

\item{na.rm}{Ignore missing data?}
}
\description{
Function copied from \strong{spatstat} package.
}
