% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{wtd_quantile}
\alias{wtd_quantile}
\title{Weighted quantile}
\usage{
wtd_quantile(x, w, probs = seq(0, 1, 0.25), na.rm = TRUE)
}
\arguments{
\item{x}{Vector of values}

\item{w}{Vector of weights}

\item{probs}{Vector of probabilities}

\item{na.rm}{Ignore missing data?}
}
\description{
Function copied from \strong{spatstat} package.
}
