% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/misc.R
\name{add_newlines}
\alias{add_newlines}
\title{Insert newlines into text every nth character.}
\usage{
add_newlines(x, total.length = 95)
}
\arguments{
\item{x}{A character string or vector.}

\item{total.length}{The total length of each line of text. Defaults to 95 (suitable for ggplot2).}
}
\description{
Returns a character string with newlines every nth character. Works for character vectors too.
}
\examples{
add_newlines()
}
\keyword{label,}
\keyword{newline,}
\keyword{string,}
\keyword{text}

