% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{all_the_same}
\alias{all_the_same}
\title{Are all elements of a vector the same?}
\usage{
all_the_same(x)
}
\arguments{
\item{x}{(expression) Some expression to run.}
}
\description{
Tests whether all elements of a vector are the same. Uses the max/min method mentioned at .
}
\examples{
all_the_same(rep(1, 100))
all_the_same(rnorm(100))
}
\keyword{equal}
\keyword{identical,}
\keyword{same,}
\keyword{vector,}

