% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{are_equal}
\alias{are_equal}
\title{Are objects equal?}
\usage{
are_equal(x, y, ...)
}
\arguments{
\item{x}{(any object) The first object.}

\item{y}{(any object) The second object.}

\item{...}{(other named parameters) Further parameters to pass to \code{all.equal}.}
}
\value{
Returns a logical scalar indicating whether the objects are equal.
}
\description{
A wrapper for \code{\link{all.equal}} that returns a logical scalar.
}
\examples{
are_equal(iris[1:4], iris[-5])
}

