% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{as_num_df}
\alias{as_num_df}
\title{Convert convertible columns in a data.frame to numeric.}
\usage{
as_num_df(df, stringsAsFactors = F, skip_factors = T)
}
\arguments{
\item{df}{A data.frame.}

\item{stringsAsFactors}{Whether to convert strings to factors. Default is F.}

\item{skip_factors}{Whether to skip factors. Default is T.}
}
\description{
Attempts to convert all columns in a data.frame to numeric. If that fails, keeps the original.
}
\examples{
as_num_df()
}
\keyword{as,}
\keyword{convert,}
\keyword{data.frame}
\keyword{numeric,}

