% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{as_num_matrix}
\alias{as_num_matrix}
\title{Convert a data.frame to a numeric matrix, including factors.}
\usage{
as_num_matrix(df)
}
\arguments{
\item{df}{A data.frame.}
}
\description{
Returns a numeric matrix.
}
\examples{
as_num_matrix()
}
\keyword{convert}
\keyword{data.frame,}
\keyword{factor,}
\keyword{numeric,}

