% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{browse_if}
\alias{browse_if}
\title{Browse if equal to}
\usage{
browse_if(condition, extended = F)
}
\arguments{
\item{condition}{(lgl) A condition}

\item{extended}{(lgl) Whether to convert to logical.}
}
\description{
Browse on equals condition. Calls browse in the parent.frame so that one can see the context. Made for easier debugging. Pipe-friendly: returns output if condition is false. Wrapper for browse_if.
}
\examples{
browse_if(T) #only
browse_if(F) #
browse_if(1) #does not automatically type convert
browse_if(1, extended = T) #can if you want it
}
