% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{browse_if_equals}
\alias{browse_if_equals}
\title{Browse if equal to}
\usage{
browse_if_equals(x, equal_to, check.names = T, check.attributes = T)
}
\arguments{
\item{x}{(object) Any object.}

\item{equal_to}{(any object) Any object to test for output equality.}

\item{check.names}{(lgl) Whether to check names.}

\item{check.attributes}{(lgl) Whether to check all attributes.}
}
\description{
Browse on equals condition. Calls browse in the parent.frame so that one can see the context. Made for easier debugging. Pipe-friendly: returns output if condition is false. Wrapper for browse_if.
}
\examples{
browse_if(1+1, 2)
}
