% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{check_missing}
\alias{check_missing}
\title{Check if arguments are missing and raise error if they are}
\usage{
check_missing(var_names,
  error_msg = "[VAR] was missing! Please supply the input and try again.")
}
\arguments{
\item{var_names}{(chr vector) Names of variables to check.}

\item{error_msg}{(chr scalar) A template of the error message to show.}
}
\description{
Checks if arguments are missing and raises an error if they are. Put this in the beginning of your functions to check the input and give useful errors without writing checking code over and over again.
}
\examples{
test_func = function(y) {
check_missing("y")
T
}
test_func(y = ) #throws error
test_func(y = 1) #returns true
}

