% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation matrix}
\usage{
cor_matrix(data, weights, CI, CI_template = "\%r [\%lower \%upper]",
  skip_nonnumeric = T, CI_round = 2)
}
\arguments{
\item{data}{(data.frame or coercible into data.frame) The data.}

\item{weights}{(numeric vector, numeric matrix/data.frame or character scalar) Weights to use for the correlations. Can be a numeric vector with weights, the name of a variable in data, or a matrix/data.frame with weights for each variable. If the latter, then harmonic means are used. If none given, defaults to rep(1, nrow(data)).}

\item{CI}{(numeric scalar) The confidence level to use as a fraction.}

\item{CI_template}{(character scalar) A template to use for formatting the confidence intervals. Defaults to "\%r [\%lower \%upper]".}

\item{skip_nonnumeric}{(logical scalar) Whether to skip non-numeric variables. Defaults to TRUE.}

\item{CI_round}{(whole number scalar) If confidence intervals are used, how many digits should be shown?}
}
\description{
Outputs a correlation matrix. Supports weights, confidence intervals and rounding.
}
\examples{
cor_matrix(iris)
cor_matrix(iris, CI = .95)
}

