% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{count_NA}
\alias{count_NA}
\title{Count missing data}
\usage{
count_NA(x)
}
\arguments{
\item{x}{(any object) An object for which to count NAs.}
}
\description{
A simple wrapper for is.na() and sum(). Returns an integer.
}
\examples{
count_NA()
}
\keyword{NA,}
\keyword{count,}
\keyword{data,}
\keyword{missing}
\keyword{number}

