% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{describe2}
\alias{describe2}
\title{Describe data}
\usage{
describe2(
  x,
  na.rm = TRUE,
  interp = FALSE,
  skew = TRUE,
  ranges = TRUE,
  trim = 0.1,
  type = 3,
  check = TRUE,
  fast = NULL,
  quant = NULL,
  IQR = FALSE,
  omit = FALSE,
  data = NULL,
  all_vars = F
)
}
\value{
Tibble subset
}
\description{
Describe data
}
\examples{
describe2(iris)
describe2(mpg)
}
