% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_add_affix}
\alias{df_add_affix}
\title{Add affix to column names.}
\usage{
df_add_affix(data, prefix = "", suffix = "")
}
\arguments{
\item{data}{(an object) An object whose colnames should be changed}

\item{prefix}{(str) A prefix to add.}

\item{suffix}{(str) A suffix to add.}
}
\description{
Add a prefix to column names on data.frame or matrix or similar.
}
\examples{
test_iris = iris[1:10, ] #small test dataset
df_add_affix(test_iris, prefix = "P_") #ad P_ prefix
df_add_affix(test_iris, suffix = "_S") #ad _S suffix
}
