% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{df_addNA}
\alias{df_addNA}
\title{Insert random NAs into a data.frame.}
\usage{
df_addNA(df, prop = 0.1)
}
\arguments{
\item{df}{(data.frame) A data.frame.}

\item{prop}{(numeric) The proportion of NAs to add.}
}
\description{
Inserts missing data into a data.frame at random, thus creating data that are Missing Completely At Random (MCAR). THis isn't how data usually are missing in the real world, but may be sufficient for some simulations.
}
\examples{
df_addNA()
}
\keyword{NA,}
\keyword{add,}
\keyword{data,}
\keyword{date.frame,}
\keyword{missing}
\keyword{simulate}

