% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_fct_split}
\alias{df_fct_split}
\title{Split factors into multiple columns}
\usage{
df_fct_split(df, fcts, prefix = "", warn_unused_levels = T,
  type = "l", rm_old = F)
}
\arguments{
\item{df}{(df) A aata frame.}

\item{fcts}{(chr) Names of columns to split up.}

\item{prefix}{(chr) A prefix to add.}

\item{warn_unused_levels}{(lgl) Whether to warn the user when some levels have no cases.}

\item{type}{(chr) Which type of column should be created.}

\item{rm_old}{(lgl) Whether to remove the old columns after the splitting.}
}
\value{
A modified data frame.
}
\description{
Split factors into multiple columns
}
\details{
This function is a more user-friendly version of \code{\link[stats:model.matrix]{stats::model.matrix()}}.

For the prefix, one can use \code{\%v} to refer to the variable name.
}
\examples{
df_fct_split(iris, "Species")
}
