% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_residualize}
\alias{df_residualize}
\title{Residualized data.frame.}
\usage{
df_residualize(
  data,
  resid.vars,
  suffix = "",
  exclude.resid.vars = T,
  return.resid.vars = T,
  print.models = F,
  exclude_vars = NULL,
  weights = NA
)
}
\arguments{
\item{data}{(data.frame or matrix) The data object.}

\item{resid.vars}{(character vector) The names of the variables to use for the residualization.}

\item{exclude.resid.vars}{(logical) Whether to exclude the residualization variables from residualization. Defaults to TRUE.}

\item{return.resid.vars}{(logical) Whether to include the residualization variables in the returned data.frame. Defaults to TRUE.}

\item{print.models}{(logical) Wether to print the lm models used in the process. Defaults to TRUE.}

\item{exclude_vars}{(character vector) Names of variables are that excluded from the residualization.}

\item{weights}{(num vector) A vector of values to use for weights. If none given, all cases will be given equal weights.}
}
\description{
Returns a residualized data.frame given a set of variables to partial out.
}
\examples{
#generate some data
df = data.frame(a = rnorm(5), b = rnorm(5), c = rnorm(5))
weightsvar = runif(5)
#residualize the data.frame
df_residualize(df, resid.vars = "c") #without weights
df_residualize(df, resid.vars = "c", weights = weightsvar) #with weights
}
