% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{df_standardize}
\alias{df_standardize}
\title{Standardize data.frame}
\usage{
df_standardize(df, exclude = NULL, messages = T, exclude_factors = T,
  w = NULL)
}
\arguments{
\item{df}{(data frame) The data frame.}

\item{exclude}{(chr vector) Names of variables to not standardize.}

\item{messages}{(lgl scalar) Whether to output messages.}

\item{exclude_factors}{(lgl scalar) Whether to exclude factors.}

\item{w}{(num vector) Weights to use, if any.}
}
\value{
Returns a standardized data.frame, i.e. one where every variable has mean 0 and sd 1.
}
\description{
Standardize all variables in a data.frame. Can use weighted standardization. Wraps \code{\link{standardize}}.
}
\examples{
head(iris) #not standardized
head(df_standardize(iris)) #standardized
head(df_standardize(iris, exclude_factors = F)) #also standardize factors (may be nonsensical)
head(df_standardize(iris, w = runif(150))) #standardized with weights
}
