% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_CAFL}
\alias{FA_CAFL}
\title{Mixedness detection with a change in absolute factor loadings approach}
\usage{
FA_CAFL(x, ..., sort = 1, include_full_sample = T)
}
\arguments{
\item{x}{A data.frame to factor analyze.}

\item{...}{Settings for fa().}

\item{sort}{Which column to sort results after. Defaults to the first column (mean absolute change). Set to anything else than 1 or 2 to avoid sorting.}

\item{include_full_sample}{Whether to include the full sample as a case. Defaults to true.}
}
\description{
Examines how the exclusion of each particular case influences the factor loadings.
}
\examples{
FA_robust_cormatrix()
}
\keyword{analysis.}
\keyword{factor}
\keyword{psychometrics,}
\keyword{robust,}

