% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fa_CFS}
\alias{fa_CFS}
\title{Returns a numerical vector of the change in the size of the first factor by removing that case.}
\usage{
fa_CFS(data, sort = T, include_full_sample = T)
}
\arguments{
\item{data}{A data.frame.}

\item{sort}{Whether to sort the results or not. Defaults to true.}

\item{include_full_sample}{Whether to include the 'case' with the full sample. Defaults to true.}
}
\description{
Returns a numerical vector of the change in the size of the first factor by removing that case.
}
\examples{
fa_CFS(iris[-5])
}
