% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_CFS}
\alias{FA_CFS}
\title{Returns a numerical vector of the change in the size of the first factor by removing that case.}
\usage{
FA_CFS(data, sort = T, include_full_sample = T)
}
\arguments{
\item{data}{A data.frame.}

\item{sort}{Whether to sort the results or not. Defaults to true.}

\item{include_full_sample}{Whether to include the 'case' with the full sample. Defaults to true.}
}
\description{
Returns a numerical vector of the change in the size of the first factor by removing that case.
}
\examples{
FA_CFS()
}
\keyword{analysis}
\keyword{factor}
\keyword{latent}
\keyword{psychology,}
\keyword{psychometrics,}
\keyword{variable,}

