% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_MAR}
\alias{FA_MAR}
\title{Calculates the mean absolute residual at the case-level.}
\usage{
FA_MAR(data, sort = T, ...)
}
\arguments{
\item{data}{a data.frame.}

\item{sort}{whether to sort the results or not. Defaults to true.}
}
\description{
Extracts the first factor, then uses the factor scores to predict the indicator values for each indicator and for each case. Returns a data.frame with residuals. Then calculates the mean absolute value of these by case.
}
\examples{
FA_MAR()
}
\keyword{analysis,}
\keyword{factor}
\keyword{latent}
\keyword{psychology,}
\keyword{psychometrics,}
\keyword{residuals}
\keyword{variable,}

