% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{FA_nfactors}
\alias{FA_nfactors}
\title{Determine number of factors}
\usage{
FA_nfactors(.fa)
}
\arguments{
\item{.fa}{(fa or list of fa objects) The object from which to determine the number of factors.}
}
\description{
Convenience function for determining the number of factors in a fa object. Vectorized.
}
\examples{
library(psych)
fa(iris[-5]) \%>\% FA_nfactors
fa(iris[-5], nfactors = 2) \%>\% FA_nfactors
}

