% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_residuals}
\alias{FA_residuals}
\title{Find residuals on case-level basis for all indicators in a factor analysis.}
\usage{
FA_residuals(data, ...)
}
\arguments{
\item{data}{a data.frame.}
}
\description{
Extracts the first factor, then uses the factor scores to predict the indicator values for each indicator and for each case. Returns a data.frame with residuals.
}
\examples{
FA_residuals()
}
\keyword{analysis,}
\keyword{factor}
\keyword{latent}
\keyword{psychology,}
\keyword{psychometrics,}
\keyword{residuals}
\keyword{variable,}

