% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{fa_robust_cormatrix}
\alias{fa_robust_cormatrix}
\title{Robust correlation matrix}
\usage{
fa_robust_cormatrix(x, method = "lmrob")
}
\arguments{
\item{x}{A data.frame to correlate.}

\item{x}{method Which robust method to use. Options are "lmrob" \link{robustbase} or "rlm" \link{MASS}. Defaults to the first.}
}
\description{
Returns a correlation matrix with robust correlations. These are derived from either rlm() \link{MASS} or lmrob() \link{robustbase}.
}
\examples{
fa_robust_cormatrix()
}
