% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{FA_robust_fa}
\alias{FA_robust_fa}
\title{Robust factor analysis.}
\usage{
FA_robust_fa(x, ..., .method = "lmrob")
}
\arguments{
\item{x}{A data.frame to factor analyze.}
}
\description{
Performs a robust factor analysis. It is done using the output from FA_robust_cormatrix(). Note that this uses a correlation matrix, so factor scores are not available.
}
\examples{
FA_robust_cormatrix()
}
\keyword{analysis.}
\keyword{factor}
\keyword{psychometrics,}
\keyword{robust,}

