% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{FA_splitsample_repeat}
\alias{FA_splitsample_repeat}
\title{Repeated splithalf reliability with factor analysis}
\usage{
FA_splitsample_repeat(data, runs = 100, save_scores = F, messages = T,
  progress = T, seed = 1, ...)
}
\arguments{
\item{data}{(data.frame or matrix) The data.}

\item{runs}{(integer scalar) The number of runs to do.}

\item{save_scores}{(boolean scalar) Whether to save scores. Default=F.}

\item{messages}{(lgl scalar) Whether to display messages, default yes.}

\item{progress}{(lgl scalar) Whether to display progress bar, default yes.}

\item{seed}{(num scalar) Seed to use for reproducible results. Default=1.}

\item{...}{Extra parameters to pass to psych::fa().}
}
\description{
Divides a dataset into 2 at random, extracts a factor from each and correlates them. Then saves the correlation. Repeats this any desired number of times. Can also return the factor scores instead of correlations.
}
\examples{
FA_splitsample_repeat(iris[-5])
}

