% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{fail_if}
\alias{fail_if}
\title{Fail if input contains NA}
\usage{
fail_if(x, extended = F, msg = "There was an error")
}
\arguments{
\item{x}{(any object) The object to test.}

\item{extended}{(lgl) Whether to convert to logical.}

\item{msg}{(chr) An error message.}
}
\description{
Inputs any object, checks if it contains NA. If yes, throws an error. If not, returns the input.
}
\examples{
fail_if(T)
fail_if(1) #does not convert types
fail_if(1, extended = T)
}
