% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAC.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Find k neighbor cases using spherical or euclidean geometry.}
\usage{
find_neighbors(df, k = 3, distance_method = "spherical", lat_var = "lat",
  lon_var = "lon")
}
\arguments{
\item{df}{A data.frame with variables.}

\item{k}{Number of neighbors to find for each case. Must be smaller than number of cases. Defaults to 3.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean.}

\item{lat_var}{A numeric vector with the latitudes.}

\item{lon_var}{A numeric vector with the longitudes.}
}
\description{
Return a list of numeric vectors. Each vector contains the k nearest neighbors to the case. When using spherical geometry, you must give names of the variables. Euclidean geometry will use all variables. Cases with missing values are removed from the data.frame.
}
\examples{
find_neighbors()
}
\keyword{autocorrelation,}
\keyword{distance,}
\keyword{knn}
\keyword{latitude,}
\keyword{longitude,}
\keyword{neighbors,}
\keyword{spatial}

