% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAC.R
\name{get_distances}
\alias{get_distances}
\title{Calculate pairwise distances for all pairs for all variables in a data.frame.}
\usage{
get_distances(df, lat_var = "lat", lon_var = "lon", weights_var = "",
  weight_method = "harmonic")
}
\arguments{
\item{df}{A data.frame with variables.}

\item{lat_var}{A numeric vector with the latitudes. Defaults to "lat".}

\item{lon_var}{A numeric vector with the longitudes. Defaults to "lon".}

\item{weights_var}{A numeric vector with the weights.}

\item{weight_method}{A character string indicating which averaging method to use when combining weights for cases. Defaults to harmonic mean. Other options are arithmic and geometric.}
}
\description{
Inputs a data.frame and outputs a data.frame with distances between all possible case pairs for each variable. One can specify that a pair of variables are spherical coordinates which are then treated in a special fashion using get_spherical_dists(). One can also specify a variable that is weights and for which a suitable mean will be calculated.
}
\examples{
get_distances()
}
\keyword{autocorrelation,}
\keyword{distance,}
\keyword{latitude,}
\keyword{longitude,}
\keyword{spatial}
\keyword{weights}

