% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{get_each_subset_minus_1}
\alias{get_each_subset_minus_1}
\title{Get every subset of the data where one case is missing.}
\usage{
get_each_subset_minus_1(df)
}
\arguments{
\item{df}{A data.frame.}
}
\description{
Return a list of data.frames.
}
\examples{
df = data.frame(1:5, letters[1:5])
get_each_subset_minus_1(df)
}
\keyword{data.frame,}
\keyword{subset}

