% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{get_each_subset_minus_1}
\alias{get_each_subset_minus_1}
\title{Get every subset of the data where one case is missing.}
\usage{
get_each_subset_minus_1(df)
}
\arguments{
\item{df}{A data.frame.}
}
\description{
Return a list of data.frames.
}
\examples{
get_each_subset_minus_1()
}
\keyword{data.frame,}
\keyword{subset}

