% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC.R
\name{get_Morans_I_multi}
\alias{get_Morans_I_multi}
\title{Calculate multiple Moran's I's. It's a wrapper function for get_Morans_I().}
\usage{
get_Morans_I_multi(df, vars, dists = NULL, lat_var = NULL, lon_var = NULL,
  distance_method = NULL, auto_detect_dist_method = T)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{vars}{A character vector with the names of the variables for which Moran's I should be calculated.}

\item{dists}{A matrix of distances between cases.}

\item{lat_var}{A string with the name of the variable which has the latitude data.}

\item{lon_var}{A string with the name of the variable which has the longitude data.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean. If using euclidean it doesn't matter which variable is coded as lat or lon.}

\item{auto_detect_dist_method}{Whether to try to autodetect the distance method. If the dataset contains variables with the names "lat" and "lon", it will be detected as spherical. If it contains "x" and "y", it will be detected as euclidean. Defaults to true.}
}
\description{
Returns a numeric vector with Moran's I coefficients.
}
