% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_denhist}
\alias{GG_denhist}
\title{Histogram with an empirical density curve and a vertical line at the mean}
\usage{
GG_denhist(df, var)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{var}{A string of the name of the variable to use.}
}
\description{
Plots a histogram with an empirical density curve and a vertical line at the mean using ggplot2.
}
\examples{
GG_denhist()
}
\keyword{density,}
\keyword{ggplot2,}
\keyword{histogram}
\keyword{plot,}

