% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_funnel}
\alias{GG_funnel}
\title{Plot a contingency table with ggplot2}
\usage{
GG_funnel(.analysis, .CI = 0.95, .study_CI = F)
}
\arguments{
\item{.analysis}{(rma object) The rma analysis from metafor.}

\item{.CI}{(chr vector) Confidence interval to use. Default = .95.}

\item{.study_CI}{(lgl vector) Whether to plot confidence intervals for individual studies. Default no.}
}
\description{
Makes a pretty contingency table with ggplot2 using geom_tile.
}

