% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_group_means}
\alias{GG_group_means}
\title{ggplot2 with group means and error bars.}
\usage{
GG_group_means(df, var, groupvar = NULL, subgroupvar = NULL, CI = 0.95,
  type = "bar", na.rm = T, msg_NA = T, split_group_labels = T,
  line_length = 95)
}
\arguments{
\item{df}{(data.frame) A data.frame with variables.}

\item{var}{(chr scalar) The name of the variable to plot.}

\item{groupvar}{(chr scaler) The name of the grouping variable.}

\item{subgroupvar}{(chr scalar) The name of the subgrouping variable, if any.}

\item{CI}{(num scalar) The confidence interval to use. Default = .95.}

\item{type}{(chr scalar) The type of plot. Options: bar (default), point, points.}

\item{msg_NA}{(log scalar) Show a message if NAs were removed? (default true)}

\item{split_group_labels}{(log scalar) Whether to automatically insert newlines into group labels if they are too long (default yes).}

\item{line_length}{(num scalar) The desired line width (default 95). Only used when split_group_labels = T.}
}
\description{
Draws a nice ggplot2 with group means and error bars.
}
\examples{
#simple examples
GG_group_means(iris, "Sepal.Length", "Species")
GG_group_means(iris, "Sepal.Length", "Species", type = "point")
GG_group_means(iris, "Sepal.Length", "Species", type = "points")
GG_group_means(iris, "Sepal.Length", "Species", type = "points", CI = .999999)
GG_group_means(iris, "Sepal.Length", "Species", type = "violin")
GG_group_means(iris, "Sepal.Length", "Species", type = "violin2")

#subgroups too
iris$type = sample(LETTERS[1:3], size = nrow(iris), replace = T)
GG_group_means(iris, var = "Sepal.Length", groupvar = "Species", subgroupvar = "type")
GG_group_means(iris, var = "Sepal.Length", groupvar = "Species", subgroupvar = "type", type = "point")
GG_group_means(iris, var = "Sepal.Length", groupvar = "Species", subgroupvar = "type", type = "points")
GG_group_means(iris, var = "Sepal.Length", groupvar = "Species", subgroupvar = "type", type = "violin")
GG_group_means(iris, var = "Sepal.Length", groupvar = "Species", subgroupvar = "type", type = "violin2")
}
