% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_heatmap}
\alias{GG_heatmap}
\title{Heatmap correlation matrix with ggplot2}
\usage{
GG_heatmap(
  data,
  add_values = T,
  reorder_vars = T,
  digits = 2,
  font_size = 4,
  color_label = "Pearson\\nCorrelation",
  legend_position = c(0.4, 0.7),
  short_x_labels = F,
  axis_labels_clean_func = str_clean,
  dodge_x_labels = 1,
  min_n = 0,
  pairwise_n = NULL
)
}
\arguments{
\item{data}{Data frame or a matrix}

\item{add_values}{Whether to add the correlation sizes as text to plot}

\item{reorder_vars}{Whether to reorder variables so strongly related ones are close to each other}

\item{digits}{How many digits to print when plotting them}

\item{font_size}{If correlation values are plotted, their size}

\item{color_label}{Which label to use for the color scale legend}

\item{legend_position}{Position of the legend. Sometimes you may need to move it a bit}

\item{short_x_labels}{Adds integers to the y axis labels, and replaces the x labels with the same integers. Useful when there are many variables.}

\item{axis_labels_clean_func}{A function to clean the labels with, typically to remove underscores. NULL means disable.}

\item{dodge_x_labels}{How much to dodge the x labels. Useful when there are many variables.}

\item{min_n}{Minimum number of observations to include a correlation}
}
\value{
a ggplot2 object
}
\description{
Heatmap correlation matrix with ggplot2
}
