% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_heatmap}
\alias{GG_heatmap}
\title{Heatmap correlation matrix with ggplot2}
\usage{
GG_heatmap(data, add_values = T, reorder_vars = T, digits = 2)
}
\arguments{
\item{data}{Data frame)}

\item{add_values}{Whether to add the correlation sizes as text to plot}

\item{reorder_vars}{Whether to reorder variables so strongly related ones are close to each other}

\item{digits}{How many digits to print when plotting them}
}
\value{
a ggplot2 object
}
\description{
Heatmap correlation matrix with ggplot2
}
