% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_ordinal}
\alias{GG_ordinal}
\title{Plot ordinal proportions}
\usage{
GG_ordinal(
  data,
  clean_factor_levels = T,
  order = "positive",
  percentages = T,
  add_values = T,
  font_size = 4,
  exclude_values_below = 0
)
}
\arguments{
\item{data}{A data frame with ordinal variables}

\item{clean_factor_levels}{Whether to clean the factor levels}

\item{order}{How to order the variables. Default is "positive".}

\item{percentages}{Whether to plot percentages or counts. Default is TRUE.}
}
\value{
A ggplot2 object
}
\description{
Plot ordinal proportions
}
\examples{
xx = tibble(
ord_1 = cut(rnorm(200), breaks = c(-Inf, -1, 0, 1, Inf), labels = c("A", "B", "C", "D")),
ord_2 = cut(rnorm(200, mean = 1), breaks = c(-Inf, -1, 0, 1, Inf), labels = c("A", "B", "C", "D")),
ord_3 = cut(rnorm(200, mean = -1), breaks = c(-Inf, -1, 0, 1, Inf), labels = c("A", "B", "C", "D")),
ord_4 = cut(rnorm(200, mean = 0), breaks = c(-Inf, -1, 0, 1, Inf), labels = c("A", "B", "C", "D"))
)
GG_ordinal(xx)
}
