% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_proportions}
\alias{GG_proportions}
\title{ggplot2 proportions plot}
\usage{
GG_proportions(x, group, drop_empty = F)
}
\arguments{
\item{x}{A factor variable}

\item{group}{Grouping variable}

\item{drop_empty}{Drop empty combinations}
}
\value{
A ggplot2 of proportions
}
\description{
ggplot2 proportions plot
}
\examples{
#plot the proportions of cylinders by year
GG_proportions(mpg$year, mpg$cyl)

#remove the 0\%'s
GG_proportions(mpg$year, mpg$cyl, drop_empty = T)
}
