% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_proportions}
\alias{GG_proportions}
\title{ggplot2 proportions plot}
\usage{
GG_proportions(
  data,
  x,
  group,
  add_values = T,
  repel = F,
  text_size = 3,
  drop_empty = F
)
}
\arguments{
\item{data}{Data frame}

\item{x}{The variable on the x axis, usually years, countries, or similar.}

\item{group}{The variable to calculation proportions of, should be a countable. Will be forced to factor.}
}
\value{

}
\description{
ggplot2 proportions plot
}
\examples{
#plot the proportions of cylinders by year
GG_proportions(mpg, "year", "cyl")
#remove the 0\%'s
GG_proportions(mpg, "year", "cyl", drop_empty = T)
#don't label the values
GG_proportions(mpg, "year", "cyl", add_values = F)
#alternative variable
GG_proportions(mpg, "year", "class")
#not pretty but gets the job done
GG_proportions(mpg, "year", "manufacturer", repel = T)

#another dataset
datasets::Titanic \%>\%
inv_table() \%>\%
GG_proportions("Class", "Survived")

datasets::Titanic \%>\%
inv_table() \%>\%
GG_proportions("Survived", "Sex")
}
