% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_save}
\alias{GG_save}
\title{Save a ggplot2 figure}
\usage{
GG_save(
  filename,
  plot = last_plot(),
  path = NULL,
  width = 10,
  height = 6.5,
  ...
)
}
\description{
Save a ggplot2 figure. Now with a sensible default size (10 x 6.5 inches). Same arguments as \code{ggplot2::ggsave}
}
