% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_scatter}
\alias{GG_scatter}
\title{Scatter plot with regression line and correlation information using ggplot2}
\usage{
GG_scatter(df, x_var, y_var, text_pos = "tl", case_names = T)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{x_var}{X variable string.}

\item{y_var}{Y variable string.}

\item{text_pos}{Where to put the text. Defaults to top right ("tl"). Other options: tr, bl, br.}

\item{case_names}{Whether to add case names or not. Defaults to true. Row names are used for case names.}
}
\description{
Plots a scatterplot with a regression line and correlation information. Returns a ggplot2 object.
}
\examples{
GG_scatter()
}
\keyword{ggplot2,}
\keyword{plot,}
\keyword{scatter}

