% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GG_scatter}
\alias{GG_scatter}
\title{Scatter plot with regression line and correlation information using ggplot2}
\usage{
GG_scatter(df, x_var, y_var, weights = NULL, color = NULL, alpha = 1,
  text_pos = NA, case_names = NULL, CI = 0.95, clean_names = T,
  check_overlap = T, repel_names = F, ...)
}
\arguments{
\item{df}{(data.frame) A data frame with variables.}

\item{x_var}{(chr scalar) X variable string.}

\item{y_var}{(chr scalar) Y variable string.}

\item{weights}{(num scalar) A set of weights to use.}

\item{color}{(chr) A variable to color points by.}

\item{alpha}{(num) The alpha to use.}

\item{text_pos}{(chr scalar) Where to put the text. Defaults to top right ("tl") if correlation is positive, or tr if negative. Can be tl, tr, bl, or br.}

\item{case_names}{(lgl scalar) Whether to add case names or not (default true).}

\item{CI}{(num scalar) Confidence interval as a fraction.}

\item{clean_names}{(lgl scalar) Whether to clean the axes names using str_clean().}

\item{check_overlap}{(lgl scalar) Whether to avoid overplotting names.}

\item{repel_names}{(lgl) If using case names, should they be repelled?}
}
\description{
Plots a scatterplot with a regression line and correlation information. Returns a \code{ggplot2} object.
}
\details{
Internally uses the ad hoc variables \code{.weights}, \code{.label} and \code{.color}. If you name your variables these, then you will get odd problems.
}
\examples{
GG_scatter(iris, "Sepal.Length", "Sepal.Width") #default plot
GG_scatter(iris, "Sepal.Length", "Sepal.Width", case_names = rep("A", 150)) #case names
GG_scatter(iris, "Sepal.Length", "Sepal.Width", case_names = "Species") #casenames from variable
GG_scatter(iris, "Sepal.Length", "Sepal.Width", case_names = "Species", repel_names = T) #casenames from variable, repelled
GG_scatter(iris, "Sepal.Length", "Sepal.Width", text_pos = "br") #other text location
GG_scatter(iris, "Sepal.Length", "Sepal.Width", CI = .99) #other CI
GG_scatter(iris, "Sepal.Length", "Sepal.Width", clean_names = F) #don't clean names
GG_scatter(iris, "Sepal.Length", "Sepal.Width", weights = 1:150) #add weights with vector
GG_scatter(iris, "Sepal.Length", "Sepal.Width", weights = "Petal.Width") #add weights with name
GG_scatter(iris, "Sepal.Length", "Sepal.Width", color = "Species") #color points
GG_scatter(iris, "Sepal.Length", "Sepal.Width", color = "Species", case_names = "Species") #color points, but labels stay black
GG_scatter(iris, "Sepal.Length", "Sepal.Width", alpha = .1) #change alpha
}
