% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{homogeneity}
\alias{homogeneity}
\title{Calculate homogeneity/heterogeneity}
\usage{
homogeneity(x, reverse = F, summary = F)
}
\arguments{
\item{x}{(a vector) A vector of values.}

\item{reverse}{(log scalar) Whether to reverse the index to index heterogeneity (default false).}

\item{summary}{(log scalar) Whether to treat data as summary statistics of the group proportions (default false). If data are given in 0-100 format, it will automatically convert.}
}
\description{
Calculate a simple index of homogeneity for nominal data, that is variously called Simpson's, Herfindahl's or Hirschman's index.
}
\examples{
homogeneity(iris$Species)
homogeneity(iris$Species, reverse = T)
homogeneity(c(.7, .2, .1), summary = T)
homogeneity(c(80, 15, 5), summary = T)
}
