% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{is_negative}
\alias{is_negative}
\title{Are values negative?}
\usage{
is_negative(x)
}
\arguments{
\item{x}{(numeric vector) Some values to test.}
}
\description{
A wrapper for "< 0".
}
\examples{
is_negative(1)
is_negative(0)
is_negative(-1)
}
