% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{is_positive}
\alias{is_positive}
\title{Are values positive?}
\usage{
is_positive(x)
}
\arguments{
\item{x}{(numeric vector) Some values to test.}
}
\description{
A wrapper for "> 0".
}
\examples{
is_positive(1)
is_positive(0)
is_positive(-1)
}

