% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{is_positive}
\alias{is_positive}
\title{Are valyes positive?}
\usage{
is_positive(x)
}
\arguments{
\item{x}{(any object compatible with <) Some values to test.}
}
\description{
A wrapper for ">0", returns a boolean. Vectorized.
}
\examples{
is_positive()
}

