% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_are.R
\name{is_simple_vector}
\alias{is_simple_vector}
\title{Is object a simple vector?}
\usage{
is_simple_vector(x)
}
\arguments{
\item{x}{(any object) An object to test.}
}
\description{
A simple wrapper for is.vector and is.list. The normal is.vector function returns true for lists which is undesirable. Returns a boolean.
}
\examples{
l = list(1:10)
v = 1:10
is.vector(v)
is.vector(l)
is_simple_vector(v)
is_simple_vector(l)
}
