% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_math.R
\name{is_zero}
\alias{is_zero}
\title{Are values zero?}
\usage{
is_zero(x)
}
\arguments{
\item{x}{(any object compatible with <) Some values to test.}
}
\description{
A wrapper for "==0", returns a boolean. Vectorized.
}
\examples{
is_zero(0)
is_zero(1)
}

