% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{Jensen_plot}
\alias{Jensen_plot}
\title{Jensen method (method of correlated vectors) plot}
\usage{
Jensen_plot(loadings, cors, reverse = TRUE, text.location = "tl")
}
\arguments{
\item{loadings}{a vector of factor loadings.}

\item{reverse}{whether to reverse indicators with negative loadings. Default to true.}

\item{text.location}{which corner to write the numerical results in. Options are "tl", "tr", "bl", "br". Defaults to "tl".}

\item{loadings}{a vector of correlations of the indicators with the criteria variable.}
}
\description{
Returns a ggplot2 scatter plot with numerical results in a corner. Also supports reversing for dealing with factors that have negative indicators.
}
\examples{
Jensen_plot()
}
\keyword{latent}
\keyword{psychology}
\keyword{psychometrics}
\keyword{variable}

