% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{lm_beta_matrix}
\alias{lm_beta_matrix}
\title{Matrix of beta coefficients of all simple linear models.}
\usage{
lm_beta_matrix(dependent, predictors, data, standardized = T, .weights = NA,
  return_models = "b", messages = T)
}
\arguments{
\item{dependent}{(character scalar) The name of the dependent variable.}

\item{predictors}{(character vector) The names of the redictor variables.}

\item{data}{(data.frame) A data.frame with the variables.}

\item{standardized}{(boolean) Whether to standardize the results. Defaults to true.}

\item{.weights}{(numeric vector) A numeric vector of weights to use. Defaults to NA, which causes it to use unit weights for all cases.}

\item{return_models}{(character scalar) What to return. all = all models, best = best model. Defaults to best.}

\item{messages}{(boolean) Whether to show messages. Default=TRUE.}
}
\description{
Returns a data.frame with beta coefficients for all possible simple (without interactions) linear models given a set of predictor variables and a dependent variable.
}
\examples{
#try all models in iris dataset to predict sepal length
lm_beta_matrix(dependent = "Sepal.Length", predictors = c("Sepal.Width", "Petal.Length", "Petal.Width", "Species"), data = iris)
}

