% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{MAT_vector2full}
\alias{MAT_vector2full}
\title{Reconstructs a symmetric matrix based on a vector of values of one of the halves.}
\usage{
MAT_vector2full(x, diag = F, byrow = F)
}
\arguments{
\item{x}{A vector of values.}

\item{diag}{Whether the diagonal values are included or not. Defaults to F.}

\item{byrow}{Whether the values were extracted by row or column. Defaults to F which matches the output from e.g. dist().}
}
\description{
Returns a symmetric matrix.
}
\examples{
MAT_vector2full()
}
\keyword{lower,}
\keyword{matrix,}
\keyword{symmetric,}
\keyword{upper}

