% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Dataset merger function}
\usage{
merge_datasets(DF1, DF2, main = 0, time = F)
}
\arguments{
\item{DF1}{the first data.frame}

\item{DF2}{the second data.frame}

\item{main}{which data.frame should be used as the main? Choose the larger one if working with large datasets. Default to using neither.}
}
\description{
This function allows you to merge two data.frames by their overlapping rownames.
}
\examples{
merge_datasets()
}
\keyword{combining}
\keyword{data.frame}
\keyword{datasets}
\keyword{merging}

