% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{merge_datasets_multi}
\alias{merge_datasets_multi}
\title{Dataset merger function for multiple data.frames.}
\usage{
merge_datasets_multi(..., main = 1, time = F)
}
\arguments{
\item{...}{(data.frames) Two or more data.frames to merge.}
}
\description{
This is a wrapper for merge_datasets().
}
\examples{
merge_datasets_multi()
}
\keyword{combining,}
\keyword{data.frame,}
\keyword{datasets,}
\keyword{merging,}
\keyword{multi,}
\keyword{wrapper}

