% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{merge_rows}
\alias{merge_rows}
\title{Merge rows in data.frame.}
\usage{
merge_rows(df, key, func = sum, numeric = TRUE)
}
\arguments{
\item{df}{(data.frame or matrix) The data object.}

\item{key}{(character scalar) The name of the key variable, which is the variable to merge rows by.}

\item{func}{(function) The function to use. Note that if you set numeric = FALSE, then the function must be able to handle non-numeric data.
numeric (logical scalar) Whether to apply the function only to the numeric columns. Default=TRUE.}
}
\description{
Find blocks of rows with a matching key and merge them using a given function.
}
\examples{
t = data.frame(key = c("a", "a", "b", "b", "c"), value = 1:5)
merge_rows(t, "key") #rows merged by sum
merge_rows(t, "key", func = mean) #rows merged by mean
}

