% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_filter}
\alias{miss_filter}
\title{Filter data by missing values per row.}
\usage{
miss_filter(data, missing = 0, reverse = F)
}
\arguments{
\item{data}{(data frame) The data.}

\item{missing}{(num) The maximum number of missing values in cases. Defaults to 0 (keep only cases with no missing values).}

\item{reverse}{(lgl) Filter based on non-NA data instead.}
}
\description{
Counts the number of missing values per row and then keeps rows that have at most a chosen number of missing values.
}
\examples{
df = data.frame(1:10, letters[1:10])
df = miss_add_random(df)
miss_filter(df)
}
