% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_filter}
\alias{miss_filter}
\title{Filter data by missing values per row.}
\usage{
miss_filter(data, missing = 0, reverse = F, by_case = T, vars = NULL)
}
\arguments{
\item{data}{(data frame) The data.}

\item{missing}{(num) The maximum number of missing values in cases. Defaults to 0 (keep only cases with no missing values). If this is a proportion between 0-1, then it is used as such.}

\item{reverse}{(lgl) Filter based on non-NA data instead.}
}
\description{
Counts the number of missing values per row and then keeps rows that have at most a chosen number of missing values.
}
\examples{
set.seed(1)
df = data.frame(ints = 1:10, letters = letters[1:10], unif = runif(10), norm = rnorm(10))
df = miss_add_random(df, prop = .25) \%>\% mutate(nomiss = T, allmiss = NA)
df
#filter cases/rows
miss_filter(df) #allow no missing values
miss_filter(df, missing = 1) #allow up to 1 value missing
miss_filter(df, missing = 0.5) #allow up to a half missing
#only use certain variables for counting
miss_filter(df, vars = c("ints", "letters"))
#filter columns/vars
miss_filter(df, missing = 0, by_case = F)
miss_filter(df, missing = 0.5, by_case = F)
}
