% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_impute}
\alias{miss_impute}
\title{Impute data using VIM::irmi}
\usage{
miss_impute(data, max_na = floor(ncol(data)/2), noise = F)
}
\arguments{
\item{data}{(data.frame) A data.frame.}

\item{max_na}{(num scalar) The maximum number of missing datapoints per case.}
}
\value{
A data.frame with missing data imputed for the desired cases.
}
\description{
Useful wrapper for VIM's irmi function. Can skip cases without changing case order depending on the number of missing values
}
\examples{
df = miss_add_random(iris[-5]) #example data, remove data at random from iris num data
miss_impute(df) #impute missing
}
