% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{miss_plot}
\alias{miss_plot}
\alias{plot_miss}
\title{Missing data histogram with ggplot2.}
\usage{
miss_plot(df, percent = T, case = T)
}
\arguments{
\item{df}{a data.frame.}

\item{percent}{whether to use percent or not. Defaults to true.}
}
\description{
Returns a ggplot2 histogram plot.
}
\examples{
miss_plot(df_addNA(iris))
}

