% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_k_fold_r2}
\alias{MOD_k_fold_r2}
\title{Calculate k fold cross validated r2}
\usage{
MOD_k_fold_r2(lmfit, folds = 10, runs = 100, seed = 1)
}
\arguments{
\item{lmfit}{(an lm fit) An lm fit object.}

\item{folds}{(whole number scalar) The number of folds to use (default 10).}
}
\description{
Using k fold cross-validation, estimate the true r2 in a new sample. This is better than using adjusted r2 values.
}
\examples{
fit = lm("Petal.Length ~ Sepal.Length", data = iris)
MOD_k_fold_r2(fit)
}

