% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_LASSO}
\alias{MOD_LASSO}
\title{Repeatedly run glmnet.cv()}
\usage{
MOD_LASSO(data, dependent, predictors, weights_ = NULL, standardize = T,
  runs = 100, alpha_ = 1, NA_ignore = T, seed = 1, messages = T,
  progress = T)
}
\arguments{
\item{data}{(data.frame) A data.frame with the data. Must contain dependent and predictor variables.}

\item{dependent}{(chr scalar) The name of the dependent variable.}

\item{predictors}{(chr vector) The names of the predictor variables.}

\item{weights_}{(num vector/chr scalar) If weights should be used, a numeric vector of values to use or chr vector.}

\item{standardize}{(log scalar) Whether to standardize the data beforehand.}

\item{runs}{(int scalar) Number of times to run.}

\item{alpha_}{(num scalar) The penalty to use. 1 = lasso regression, 0 = ridge regression.}

\item{NA_ignore}{(log scalar) Whether to remove cases with missing data.´´}

\item{seed}{(int scalar) The seed to use. For reproducible results.}

\item{messages}{(log scalar) Whether to show messages.}

\item{progress}{(log scalar) Whether to show a progress bar.}
}
\description{
Returns a data frame of beta coefficients from glmnet.cv() fits.
}
\details{
The messages parameter is depreceated in favor of the progress parameter.
}
\examples{
MOD_LASSO(iris, "Sepal.Length", predictors = colnames(iris)[-1])
}
