% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{MOD_repeat_cv_glmnet}
\alias{MOD_repeat_cv_glmnet}
\title{Repeatedly run glmnet.cv()}
\usage{
MOD_repeat_cv_glmnet(df, dependent, predictors, weights_ = NA,
  standardize = T, runs = 100, alpha_ = 1, NA_ignore = T,
  messages = T)
}
\arguments{
\item{dependent}{(character scalar) The name of the dependent variable.}

\item{predictors}{(character vector) The names of the predictor variables.}

\item{weights_}{(numeric vector) If weights should be used, a numeric vector of values to use. Defaults to equal weights.}

\item{standardize}{(boolean) Whether to standardize the data beforehand. Defaults to true.}

\item{runs}{(numeric/integer scalar) Number of times to run. Defaults to 100.}

\item{alpha_}{(numeric scalar) The penalty to use. 1 = lasso regression, 0 = ridge regression. Defaults to 1.}

\item{NA_ignore}{(boolean) Whether to remove cases with missing data. Defaults to T.}

\item{messages}{(boolean) Whether to send messages to the user.}

\item{data}{(data.frame) A data.frame with the data. Must contain dependent and predictor variables.}
}
\description{
Returns a data frame of beta coefficients from glmnet.cv() fits.
}
\examples{
MOD_repeat_cv_glmnet()
}
\keyword{cross-validation,}
\keyword{fit,}
\keyword{glmnet,}
\keyword{model,}
\keyword{repeat}

