% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{new_lines_adder}
\alias{new_lines_adder}
\title{Insert newlines into text every nth character.}
\usage{
new_lines_adder(x, interval)
}
\arguments{
\item{x}{A character string.}

\item{interval}{How often the newlines are added.}
}
\description{
Returns a character string with newlines every nth character. See also add_newlines().
}
\examples{
new_lines_adder()
}
\keyword{label,}
\keyword{newline,}
\keyword{string,}
\keyword{text}

