% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{output_sorted_var}
\alias{output_sorted_var}
\title{Write a nice sorted vector with names.}
\usage{
output_sorted_var(df, var, filename)
}
\arguments{
\item{df}{a data frame.}

\item{var}{a string with the name of the variable to write.}

\item{var}{a string with the desired filename to write to.}
}
\description{
This is a convenience function for outputting lists
}
\examples{
output_sorted_var()
}
\keyword{convenience,}
\keyword{output}

