% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_loadings_multi}
\alias{plot_loadings_multi}
\title{Plot multiple factor loadings in one plot.}
\usage{
plot_loadings_multi(fa_objects, fa_labels, reverse_vector = NA,
  reorder = "all")
}
\arguments{
\item{fa_objects}{a list of factor analyses objects from the fa() function from the psych package.}

\item{fa_labels}{a character vector for names of the analyses. Defaults to fa.1, fa.2, etc..}

\item{reorder}{(character scalar or NA) Which factor analysis to order the loadings by. Can be integers, "all" or NA for not reordering.}

\item{reverse.vector}{a numeric vector to multiple factor loadings with. Use e.g. c(1, -1) to reverse the second factor. Defaults not reversing.}
}
\description{
Returns a ggplot2 plot with sorted loadings colored by the analysis they belong to. Supports reversing óf any factors that are reversed. Dodges to avoid overplotting. Only works for factor analyses with 1 factor solutions!
}
\examples{
plot_loadings_multi()
}
\keyword{analysis,}
\keyword{factor}
\keyword{ggplot2}
\keyword{latent}
\keyword{plot,}
\keyword{psychology,}
\keyword{psychometrics,}
\keyword{variable,}

