% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rank_df}
\alias{rank_df}
\title{Rank numeric variables of a data frame.}
\usage{
rank_df(df, ...)
}
\arguments{
\item{df}{A data.frame.}

\item{...}{Other parameters for rank().}
}
\description{
Returns a data.frame where numeric variables have been converted to their ranks.
}
\examples{
rank_df()
}
\keyword{data.frame}
\keyword{rank,}

