% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{remove_redundant_vars2}
\alias{remove_redundant_vars2}
\title{Remove redundant variables from a data.frame based on a threshold value. This is done by calculating all the intercorrelations, then finding those that correlate at or above the threshold (absolute value), then removing the second pair of each variable and not removing more variables than strictly necessary.}
\usage{
remove_redundant_vars2(df, threshold = 0.9, cor_method = "pearson",
  messages = T)
}
\arguments{
\item{df}{(data.frame) A data.frame with numeric variables.}

\item{threshold}{(numeric scalar) A threshold above which intercorrelations are removed. Defaults to .9.}

\item{cor_method}{(character scalar) The correlation method to use. Parameter is fed to cor(). Defaults to "pearson".}

\item{messages}{(boolean) Whether to print diagnostic messages.}
}
\description{
Remove redundant variables from a data.frame based on a threshold value. This is done by calculating all the intercorrelations, then finding those that correlate at or above the threshold (absolute value), then removing the second pair of each variable and not removing more variables than strictly necessary.
}
\examples{
remove_redundant_vars()
}

