% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modeling.R
\name{residualize_DF}
\alias{residualize_DF}
\title{Residualized data.frame.}
\usage{
residualize_DF(data, resid.vars, suffix = "", exclude.resid.vars = T,
  return.resid.vars = T, print.models = T)
}
\arguments{
\item{data}{A data.frame or matrix.}

\item{resid.vars}{A character vector of the variables to partial out.}

\item{exclude.resid.vars}{Whether to exclude the residualization variables from residualization. Defaults to true.}

\item{return.resid.vars}{Whether to include the residualization variables in the returned data.frame. Defaults to true.}

\item{print.models}{Wether to print the lm models used in the process. Defaults to true.}
}
\description{
Returns a residualized data.frame given a set of variables to partial out.
}
\examples{
residualize_DF()
}
\keyword{modeling,}
\keyword{partialing}
\keyword{residualize,}

