% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{round_df}
\alias{round_df}
\title{Round numeric variables of a data frame.}
\usage{
round_df(df, digits = 3)
}
\arguments{
\item{df}{A data.frame or matrix.}

\item{digits}{The number of digits to round to.}
}
\description{
Returns a data.frame where numeric variables have been rounded to the desired number of digits.
}
\examples{
round_df()
}
\keyword{data.frame}
\keyword{round,}

