% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAC.R
\name{SAC_slr}
\alias{SAC_slr}
\title{Perform spatial local regression.}
\usage{
SAC_slr(df, dependent, predictors, k = 3, output = "trim10", dists, lat_var,
  lon_var, distance_method, auto_detect_dist_method = T,
  weights_method = "inverse", weights_var = "", include_self = F,
  verbose = T)
}
\arguments{
\item{df}{A data.frame with variables.}

\item{dependent}{A character vector with the name of the dependent variable.}

\item{predictors}{A character vector with the names of the predictor variables.}

\item{k}{The number of neighbors to use in the regressions. Defaults to 3.}

\item{output}{Which meta-analysis measure to use. Defaults to trim10, a 10 percent trimmed mean. Other options are: mean, median, and trim followed by any desired number. Can also use 'values' to get the values which returns a data.frame.}

\item{dists}{A matrix of distances between cases.}

\item{lat_var}{A string with the name of the variable which has the latitude/east-west data.}

\item{lon_var}{A string with the name of the variable which has the longitude/north-south data.}

\item{distance_method}{Which geometric system to use to calculate distances. Defaults to spherical. Can be either spherical or euclidean. If using euclidean it doesn't matter which variable is coded as lat or lon.}

\item{auto_detect_dist_method}{Whether to try to autodetect the distance method. If the dataset contains variables with the names "lat" and "lon", it will be detected as spherical. If it contains "x" and "y", it will be detected as euclidean. Defaults to true.}

\item{weights_method}{A string with the name of the weights method to use. Defaults to "inverse", i.e. weighing cases by their inverse distance to the case being considered. Can also be "none".}
}
\description{
Returns either summary statistics for each predictor or a data.frame with the betas from each local regression.
}

