% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plot_to_file}
\alias{save_plot_to_file}
\title{Save a non-ggplot2 plot to a file}
\usage{
save_plot_to_file(code, filename, width = 1000, height = 750, ...)
}
\arguments{
\item{code}{A code chunk that produces a plot}

\item{filename}{The desired filename}

\item{width}{Width of image}

\item{height}{Height of image}
}
\value{
Nothing
}
\description{
Save a non-ggplot2 plot to a file
}
\examples{
plot(1:3)
save_plot_to_file(plot(1:3), filename = "test.png")
file.remove("test.png")
}
