% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{score_items}
\alias{score_items}
\title{Score multiple choice items}
\usage{
score_items(df, key)
}
\arguments{
\item{df}{(data.frame) A data.frame with responses.}

\item{key}{(vector) A vector with the correct responses.}
}
\description{
Score a data.frame with multiple choice items using an answer key. Each column has a answers for one question.
}
\examples{
score_items()
}
\keyword{choice}
\keyword{date.frame,}
\keyword{multiple}
\keyword{score,}

