% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{SMD_matrix}
\alias{SMD_matrix}
\title{Standardized mean differences}
\usage{
SMD_matrix(
  x,
  group,
  central_tendency = wtd_mean,
  dispersion = "sd",
  dispersion_method = "all",
  extended_output = F,
  CI = 0.95,
  str_template = "\%d [\%lower \%upper]",
  str_round_to = 2,
  reliability = 1,
  se_analytic = T,
  ...
)
}
\arguments{
\item{x}{(numeric vector) A vector of values.}

\item{group}{(vector) A vector of group memberships.}

\item{central_tendency}{(function) A function to use for calculating the central tendency. Must support a parameter called na.rm. Ideal choices: mean, median.}

\item{dispersion}{(character or numeric scalar) Either the name of the metric to use (sd or mad) or a value to use.}

\item{dispersion_method}{(character scalar) If using one of the built in methods for dispersion, then a character indicating whether to use the pooled value from the total dataset (all), the pairwise comparison (pair), or the sd from the total dataset (total).}

\item{extended_output}{(lgl) Whether to output a list of matrices. Useful for computational reuse.}

\item{CI}{(num) Confidence interval coverage.}

\item{str_template}{(chr) A string template to use.}

\item{reliability}{(num) A reliability to use for correcting for measurement error. Done via \code{\link[kirkegaard:a4me_cohen_d]{kirkegaard::a4me_cohen_d()}}.}

\item{se_analytic}{(lgl) Use analytic standard errors. If not, then it will bootstrapping (slower). Always uses bootstrapping for non-mean functions.}

\item{...}{(other arguments) Additional arguments to pass to the central tendency function.}
}
\description{
Calculate standardized mean differneces between all groups.
}
\examples{
SMD_matrix(iris$Sepal.Length, iris$Species)
SMD_matrix(iris$Sepal.Length, iris$Species, extended_output = T)
#pairwise SDs
SMD_matrix(iris$Sepal.Length, iris$Species, extended_output = T, dispersion_method = "pair")
}
