% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sort_df}
\alias{sort_df}
\title{Sort a data.frame by one or more variables.}
\usage{
sort_df(df, vars = names(df), decreasing = F)
}
\arguments{
\item{df}{(data.frame) A data.frame.}

\item{vars}{(strings/integers) variables to use for sorting.}

\item{decreasing}{Whether to use decreasing order. Default=F.}
}
\description{
A wrapper for order(). Improved from the version in the reshape package.
}
\examples{
sort_df()
}
\keyword{NA,}
\keyword{add,}
\keyword{data,}
\keyword{date.frame,}
\keyword{missing}
\keyword{simulate}

