% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector}
\usage{
standardize(x, w = NULL, robust = F, sample = T,
  focal_group = NULL)
}
\arguments{
\item{x}{(num vector) A vector of values.}

\item{w}{(num vector) A vector of weights.}

\item{robust}{(log vector) Whether to use robust measures (default false). See \code{\link{mad}} and \code{\link{median}}.}

\item{sample}{(log scalar) Whether this is a sample as opposed to a population (default true).}

\item{focal_group}{(lgl vector) A subset of the data to standardize the values to. This is useful when you want one subgroup to be the focal group, using their mean/sd as 0/1.}
}
\description{
Standardize a vector. Can use weights and robust measures of central tendency and dispersion. Returns a clean vector as opposed to base-r's \code{\link{scale}}.
}
\examples{
set.seed(1)
X = rnorm(100, mean = 10, sd = 5)
set.seed(1)
W = runif(100)
standardize(X, W)
standardize(X, robust = T) #almost the same for these data
}
