% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{std_df}
\alias{std_df}
\title{Standardize data.frame}
\usage{
std_df(df, exclude = "")
}
\arguments{
\item{df}{A data.frame or matrix.}

\item{exclude}{A character vector of variables to exclude from the process.}
}
\description{
Returns a standardized data.frame, i.e. one where every variable has mean 0 and sd 1.
}
\examples{
std_df()
}
\keyword{data.frame,}
\keyword{standardize,}
\keyword{z-score}

