% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_functions.R
\name{std_df}
\alias{std_df}
\title{Standardize data.frame}
\usage{
std_df(df, exclude = "", messages = T, exclude_factors = T, w)
}
\arguments{
\item{df}{(data.frame) The data.frame.}

\item{exclude}{(chr vector) Names of variables to not standardize.}

\item{messages}{(log scalar) Whether to output messages (default T).}

\item{exclude_factors}{(log scalar) Whether to exclude factors (default T).}

\item{w}{(num vector) Weights to use, if any.}
}
\value{
Returns a standardized data.frame, i.e. one where every variable has mean 0 and sd 1.
}
\description{
Standardize all variables in a data.frame. Can use weighted standardization. Wraps \code{\link{standardize}}.
}
\examples{
head(iris) #not standardized
head(std_df(iris)) #standardized
head(std_df(iris, exclude_factors = F)) #also standardize factors (may be nonsensical)
head(std_df(iris, w = runif(150))) #standardized with weights
}

