% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_round}
\alias{str_round}
\title{Round a number to desired number of shown digits}
\usage{
str_round(x, round_to = str_round_to)
}
\arguments{
\item{x}{A vector of values}

\item{round_to}{Number of digits to show}
}
\value{
A character vector
}
\description{
Round a number to desired number of shown digits
}
\examples{
str_round(1.123, 2)
str_round(1.100, 2)
str_round(1.1000000, 2)
str_round(1.1000000, 3)
}
