% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_uniquify}
\alias{str_uniquify}
\title{Make strings unique}
\usage{
str_uniquify(string, suffix = " (\%d)")
}
\arguments{
\item{string}{(chr vector) A character vector.}

\item{suffix}{(chr vector) A character to use to make unique suffixes. Must contain \%d.}
}
\value{
A character vector.
}
\description{
Detects duplicate strings and makes them unique by adding a number at the end.
}
\details{
This function loops over the groups of identical strings and adds the suffix if the group has more than 1 member. This suffix addition is done using \code{sprintf}.
}
\examples{
Example vector with some duplicates
(x = sample(LETTERS[1:10], size = 20, replace = T))
#uniquify
x \%>\% str_uniquify
#custom suffix using a second \%d.
x \%>\% str_uniquify(" [\%d/\%d]")
}

