% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{table2}
\alias{table2}
\title{Better table}
\usage{
table2(x, prop = F, include_NA = T)
}
\arguments{
\item{x}{(a vector) A vector.}

\item{prop}{(lgl) Whether to return proportions instead of percents. Default no.}

\item{include_NA}{(lgl) Whether to count NA values. Default yes.}
}
\value{
A data frame (tbl).
}
\description{
Returns a data frame table with both counts and percents.
}
\examples{
sample(letters[1:10], size = 100, replace = T) \%>\% table2
}

