% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{throws_error}
\alias{throws_error}
\title{Does call return an error?}
\usage{
throws_error(x, silent_try = T)
}
\arguments{
\item{x}{(character scalar) Code to run in a string.}

\item{silent_try}{(log scalar) Whether to use a silent try (default true).}
}
\description{
Parses a string as a call and determines whether it returns an error or not. Returns a boolean.
}
\examples{
throws_error("log('123')") #cannot take a logarithm of a string
}

