% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{try_browse}
\alias{try_browse}
\title{Browse on error}
\usage{
try_browse(expr)
}
\arguments{
\item{expr}{(expr) Any expression that might cause an error.}
}
\description{
Browse on error. Calls browse in the parent.frame so that one can see the context the for error.
}
\examples{
try_browse({log("p")})
}

