% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{which_min2}
\alias{which_min2}
\title{Generalized which.min(). Returns the index of the cell or number with the lowest value.}
\usage{
which_min2(x)
}
\arguments{
\item{x}{a numeric vector, matrix or data.frame.}
}
\description{
A generalization of which.min() that works on matrices and data.frames too.
}
\examples{
which_min2()
}
\keyword{index}
\keyword{min,}

