% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{write_clipboard}
\alias{write_clipboard}
\title{Write object to clipboard}
\usage{
write_clipboard(x, digits = 2, clean_names = F, clean_what = c("_",
  "\\\\."), pad_digits = T, print = F)
}
\arguments{
\item{x}{(any object that works with write.table) Something to write to the clipboard.}

\item{digits}{(int scalar) A number of digits to round the data to.}

\item{clean_names}{(log scalar) Whether to clean the names. Default=F.}

\item{clean_what}{(chr vector) Which things to clean. Defaults to underscores and dots.}

\item{pad_digits}{(log scalar) Whether to pad zeros to the digits (for prettier tables; default = T).}

\item{print}{(log scalar) Whether to also print the output in R (default F).}
}
\description{
A wrapper function to \code{\link{write.table}} for writing to the clipboard for pasting in a spreadsheet.
}
\examples{
write_clipboard(cor(iris[-5]))
write_clipboard(miss_add_random(iris))
}

