% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{write_clipboard}
\alias{write_clipboard}
\title{Write object to clipboard}
\usage{
write_clipboard(x, digits = 3, clean_names = F, clean_what = c("_",
  "\\\\."))
}
\arguments{
\item{x}{(any object that works with write.table) Something to write to the clipboard.}

\item{digits}{(integer) A number of digits to round the data to.}

\item{clean_names}{(boolean) Whether to clean the names. Default=F.}

\item{clean_what}{(character vector) Which things to clean. Defaults to underscores and dots.}
}
\description{
A wrapper function to write.table() for writing to the clipboard for pasting in a spreadsheet.
}
\examples{
write_clipboard()
}
\keyword{clipboard}
\keyword{export,}
\keyword{output,}
\keyword{write,}

