% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_IO.R
\name{write_clipboard}
\alias{write_clipboard}
\title{Write object to clipboard}
\usage{
write_clipboard(x)
}
\arguments{
\item{x}{An matrix or data.frame, or something similar.}
}
\description{
A wrapper function to write.table() for writing to the clipboard for pasting in a spreadsheet.
}
\examples{
write_clipboard()
}
\keyword{clipboard}
\keyword{export,}
\keyword{output,}
\keyword{write,}

