% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_sessioninfo}
\alias{write_sessioninfo}
\title{Write session information to a file.}
\usage{
write_sessioninfo(filename, print = FALSE)
}
\arguments{
\item{filename}{(character scalar) The filename of the file to write to.}

\item{print}{(logical scalar) Whether to also print the output. Default=FALSE.}
}
\value{
Logical scalar.
}
\description{
A wrapper for writeLines() and capture.output().
}
\examples{
write_sessioninfo("session_info.txt", print = TRUE)
}

