% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{wtd_median}
\alias{wtd_median}
\title{Weighted median}
\usage{
wtd_median(x, w = rep(1, length(x)))
}
\arguments{
\item{x}{Vector of values}

\item{w}{Vector of weights}
}
\description{
Function copied from \strong{spatstat} package.
}
