% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{wtd_sd}
\alias{wtd_sd}
\title{Calculate weighted standard deviation}
\usage{
wtd_sd(x, w = NULL, sample = T, error = F)
}
\arguments{
\item{x}{(num vector) A vector of values.}

\item{w}{(num vector) A vector of weights.}

\item{sample}{(log scalar) Whether this is a sample as opposed to a population (default true).}

\item{error}{(lgl scalr) Whether to throw an error if there is no data at all or no pairwise complete cases. Default yes.}
}
\description{
Calculated the weighted standard deviation using a vector of values and a vector of weights.
}
\examples{
set.seed(1)
X = rnorm(100)
set.seed(1)
W = runif(100)
sd(X) #0.898
wtd_sd(X, W) #0.894, slightly different
wtd_sd(X) #0.898, not using weights
}
