% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{wtd_sum}
\alias{wtd_sum}
\title{Calculate a weighted sum}
\usage{
wtd_sum(x, w, error = T)
}
\arguments{
\item{x}{(num vector) A vector of values.}

\item{w}{(num vector) A vector of weights.}

\item{error}{(lgl scalr) Whether to throw an error if there is no data at all or no pairwise complete cases. Default yes.}
}
\description{
This is an improvement on \code{\link{sum}} in \code{base-r}.
}
\details{
It automatically handles missing data. It returns a useful error message if there are no complete cases.
}
\examples{
set.seed(1)
X = rnorm(100)
set.seed(1)
W = runif(100)
wtd_sum(X) # not using weights
sum(X) #same as above
wtd_sum(X, W) #different
}

